% Generated by roxygen2 (4.0.0): do not edit by hand
\name{a2m}
\alias{a2m}
\title{Converts an array of coordinates to a matrix}
\usage{
a2m(a)
}
\arguments{
\item{a}{An \code{array} of coordinates.}
}
\value{
A matrix (see above)
}
\description{
Converts an array of coordinates into a matrix.
All the individuals (the 3rd dimension of the array) becomes rows, and
columns are (all) x coordinates and (all) y coordinates, so that we have
x1, x2, ..., xn, y1, y2, ..., yn. Rows and colums are names anyway.
}
\examples{
data(wings)
a <- l2a(wings$coo)
a
a2m(a)
}
\seealso{
\link{m2a} the reverse function
}
\keyword{Babel}

