\name{Momocs Package}
\alias{Momocs-package}
\alias{Momocs}
\docType{package}
\title{
Outline Analysis using R.
}
\description{
Momocs is intended to ease and popularize shape analysis of outlines (especially using elliptical Fourier analysis). It mostly hinges on the core functions developped in \emph{Morphometrics with R} (Claude, 2008). From outline extraction of images and elliptical analysis, radii variation and tangent angle Fourier calculation to multivariate analysis and calculation of the morphological space, Momocs provides a complete and convenient toolkit to specialists within every field that are, or may be, interested in morphological comparisons of outlines.

Have a look to Momocs' website: \url{http://www.vincentbonhomme.fr/Momocs}. You will find a graphical introduction to Momocs.
}

\author{
\enumerate{
\item Vincent Bonhomme, French Institute of Pondicherry, India. <bonhomme.vincent@gmail.com>
\item Sandrine Picq, UMR CBAE, Montpellier, France.
\item Julien Claude UMR, ISEM, Universite de Montpellier II, France.
}}

\references{
The companion paper of Momcs is submitted.

Have a look to Julien's book that greatly inspired Momocs: Claude, J. (2008) \emph{Morphometrics with R}, Use R! series, Springer 316 pp.

An exhaustive collection of the functions published in the book can be found here: \url{http://www.isem.cnrs.fr/IMG/txt/Rfunctions.txt}.
}

\keyword{Package}