% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-misc.R
\name{andnow}
\alias{andnow}
\alias{andnow_method}
\title{And now, what to do?}
\usage{
andnow(x)

andnow_method(x)
}
\arguments{
\item{x}{any object, or class (quoted or not)}
}
\description{
\link{andnow}, given an object, return available methods for its class(es);
\link{andnow_method}, given a function/method name, return supported classes.
}
\examples{
#methods for data.frame
andnow(iris)

#methods for Coo objects
andnow(bot)

#classes supported by efourier
andnow_method("efourier")

# methods for plot
andnow_method("plot")
}
