% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{arrange}
\alias{arrange}
\alias{arrange.Coe}
\alias{arrange.Coo}
\alias{arrange.PCA}
\alias{arrange.default}
\title{Arrange (ala dplyr) rows with matching conditions, using the $fac slot on Momocs classes}
\usage{
arrange(.data, ...)

\method{arrange}{default}(.data, ...)

\method{arrange}{Coo}(.data, ...)

\method{arrange}{Coe}(.data, ...)

\method{arrange}{PCA}(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{Logical conditions}
}
\value{
a a \code{Coo}, \code{Coe}, \code{PCA} object
}
\description{
Extends this dplyr verb to \link{Coo} objects. See examples.
}
\details{
dplyr verbs are maintained.
}
\note{
this method is quite experimental, check the result and do not hesitate to report a bug.
}
\examples{
data(olea)
head(olea$fac)
head(arrange(olea$fac, view, desc(domes)))
}
\seealso{
\link{select}, \link{filter}, \link{slice}, \link{chop}, \link{combine}.
}

