% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{as_df}
\alias{as_df}
\title{Convert Momocs objects to data.frames}
\usage{
as_df(x)
}
\arguments{
\item{x}{an object, typically a Momocs class}
}
\value{
a data.frame
}
\description{
Used in particular for ggplot2 compatibility
}
\examples{
data(bot)
head(as_df(bot))
bot.f <- efourier(bot, 10)
head(as_df(bot.f))
bot.p <- PCA(bot.f)
head(as_df(bot.p))
bot.l <- LDA(bot.p, "type")
head(as_df(bot.l))
}
\seealso{
Other babel functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{chc2Out}},
  \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{l2a}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2l}},
  \code{\link{nef2Coe}}, \code{\link{ntsrow2Coo}},
  \code{\link{pix2chc}}, \code{\link{tie_jpg_txt}},
  \code{\link{tps2coo}}
}

