% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{pix2chc}
\alias{pix2chc}
\alias{chc2pix}
\title{Convert (x; y) coordinates to chaincoded coordinates}
\usage{
pix2chc(coo)

chc2pix(chc)
}
\arguments{
\item{coo}{(x; y) coordinates passed as a matrix}

\item{chc}{chain coordinates}
}
\description{
Useful to convert (x; y) coordinates to chain-coded coordinates.
}
\note{
Note this function will be deprecated from Momocs
when \code{Momacs} and \code{Momit}
will be fully operationnal.
}
\examples{
pix2chc(shapes[1]) \%T>\% print \%>\% # from pix to chc
chc2pix()                         # and back
}
\references{
Kuhl, F. P., & Giardina, C. R. (1982).
Elliptic Fourier features of a closed contour.
\emph{Computer Graphics and Image Processing}, 18(3), 236-258.
}
\seealso{
\link{chc2pix}

Other import functions: 
\code{\link{import_Conte}()},
\code{\link{import_StereoMorph_curve1}()},
\code{\link{import_jpg1}()},
\code{\link{import_jpg}()},
\code{\link{import_tps}()},
\code{\link{import_txt}()}

Other import functions: 
\code{\link{import_Conte}()},
\code{\link{import_StereoMorph_curve1}()},
\code{\link{import_jpg1}()},
\code{\link{import_jpg}()},
\code{\link{import_tps}()},
\code{\link{import_txt}()}
}
\concept{import functions}
