\name{bezier.i}
\alias{bezier.i}
\title{Calculates a shape from Bezier coefficients}
\usage{
bezier.i(B, nb.pts = 120)
}
\arguments{
  \item{B}{a matrix}

  \item{nb.pts}{the number of points to return for drawing
  he shape}
}
\value{
a matrix of (x; y) coordinates
}
\description{
todo
}
\keyword{morphoCore}

