\name{bezier}
\alias{bezier}
\title{Calculates Bezier coefficients from a shape
#todo}
\usage{
bezier(coo, n)
}
\arguments{
  \item{coo}{a matrix or a list of (x; y) coordinates}

  \item{n}{the degree}
}
\value{
a list with J and B
}
\description{
Calculates Bezier coefficients from a shape #todo
}
\keyword{morphoCore}

