% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bezier_i}
\alias{bezier_i}
\title{Calculates a shape from Bezier coefficients}
\usage{
bezier_i(B, nb.pts = 120)
}
\arguments{
\item{B}{a matrix of Bezier vertices, such as those produced by \link{bezier}}

\item{nb.pts}{the number of points to sample along the curve.}
}
\value{
a matrix of (x; y) coordinates
}
\description{
Calculates a shape from Bezier coefficients
}
\note{
Directly borrowed for Claude (2008), and called \code{beziercurve} there.
Not implemented for open outlines but may be useful for other purposes.
}
\examples{
set.seed(34)
x <- coo_sample(efourier_shape(), 5)
plot(x, ylim=c(-3, 3), asp=1, type='b', pch=20)
b <- bezier(x)
bi <- bezier_i(b$B)
lines(bi, col='red')
}
\references{
Claude, J. (2008) \emph{Morphometrics with R}, Use R! series,
}
\keyword{Bezier}

