\docType{data}
\name{bot}
\alias{bot}
\title{Outline coordinates of 20 beer and 20 whisky bottles.}
\format{An Out object containing the outlines coordinates and a grouping factor
for 20 beer and 20 whisky bottles}
\source{
Images have been grabbed on the internet and prepared by
the package's authors. No particular choice has been made
on the dimension of the original images or the brands cited
here.
}
\description{
Outline coordinates of 20 beer and 20 whisky bottles.
}
\keyword{datasets}

