% Generated by roxygen2 (4.0.0): do not edit by hand
\name{boxplot.OpnCoe}
\alias{boxplot.OpnCoe}
\title{Boxplot on OpnCoe matrices of polynomial coefficients}
\usage{
\method{boxplot}{OpnCoe}(x, retain, drop, palette = col.gallus,
  title = "Variation of polynomials coefficients", ...)
}
\arguments{
\item{x}{the \link{OpnCoe} object}

\item{retain}{numeric the number of harmonics to retain}

\item{drop}{numeric the number of harmonics to drop}

\item{palette}{a color \link{palette}}

\item{title}{a title for the plot}

\item{...}{useless here but maintain the consistency with generic boxplot}
}
\description{
Allows to explore variability of these coefficients, typically after a
rawPolynomials or orthoPolynomials on Opn objects.
}
\examples{
data(olea)
olea.p <- rawPolynomials(olea, 5)
boxplot(olea.p)
}
\seealso{
\link{hist.OpnCoe}
}
\keyword{OpnCoe}
\keyword{graphics}

