% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coe.R
\name{boxplot.OutCoe}
\alias{boxplot.OutCoe}
\alias{boxplot.Coe}
\title{Boxplot of morphometric coefficients}
\usage{
\method{boxplot}{OutCoe}(x, ...)
}
\arguments{
\item{x}{the \link{Coe} object}

\item{...}{useless here}
}
\value{
a ggplot2 object
}
\description{
Explores the distribution of coefficient values.
}
\examples{
# on OutCoe
bot \%>\% efourier(9) \%>\% rm_harm(1) \%>\% boxplot()

data(olea)
op <- opoly(olea)
boxplot(op)
}
\seealso{
Other Coe_graphics: \code{\link{hcontrib}}
}
