% Generated by roxygen2 (4.0.2): do not edit by hand
\name{breed}
\alias{breed}
\alias{breed.Coe}
\alias{breed.default}
\title{Mutate Coe (and others) objects}
\usage{
breed(x, ...)

\method{breed}{default}(x, margin = 2, size, rate = 1, ...)

\method{breed}{Coe}(x, size, rate = 1, ...)

breed(x, ...)

\method{breed}{default}(x, margin = 2, size, rate = 1, ...)

\method{breed}{Coe}(x, size, rate = 1, ...)
}
\arguments{
\item{x}{the object to permute}

\item{...}{useless here}

\item{margin}{numeric whether 1 or 2 (rows or columns)}

\item{size}{numeric the required size for the final object, same size by default}

\item{rate}{numeric the number of sd for \link{rnorm}, 1 by default.}

\item{x}{the object to permute}

\item{margin}{numeric whether 1 or 2 (rows or columns)}

\item{size}{numeric the required size for the final object, same size by default}

\item{rate}{numeric the number of sd for \link{rnorm}, 1 by default.}

\item{...}{useless here}
}
\description{
This methods applies column-wise on the \code{coe} of any
\link{Coe} object but relies on a function that can be used on any matrix. It
simply uses \link{rnorm} with the mean and sd calculated for every column (or row).
For a \code{Coe} object, on every colum, randomly generates coefficients values
centered on the mean of the column, and with a sd equals to it standard deviates
multiplied by \code{rate}.

This methods applies column-wise on the \code{coe} of any
\link{Coe} object but relies on a function that can be used on any matrix. It
simply uses \link{rnorm} with the mean and sd calculated for every column (or row).
For a \code{Coe} object, on every colum, randomly generates coefficients values
centered on the mean of the column, and with a sd equals to it standard deviates
multiplied by \code{rate}.
}
\examples{
m <- matrix(1:12, nrow=3)
breed(m, margin=2, size=4)
breed(m, margin=1, size=10)

data(bot)
bot.f <- efourier(bot, 12)
bot.m <- breed(bot.f, 80)
bot.m
panel(bot.m)
m <- matrix(1:12, nrow=3)
breed(m, margin=2, size=4)
breed(m, margin=1, size=10)

data(bot)
bot.f <- efourier(bot, 12)
bot.m <- breed(bot.f, 80)
bot.m
panel(bot.m)
}
\seealso{
\link{perm}

\link{perm}
}
\keyword{Coe}

