% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core-outopn-calibrate.R
\name{calibrate_deviations}
\alias{calibrate_deviations}
\alias{calibrate_deviations.Out}
\title{Quantitative calibration, through deviations, for Out objects}
\usage{
calibrate_deviations(Coo, ...)

\method{calibrate_deviations}{Out}(Coo, method = c("efourier", "rfourier",
  "tfourier"), id = 1, harm.range, norm.centsize = TRUE,
  dist.method = edm_nearest, dist.nbpts = 120, ...)
}
\arguments{
\item{Coo}{the \code{Out} object on which to calibrate_deviations}

\item{...}{only used for the generic}

\item{method}{any method from \code{c('efourier', 'rfourier', 'tfourier')}}

\item{id}{the shape on which to perform calibrate_deviations}

\item{harm.range}{vector of harmonics on which to perform calibrate_deviations.
If not provided, the harmonics corresponding to 0.9, 0.95 and 0.99% of harmonic power
are used.}

\item{norm.centsize}{logical whether to normalize deviation by the centroid size}

\item{dist.method}{a method such as \link{edm_nearest} to calculate deviations}

\item{dist.nbpts}{numeric the number of points to use for deviations calculations}
}
\value{
a ggplot object
}
\description{
Calculate deviations from original and reconstructed shapes using a
range of harmonic number.
}
\examples{
data(bot)
calibrate_deviations(bot)
\dontrun{
library(ggplot2)
gg <- calibrate_deviations(bot, id=1:20)$gg
gg + geom_hline(yintercept=c(0.001, 0.005), linetype=3)
gg + labs(col="Number of harmonics", fill="Number of harmonics",
           title="Harmonic power") + theme_bw()
gg + coord_polar()
}
}
\keyword{Out}

