% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core-outopn-calibrate.R
\name{calibrate_r2}
\alias{calibrate_r2}
\title{Quantitative calibration, through r2, for Opn objects}
\usage{
calibrate_r2(x, ...)
}
\arguments{
\item{x}{and Opn object}

\item{...}{useless here}

\item{method}{one of 'npoly' or 'opoly'}

\item{id}{the ids of shapes on which to calculate r2 (all by default)}

\item{degree.range}{on which to calculate r2}

\item{thresh}{the threshold to return diagnostic}

\item{plot}{logical whether to print the plot}

\item{verbose}{logical whether to print messages}
}
\description{
Estimates the r2 to calibrate the degree for \code{npoly} and \code{opoly} methods.
Also returns a plot
}
\details{
May be long, so you can estimate it on a sample.
}
\examples{
\dontrun{
data(olea)
calibrate_r2(olea, "opoly", degree.range=1:12, thresh=c(0.9, 0.99)

}
}
\seealso{
\link{calibrate_harmonicpower}
}

