% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calibrate_reconstructions}
\alias{calibrate_reconstructions}
\alias{calibrate_reconstructions.Opn}
\alias{calibrate_reconstructions.Out}
\title{Calibrate using reconstructed shapes}
\usage{
calibrate_reconstructions(x, ...)

\method{calibrate_reconstructions}{Out}(x, method = c("efourier", "rfourier",
  "tfourier"), id, range = 1:9, plot.method = c("panel", "stack")[1],
  legend = TRUE, legend.title = "Nb of harmonics", palette = col_heat,
  shp.col = NA, shp.border = "#1A1A1A", ...)

\method{calibrate_reconstructions}{Opn}(x, method = c("npoly", "opoly"), id,
  range = c(2, 3, 4, 6, 8, 10), baseline1 = c(-1, 0), baseline2 = c(1, 0),
  plot.method = c("panel", "stack")[1], legend = TRUE,
  legend.title = "Degree", palette = col_india, shp.border = "#1A1A1A",
  ...)
}
\arguments{
\item{x}{the \code{Coo} object on which to calibrate_reconstructions}

\item{...}{additional parameters to fed \link{coo_plot}}

\item{method}{any method from \code{c('efourier', 'rfourier', 'tfourier')}}

\item{id}{the shape on which to perform calibrate_reconstructions}

\item{range}{vector of harmonics on which to perform calibrate_reconstructions}

\item{plot.method}{either \code{'\link{panel}'} or \code{'\link{stack}'}}

\item{legend}{logical whether to plot a legend}

\item{legend.title}{if TRUE above, its title}

\item{palette}{a color \link{palette}}

\item{shp.col}{a color for the shape (\code{NA} by default)}

\item{shp.border}{a color for the border of the shape}

\item{baseline1}{\eqn{(x; y)} coordinates for the first point of the baseline}

\item{baseline2}{\eqn{(x; y)} coordinates for the second point of the baseline}
}
\description{
Calculate and displays reconstructed shapes using a
range of harmonic number.
}
\examples{
data(bot)
calibrate_reconstructions(bot)
}
\keyword{Out}

