% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{chop}
\alias{chop}
\title{Rougher slicing that accepts a classifier ie a column name from the \code{$fac} on Momocs classes}
\usage{
chop(.data, fac)
}
\arguments{
\item{.data}{a \code{Coo} or \code{Coe} object}

\item{fac}{a column name from the \code{$fac}}
}
\value{
a named list of \code{Coo} or \code{Coe} objects
}
\description{
Returns a named (after every level) list that can be lapply-ed and combined. See examples.
}
\examples{
data(bot)
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rename}}, \code{\link{sample_frac}},
  \code{\link{sample_n}}, \code{\link{select}},
  \code{\link{slice}}, \code{\link{transmute}}
}

