% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Coe}
\alias{Coe}
\title{Coe class}
\usage{
Coe(...)
}
\arguments{
\item{...}{anything and, anyway, this function will simply returns a message.
# to see all methods for Coo objects.
methods(class="Coe")}
}
\description{
\code{Coe} class is a "super class" for \link{OutCoe} \link{OpnCoe} and \link{LdkCoe} classes,
matrices of coefficients (in their \code{$coe} slot), along with other informations,
(e.g. an inherited \code{$fac}) obtained with morphometrics methods on
\link{Out}, \link{Opn} and \link{Ldk} objects.
}
\details{
It allows to recycle most of the methods, since both of them apply on
matrices of coefficients. In other words, \code{OutCoe} and \code{OpnCoe}
classes are all, primarily, \code{Coe} objects on which we define generic
\emph{and} specific methods.

More, generally, see \link{OutCoe}, \link{OpnCoe} and \link{LdkCoe}
for documentation on these classes.
}

