% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col.summer}
\alias{Palettes}
\alias{col.autumn}
\alias{col.bw}
\alias{col.cold}
\alias{col.gallus}
\alias{col.gw}
\alias{col.hot}
\alias{col.india}
\alias{col.qual}
\alias{col.sari}
\alias{col.solarized}
\alias{col.spring}
\alias{col.summer}
\alias{col.summer2}
\title{Some color palettes.}
\usage{
col.summer(n)

col.summer2(n)

col.spring(n)

col.autumn(n)

col.solarized(n)

col.gallus(n)

col.qual(n)

col.hot(n)

col.cold(n)

col.sari(n)

col.india(n)

col.bw(n)

col.gw(n)
}
\arguments{
\item{n}{the number of colors to generate from the color palette}
}
\value{
colors (hexadecimal format)
}
\description{
Colors, colors, colors.
}
\note{
\code{col.solarized} is based on Solarized: \url{http://ethanschoonover.com/solarized};
\code{col.div} on ColorBrewer2: \url{http://colorbrewer2.org/}; \code{col.gallus2} is based on
Sron \url{http://www.sron.nl/~pault/colourschemes.pdf}; \code{col.qual}, \code{col.heat},
\code{col.cold} and \code{col.gallus} use the colorspace package.
}
\keyword{Graphics}

