% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{combine}
\alias{combine}
\alias{combine.Ldk}
\alias{combine.Opn}
\alias{combine.OpnCoe}
\alias{combine.Out}
\alias{combine.OutCoe}
\alias{combine.list}
\title{Combine Out objects}
\usage{
combine(...)

\method{combine}{list}(...)

\method{combine}{Out}(...)

\method{combine}{Opn}(...)

\method{combine}{Ldk}(...)

\method{combine}{OutCoe}(...)

\method{combine}{OpnCoe}(...)
}
\arguments{
\item{...}{a list of Out(Coe), Opn(Coe), Ldk objects (but of the same class)}
}
\description{
e.g. after a slicing, either manual or using \link{slice}. Note that on Coo object,
it combines row-wise (ie, merges shapes) ; but on Coe it combines column-wise
(merges coefficients). In the latter case, Coe must ahve the same number of shapes (not
necessarily the same number of coefficients). Also the $fac of the first Coe is retrieved.
A separate version may come at some point.
}
\examples{
data(bot)
w <- subset(bot, type=="whisky")
b <- subset(bot, type=="beer")
combine(w, b)
# or, if you have many levels
bot_s <- slice(bot, type)
}
\seealso{
\link{subset.Coo}
}

