% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{complex}
\alias{complex}
\alias{cpx2coo}
\alias{coo2cpx}
\title{Convert complex to/from cartesian coordinates}
\usage{
cpx2coo(Z)

coo2cpx(coo)
}
\arguments{
\item{Z}{coordinates expressed in the complex form}

\item{coo}{coordinates expressed in the cartesian form}
}
\value{
coordinates expressed in the cartesian/complex form
}
\description{
Convert complex to/from cartesian coordinates
}
\examples{
shapes[4] \%>\%            # from cartesian
    coo_sample(24) \%>\%
    coo2cpx() \%T>\%       # to complex
    cpx2coo()            # and back
}
\seealso{
Other bridges functions: \code{\link{as_df}},
  \code{\link{bridges}}, \code{\link{export}}
}
