% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.arrows}
\alias{coo.arrows}
\title{Plots (lollipop) differences between two configurations}
\usage{
coo.arrows(coo1, coo2, length = 0.1, angle = 20, ...)
}
\arguments{
\item{coo1}{A \code{list} or a \code{matrix} of coordinates.}

\item{coo2}{A \code{list} or a \code{matrix} of coordinates.}

\item{length}{a length for the arrows.}

\item{angle}{an angle for the arrows}

\item{...}{optional parameters to fed \link{arrows}.}
}
\description{
Draws "arrows" between two configurations.
}
\examples{
data(olea)
coo.arrows(coo.sample(olea[3], 50), coo.sample(olea[6], 50))
}
\seealso{
\link{coo.arrows}
}
\keyword{Graphics}

