% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.centdist}
\alias{coo.centdist}
\title{Returns the distance between everypoints and the centroid
For every point of the shape, returns the (centroid-points) distance.}
\usage{
coo.centdist(coo)
}
\arguments{
\item{coo}{a \code{matrix} of \eqn{(x; y)} coordinates.}
}
\value{
a \code{matrix} of \eqn{(x; y)} coordinates.
}
\description{
Returns the distance between everypoints and the centroid
For every point of the shape, returns the (centroid-points) distance.
}
\examples{
data(bot)
b <- coo.sample(bot[1], 64)
d <- coo.centdist(b)
barplot(d)
}
\keyword{coo_descriptors}

