% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.interpolate}
\alias{coo.interpolate}
\title{Interpolates coordinates}
\usage{
coo.interpolate(coo, n)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{n}{an integer, the number fo points to interpolate.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Interpolates n coordinates 'among existing points'between' existing points,
along the perimeter of the coordinates provided and keeping the first point
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.scale(bot))
coo.plot(b)
coo.plot(coo.scale(b))
data(bot)
stack(bot)
stack(coo.interpolate(coo.sample(bot, 12), 120))
coo.plot(bot[1])
coo.plot(coo.interpolate(coo.sample(bot[1], 12), 120))
}
\keyword{ShapeUtilities}

