% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.length}
\alias{coo.length}
\title{Calculates the length of a shape}
\usage{
coo.length(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
the length (in pixels) of the shape
}
\description{
Nothing more than \code{coo.lw(coo)[1]}.
}
\examples{
data(bot)
coo.length(bot[1])
}
\seealso{
\link{coo.lw}, \link{coo.width}
}
\keyword{ShapeDescriptors}

