% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.overlap}
\alias{coo.overlap}
\title{Calculate area overlap between two shapes}
\usage{
coo.overlap(coo1, coo2)
}
\arguments{
\item{coo1}{the first shape}

\item{coo2}{the second shape}
}
\value{
the area of the overlap
}
\description{
Simply calculates (area(coo1) + area(coo2) - area(union(coo1, coo2)))
}
\examples{
data(bot)
b1 <- bot[1]
b2 <- coo.trans(b1, 50)
coo.plot(b1)
coo.draw(b2)
coo.overlap(b1, b2)
}

