\name{coo.rectangularity}
\alias{coo.rectangularity}
\title{Calculates the rectangularity of a shape}
\usage{
coo.rectangularity(coo)
}
\arguments{
  \item{coo}{a matrix or a list of (x,y) coordinates.}
}
\value{
a numeric, the rectangularity.
}
\description{
Calculates the rectangularity of a shape
}
\examples{
data(bot)
coo.rectangularity(bot[4])
}
\keyword{coo_descriptors}

