% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.samplerr}
\alias{coo.samplerr}
\title{Samples coordinates (regular radius)}
\usage{
coo.samplerr(coo, n)
}
\arguments{
\item{coo}{either a \code{matrix} of \eqn{(x; y)} coordinates, or a \link{Coo} object.}

\item{n}{integer, the number of points to sample.}
}
\value{
a \code{matrix} of \eqn{(x; y)} coordinates or an Coo object.
}
\description{
Samples n coordinates with a regular angle.
}
\examples{
data(bot)
stack(bot)
bot <- coo.center(bot)
stack(coo.samplerr(bot, 12))
coo.plot(bot[1])
coo.plot(rr <- coo.samplerr(bot[1], 12))
cpos <- coo.centpos(bot[1])
segments(cpos[1], cpos[2], rr[, 1], rr[, 2])
}
\keyword{coo_utilities}

