% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.scalex}
\alias{coo.scalex}
\alias{coo.scaley}
\title{Scales coordinates in one direction}
\usage{
coo.scalex(coo, k = 1)

coo.scaley(coo, k = 1)
}
\arguments{
\item{coo}{a matrix or a list of coordinates}

\item{k}{scaling factor}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
\code{coo.scalex} applies a scaling a matrix of (x; y) (or a list),
parallel to the x-axis,
\code{coo.scaley} does it parallel to the y-axis.
}
\examples{
data(shapes)
coo <- coo.template(shapes[11])
coo.plot(coo, xlim=c(-1, 1))
coo.draw(coo.scalex(coo, 0.5), border="blue")
coo.draw(coo.scalex(coo, 1.2), border="blue", lty=2)
coo.draw(coo.scaley(coo, 0.5), border="red")
coo.draw(coo.scaley(coo, 1.2), border="red", lty=2)
}
\keyword{ShapeUtilities}

