% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.shearx}
\alias{coo.shearx}
\alias{coo.sheary}
\title{Shears shapes}
\usage{
coo.shearx(coo, k = 1)

coo.sheary(coo, k)
}
\arguments{
\item{coo}{a matrix or a list of coordinates}

\item{k}{shear factor}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
\code{coo.shearx} applies a shear mapping on a matrix of (x; y) coordinates (or a list), parallel
to the x-axis (i.e. x' = x + ky; y' = y + kx). \code{coo.sheary} does it parallel to the y-axis.
}
\examples{
data(shapes)
coo <- coo.template(shapes[11])
coo.plot(coo)
coo.draw(coo.shearx(coo, 0.5), border="blue")
coo.draw(coo.sheary(coo, 0.5), border="red")
}
\keyword{ShapeUtilities}

