% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.slidedirection}
\alias{coo.slidedirection}
\title{Slides coordinates in a particular direction}
\usage{
coo.slidedirection(coo, direction)
}
\arguments{
\item{coo}{a matrix (or a list) or (x; y) coordinates or a Coo object}

\item{direction}{a character among \code{"N"} (by default), \code{"S"}, \code{"E"}, or \code{"W"}.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Shapes are centered and then, according to direction, the point northwards, southwards,
eastwards or westwards the centroid, becomes the first point with \link{coo.slide}.
}
\examples{
data(bot)
b <- coo.rotate(bot[1], pi/6) # dummy example just to make it obvious
coo.plot(b) # not the first point
coo.plot(coo.slidedirection(b, "N"))
coo.plot(coo.slidedirection(b, "E"))
coo.plot(coo.slidedirection(b, "W"))
coo.plot(coo.slidedirection(b, "S"))

# on Coo objects
stack(bot)
stack(coo.slidedirection(bot, "E"))
}
\keyword{ShapeUtilities}

