% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.symmetry}
\alias{coo.symmetry}
\title{Estimates radial symmetry}
\usage{
coo.symmetry(coo, order.max = 12)
}
\arguments{
\item{coo}{a matrix of a shape}

\item{order.max}{the n maximal order (estimates will be calculated for the 2:order.max range)}
}
\value{
a list with $ov (the mean overlapping index), $sd the standard deviation, and $sym the
n orders of radial symmetry
}
\description{
This function implements a simple estimate of radial symmetry occurence,
by calculating overlapping of rotated shapes against reference shapes. For n-order
radia symmetry, it averages the overlap (normalized by the area of the orginal shapes) of
2*pi / (n-1) radians rotated shapes. See references below for a detailed explanation.
}
\examples{
data(bot)
x <- coo.symmetry(bot[1])
w <- barplot(x$ov)
axis(1, at=w, labels=x$sym)
segments(w, x$ov - x$sd, w, x$ov + x$sd)
}
\references{
Rosin, P. L. (2005). Computing global shape measures. In C. H. Chen and P. S. P. Wang (Eds.),
Handbook of Pattern Recognition and Computer Vision (pp. 177-196).
}

