% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.tangle}
\alias{coo.tangle}
\title{Calculates the tangent angle along the perimeter of a shape}
\usage{
coo.tangle(coo)
}
\arguments{
\item{coo}{a matrix of coordinates}
}
\value{
a numeric, the tangent angle along the perimeter
}
\description{
Calculated using complex numbers and returned in radians
minus the first one (modulo 2*pi).
}
\examples{
data(bot)
b <- bot[1]
phi  <- coo.tangle(b)
phi2 <- coo.tangle(coo.smooth(b, 2))
plot(phi, type="l")
plot(phi2, type="l", col="red") # ta is very sensible to noise
}
\seealso{
\link{tfourier}
}

