% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.union}
\alias{coo.union}
\title{Calculate the union between two shapes}
\usage{
coo.union(coo1, coo2)
}
\arguments{
\item{coo1}{the first shape}

\item{coo2}{the second shape}
}
\value{
the area of the overlap
}
\description{
If the two shapes overlaps returns the shape of their union. If not, returns NULL.
}
\examples{
data(bot)
b1 <- bot[1]
ba <- coo.union(b1, coo.trans(b1, 200))
coo.plot(ba)
coo.union(b1, coo.trans(b1, 1e3))
}

