% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.up}
\alias{coo.up}
\title{Retains coordinates with positive y-coordinates}
\usage{
coo.up(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
Useful when shapes are aligned along the x-axis (e.g. because of a
bilateral symmetry) #' and when one wants to retain just the upper side.
}
\examples{
data(bot)
b <- coo.alignxax(bot[1])
coo.plot(b)
coo.draw(coo.up(b), border="red")
}
\keyword{ShapeUtilities}

