% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.width}
\alias{coo.width}
\title{Calculates the width of a shape}
\usage{
coo.width(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
the width (in pixels) of the shape
}
\description{
Nothing more than \code{coo.lw(coo)[2]}.
}
\examples{
data(bot)
coo.width(bot[1])
}
\seealso{
\link{coo.lw}, \link{coo.length}.
}
\keyword{ShapeDescriptors}

