% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{coo2cpx}
\alias{coo2cpx}
\title{Converts cartesian to complex coordinates}
\usage{
coo2cpx(coo)
}
\arguments{
\item{coo}{coordinates expressed in the cartesian form}
}
\value{
coordinates expressed in the complex form
}
\description{
Converts cartesian to complex coordinates
}
\examples{
shapes[4] \%>\% coo_sample(24) \%>\% coo2cpx()
shapes[4] \%>\% coo_sample(24) \%>\% coo2cpx() \%>\% cpx2coo()
}

