% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_align}
\alias{coo_align}
\title{Aligns coordinates}
\usage{
coo_align(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Aligns the coordinates along their longer axis using var-cov matrix and eigen values.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo_align(bot))
coo_plot(b)
coo_plot(coo_align(b))
}
\seealso{
\link{coo_aligncalliper}, \link{coo_alignxax}
}
\keyword{ShapeUtilities}

