% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_aligncalliper}
\alias{coo_aligncalliper}
\title{Aligns shapes along their 'calliper length'}
\usage{
coo_aligncalliper(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
And returns them registered on bookstein coordinates.
See \link{coo_bookstein}.
}
\examples{
\dontrun{
data(bot)
b <- bot[1]
coo_plot(b)
coo_plot(coo_aligncalliper(b))
bot.al <- coo_aligncalliper(bot)
stack(bot.al)
}
}
\seealso{
\link{coo_align}, \link{coo_alignxax}, \link{coo_calliper}
}
\keyword{ShapeUtilities}

