% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_alignminradius}
\alias{coo_alignminradius}
\title{Aligns shapes using their shortest radius}
\usage{
coo_alignminradius(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
And returns them slided with the first coordinate on the east.
May be used as an aligning strategy on shapes with a clear 'invaginate' part.
}
\examples{
\dontrun{
data(hearts)
stack(coo_alignminradius(hearts))
}
}
\seealso{
\link{coo_align}, \link{coo_alignxax}, \link{coo_calliper}
}
\keyword{ShapeUtilities}

