% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_alignxax}
\alias{coo_alignxax}
\title{Aligns shapes along the x-axis}
\usage{
coo_alignxax(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Align the longest axis of a shape along the x-axis.
}
\examples{
\dontrun{
data(bot)
b <- bot[1]
coo_plot(b)
coo_plot(coo_alignxax(b))
}
}
\seealso{
\link{coo_align}, \link{coo_aligncalliper}
}
\keyword{ShapeUtilities}

