% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{coo_arrows}
\alias{coo_arrows}
\title{Plots (lollipop) differences between two configurations}
\usage{
coo_arrows(coo1, coo2, length = coo_centsize(coo1)/15, angle = 20, ...)
}
\arguments{
\item{coo1}{A \code{list} or a \code{matrix} of coordinates.}

\item{coo2}{A \code{list} or a \code{matrix} of coordinates.}

\item{length}{a length for the arrows.}

\item{angle}{an angle for the arrows}

\item{...}{optional parameters to fed \link{arrows}.}
}
\value{
a plot
}
\description{
Draws 'arrows' between two configurations.
}
\examples{
coo_arrows(coo_sample(olea[3], 50), coo_sample(olea[6], 50))
title("Hi there !")
}
\seealso{
Other plotting functions: 
\code{\link{coo_draw}()},
\code{\link{coo_listpanel}()},
\code{\link{coo_lolli}()},
\code{\link{coo_plot}()},
\code{\link{coo_ruban}()},
\code{\link{ldk_chull}()},
\code{\link{ldk_confell}()},
\code{\link{ldk_contour}()},
\code{\link{ldk_labels}()},
\code{\link{ldk_links}()},
\code{\link{plot_devsegments}()},
\code{\link{plot_table}()}
}
\concept{plotting functions}
