% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_centdist}
\alias{coo_centdist}
\title{Returns the distance between everypoints and the centroid
For every point of the shape, returns the (centroid-points) distance.}
\usage{
coo_centdist(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
Returns the distance between everypoints and the centroid
For every point of the shape, returns the (centroid-points) distance.
}
\examples{
data(bot)
b <- coo_sample(bot[1], 64)
d <- coo_centdist(b)
barplot(d)
}
\keyword{ShapeDescriptors}

