% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_center}
\alias{coo_center}
\alias{coo_centre}
\title{Centers coordinates}
\usage{
coo_center(coo)

coo_centre(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a shape centered on the origin. The two functions are strictly equivalent.
}
\examples{
coo_plot(bot[1])
# same as
coo_plot(coo_centre(bot[1]))
# this
coo_plot(coo_center(bot[1]))
# on Coo objects
stack(bot)
stack(coo_center(bot))
}

