% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_centpos}
\alias{coo_centpos}
\title{Returns the position of the centroid}
\usage{
coo_centpos(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
(x; y) coordinates as \code{numeric}.
}
\description{
Returns the (x; y) centroid coordinates of a shape.
}
\examples{
data(bot)
b <- bot[1]
coo_plot(b)
xy <- coo_centpos(b)
points(xy[1], xy[2], cex=2, col='blue')
xy
}
\keyword{ShapeUtilities}

