% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_check}
\alias{coo_check}
\title{Checks 'coo' shapes}
\usage{
coo_check(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a list.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
A simple utility, used internally, mostly in the coo functions and methods.
Returns a matrix of coordinates, when passed with either a list or a \code{matrix} of coordinates.
}
\examples{
#coo_check('Not a shape')
#coo_check(matrix(1:10, ncol=2))
#coo_check(list(x=1:5, y=6:10))
}
\seealso{
\link{ldk_check}
}
\keyword{ShapeUtilities}

