% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_circularity.norm}
\alias{coo_circularity.norm}
\title{Calculates the 'normalized' circularity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_circularity.norm(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the circularity normalized to the unit circle.
}
\description{
Returns the 'circularity', also called compactness
and shape factor, but normalized to the unit circle.
}
\examples{
data(bot)
coo_circularity.norm(bot[1])
}
\keyword{ShapeDescriptors}

