% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_circularity}
\alias{coo_circularity}
\alias{coo_compactness}
\alias{coo_shapefactor}
\title{Calculates the circularity of a shape}
\usage{
coo_circularity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the circularity.
}
\description{
Returns the 'circularity measure'. Also called 'compactness'
and 'shape factor' sometimes.
}
\examples{
data(bot)
coo_circularity(bot[1])
}
\keyword{ShapeDescriptors}

