% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{coo_draw}
\alias{coo_draw}
\title{Adds a shape to the current plot}
\usage{
coo_draw(coo, ...)
}
\arguments{
\item{coo}{a \code{list} or a \code{matrix} of coordinates.}

\item{...}{optional parameters for \link{coo_plot}}
}
\description{
\code{coo_draw} is simply a \link{coo_plot} with \code{plot.new=FALSE}, ie
that adds a shape on the active plot.
}
\examples{
data(bot)
b1 <- bot[4]
b2 <- bot[5]
coo_plot(b1)
coo_draw(b2, border='red') # all coo_plot arguments will work for coo_draw
}

