% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{coo_draw_rads}
\alias{coo_draw_rads}
\title{Draw radii to the current plot}
\usage{
coo_draw_rads(coo, ...)
}
\arguments{
\item{coo}{a shape}

\item{...}{arguments to feed \link{segments}}
}
\description{
Given a shape, all centroid-points radii are drawn using \link{segments}
that can be passed with options
}
\examples{
shp <- shapes[4] \%>\% coo_sample(24) \%T>\% coo_plot
coo_draw_rads(shp, col=col_summer(24))

}

