% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_eccentricity.boundingbox}
\alias{coo_eccentricity.boundingbox}
\title{Calculates the eccentricity (bounding box) of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_eccentricity.boundingbox(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the eccentricity (boundingbox)
}
\description{
Calculated using the width / length ratio. See \link{coo_lw}
}
\examples{
data(bot)
coo_eccentricity.boundingbox(bot[1])
}
\seealso{
\link{coo_eccentricity.eigen}
}
\keyword{ShapeDescriptors}

