% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_elongation}
\alias{coo_elongation}
\title{Calculates the elongation of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_elongation(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the circularity normalized to the unit circle.
}
\description{
Calculates the elongation of a shape
}
\examples{
data(bot)
coo_elongation(bot[1])
}
\keyword{ShapeDescriptors}

