% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_extract}
\alias{coo_extract}
\title{Extract coordinates from a shape}
\usage{
coo_extract(coo, ids)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates or a \link{Coo} object.}

\item{ids}{\code{integer}, the ids of points to sample.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Extract ids coordinates from a single shape or a Coo object.
}
\details{
It probably only make sense for Coo objects with the same number of coordinates
and them being homologous, typically on Ldk.
}
\examples{
b <- bot[1]
stack(bot)
stack(coo_sample(bot, 24))
coo_plot(b)
coo_plot(coo_sample(b, 24))
}
\seealso{
Other sampling functions: \code{\link{coo_interpolate}},
  \code{\link{coo_samplerr}}, \code{\link{coo_sample}}
}

