% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_force2close}
\alias{coo_force2close}
\title{Forces shapes to close}
\usage{
coo_force2close(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
An exotic function that distribute the distance between the first and the last points
of unclosed shapes, so that they become closed. May be useful (?) e.g. for t/rfourier methods
where reconstructed shapes may not be closed.
}
\examples{
data(bot)
b <- coo_sample(bot[1], 64)
b <- b[1:40,]
coo_plot(b)
coo_draw(coo_force2close(b), border='red')
}
\keyword{ShapeUtilities}

