% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_interpolate}
\alias{coo_interpolate}
\title{Interpolates coordinates}
\usage{
coo_interpolate(coo, n)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{n}{code{integer}, the number fo points to interpolate.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Interpolates n coordinates 'among existing points'between' existing points,
along the perimeter of the coordinates provided and keeping the first point
}
\examples{
b <- bot[1]
stack(bot)
stack(coo_scale(bot))
coo_plot(b)
coo_plot(coo_scale(b))
stack(bot)
stack(coo_interpolate(coo_sample(bot, 12), 120))
coo_plot(bot[1])
coo_plot(coo_interpolate(coo_sample(bot[1], 12), 120))
}

