% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_jitter}
\alias{coo_jitter}
\title{Jitters shapes}
\usage{
coo_jitter(coo, ...)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{...}{additional parameter for \link{jitter}}
}
\value{
a \code{matrix} of (x; y) coordinates or a Coo object
}
\description{
A simple wrapper around \link{jitter}.
}
\examples{
b <-bot[1]
coo_plot(b, zoom=0.2)
coo_draw(coo_jitter(b, amount=3), border="red")

# for a Coo example, see \\link{get_pairs}
}
\seealso{
\link{get_pairs}
}

