% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_left}
\alias{coo_left}
\title{Retains coordinates with negative x-coordinates}
\usage{
coo_left(coo, slidegap = FALSE)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates or a \link{Coo} object}

\item{slidegap}{logical whether to apply \link{coo_slidegap} after coo_left}
}
\value{
a \code{matrix} of (x; y) coordinates or a \link{Coo} object (\link{Out} are returned as \link{Opn})
}
\description{
Useful when shapes are aligned along the y-axis (e.g. because of a
bilateral symmetry) and when one wants to retain just the lower side.
}
\note{
When shapes are "sliced" along the y-axis, it usually results on open curves and thus to huge/artefactual
gaps between points neighboring this axis. This is usually solved with \link{coo_slidegap}. See examples there.

Also, when apply a coo_left/right/up/down on an \link{Out} object, you then obtain an \link{Opn} object, which is done
automatically.
}
\examples{
data(bot)
b <- coo_center(bot[1])
coo_plot(b)
coo_draw(coo_left(b), border='red')
}
\seealso{
\link{coo_right}, \link{coo_up}, \link{coo_down}
}

