% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_length}
\alias{coo_length}
\title{Calculates the length of a shape}
\usage{
coo_length(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
the length (in pixels) of the shape
}
\description{
Nothing more than \code{coo_lw(coo)[1]}.
}
\examples{
data(bot)
coo_length(bot[1])
}
\seealso{
\link{coo_lw}, \link{coo_width}
}
\keyword{ShapeDescriptors}

