% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_nb}
\alias{coo_nb}
\title{Returns the number of coordinates}
\usage{
coo_nb(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\description{
A simple wrapper: for a single shape, returns its nrow ; for a Coo object, return sapply + nrow.
}
\examples{
data(bot)
coo_nb(bot[1])
coo_nb(bot)
}
\keyword{ShapeUtilities}

