% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_perim.cum}
\alias{coo_perim.cum}
\title{Calculates the cumulative chrodal distance a shape.}
\usage{
coo_perim.cum(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric} the cumulate sum of chrodal distances
}
\description{
Just a wrapper for cumsum(coo_perim.pts). See \link{coo_perim.pts}.
}
\examples{
data(bot)
b <- coo_sample(bot[1], 24)
coo_perim.cum(b)
}
\keyword{ShapeDescriptors}

