% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_perim.pts}
\alias{coo_perim.pts}
\title{Calculates the chordal distance along a shape.}
\usage{
coo_perim.pts(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric} the distance between every point.
}
\description{
Calculates the euclidean distance between every points of a shape for coo_perim.pts.
The cumulative sum for coo_perim.cum
}
\examples{
data(bot)
b <- coo_sample(bot[1], 24)
coo_perim.pts(b)
}
\keyword{ShapeDescriptors}

