% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_perimcum}
\alias{coo_perimcum}
\title{Calculates the cumulative chrodal distance a shape.}
\usage{
coo_perimcum(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric} the cumulate sum of chrodal distances
}
\description{
Just a wrapper for cumsum(coo_perimpts). See \link{coo_perimpts}.
}
\examples{
data(bot)
b <- coo_sample(bot[1], 24)
coo_perimcum(b)
}
\keyword{ShapeDescriptors}

