% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_perimpts}
\alias{coo_perimpts}
\title{Calculates the chordal distance along a shape.}
\usage{
coo_perimpts(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric} the distance between every point.
}
\description{
Calculates the euclidean distance between every points of a shape for coo_perimpts.
The cumulative sum for coo_perimcum
}
\examples{
data(bot)
b <- coo_sample(bot[1], 24)
coo_perimpts(b)
}
\keyword{ShapeDescriptors}

