% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_rev}
\alias{coo_rev}
\title{Reverses coordinates}
\usage{
coo_rev(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
Returns the reverse suite of coordinates, i.e. change shape's orientation
}
\examples{
data(bot)
b <- coo_sample(bot[1], 4)
b
coo_rev(b)
}
\keyword{ShapeUtilities}

