% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_rotatecenter}
\alias{coo_rotatecenter}
\title{Rotates shapes with a custom center}
\usage{
coo_rotatecenter(coo, theta, center = c(0, 0))
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{theta}{\code{numeric} the angle (in radians) to rotate shapes.}

\item{center}{the (x; y) position of the center}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
rotates a shape of 'theta' angles (in radians) and with a (x; y) 'center'.
}
\examples{
data(bot)
b <- bot[1]
coo_plot(b)
coo_draw(coo_rotatecenter(b, -pi/2, c(200, 200)), border='red')
}
\seealso{
\link{coo_rotate}
}
\keyword{ShapeUtilities}

