% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_shearx}
\alias{coo_shearx}
\alias{coo_sheary}
\title{Shears shapes}
\usage{
coo_shearx(coo, k = 1)

coo_sheary(coo, k)
}
\arguments{
\item{coo}{a matrix or a list of coordinates}

\item{k}{shear factor}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
\code{coo_shearx} applies a shear mapping on a matrix of (x; y) coordinates (or a list), parallel
to the x-axis (i.e. x' = x + ky; y' = y + kx). \code{coo_sheary} does it parallel to the y-axis.
}
\examples{
data(shapes)
coo <- coo_template(shapes[11])
coo_plot(coo)
coo_draw(coo_shearx(coo, 0.5), border="blue")
coo_draw(coo_sheary(coo, 0.5), border="red")
}
\keyword{ShapeUtilities}

