% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slice}
\alias{coo_slice}
\title{Slices shapes between successive coordinates}
\usage{
coo_slice(coo, ids)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates}

\item{ids}{numeric of length >= 2, where to slice the shape}
}
\description{
Takes a shape with n coordinates. When you pass this function with at least
two ids (<= n), the shape will be open on the corresponding coordinates and
slices returned as a list
}
\examples{
data(hearts)
sh <- coo_slice(hearts[1], c(12, 24, 36, 48))
panel(Opn(sh))
}

