% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_slide}
\alias{coo_slide}
\title{Slides coordinates}
\usage{
coo_slide(coo, id1)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{id1}{the id of the point that will become the new first point.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Slides the coordinates so that the id1-th point become the first one.
}
\examples{
data(hearts)
stack(hearts)
stack(coo_slide(hearts, 1))
coo_plot(hearts[4])
coo_plot(coo_slide(hearts[4], 50))
}
\keyword{ShapeUtilities}

