% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slidedirection}
\alias{coo_slidedirection}
\title{Slides coordinates in a particular direction}
\usage{
coo_slidedirection(coo, direction, center, id)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{direction}{\code{character} among \code{'N'} (by default), \code{'S'}, \code{'E'}, or \code{'W'}.}

\item{center}{\code{logical} whether to center or not before sliding}

\item{id}{\code{numeric} whether to return the id of the point or the slided shapes}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Shapes are centered and then, according to direction, the point northwards, southwards,
eastwards or westwards the centroid, becomes the first point with \link{coo_slide}.
}
\examples{
b <- coo_rotate(bot[1], pi/6) # dummy example just to make it obvious
coo_plot(b) # not the first point
coo_plot(coo_slidedirection(b, 'N'))
coo_plot(coo_slidedirection(b, 'E'))
coo_plot(coo_slidedirection(b, 'W'))
coo_plot(coo_slidedirection(b, 'S'))

# on Coo objects
stack(bot)
stack(coo_slidedirection(bot, 'E'))
}
\seealso{
Other sliding functions: \code{\link{coo_slidegap}},
  \code{\link{coo_slide}}
}

