% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slidegap}
\alias{coo_slidegap}
\title{Slides coordinates using the widest gap}
\usage{
coo_slidegap(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates or a \link{Coo} object}
}
\value{
a \code{matrix} of (x; y) coordinates or a \link{Coo} object.
}
\description{
When slicing a shape using two landmarks, or functions such as \link{coo_up},
an open curve is obtained and the rank of points make wrong/artefactual results.
This functions uses the widest gap between points and use the latter as starting
and ending points. Examples are self-speaking.
}
\examples{
data(shapes)
cat <- coo_center(shapes[4])
coo_plot(cat)

# we only retain the bottom of the cat
cat_down <- coo_down(cat, slidegap=FALSE)

# see? the segment on the x-axis coorespond to the widest gap.
coo_plot(cat_down)

# that's what we meant
coo_plot(coo_slidegap(cat_down))
}
\keyword{ShapeUtilities}

