% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_solidity}
\alias{coo_solidity}
\title{Calculates the solidity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_solidity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the solidity of a shape.
}
\description{
Returns the ids of points that define the convex hull of a shape.
}
\examples{
data(bot)
coo_solidity(bot[1])
}
\keyword{ShapeDescriptors}

