% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_tangle}
\alias{coo_tangle}
\title{Calculates the tangent angle along the perimeter of a shape}
\usage{
coo_tangle(coo)
}
\arguments{
\item{coo}{a matrix of coordinates}
}
\value{
a numeric, the tangent angle along the perimeter
}
\description{
Calculated using complex numbers and returned in radians
minus the first one (modulo 2*pi).
}
\examples{
data(bot)
b <- bot[1]
phi  <- coo_tangle(b)
phi2 <- coo_tangle(coo_smooth(b, 2))
plot(phi, type='l')
plot(phi2, type='l', col='red') # ta is very sensible to noise
}
\seealso{
\link{tfourier}
}

