% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_theta.pts}
\alias{coo_theta.pts}
\title{Calculates the angle of every edge of a shape}
\usage{
coo_theta.pts(coo, method = c("atan2", "acos")[1])
}
\arguments{
\item{coo}{a \code{matrix} or a list of (x; y) coordinates.}

\item{method}{one of 'atan2' or 'acos' for a signed or not angle.}
}
\value{
\code{numeric} the angles in radians for every edge.
}
\description{
Returns the angle (in radians) of every edge of a shape,
}
\examples{
data(bot)
b <- coo_sample(bot[1], 64)
coo_theta.pts(b)
}
\keyword{ShapeDescriptors}

