% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_theta3}
\alias{coo_theta3}
\title{Calculate the angle formed by three (x; y) coordinates}
\usage{
coo_theta3(m, method = c("atan2", "acos")[1])
}
\arguments{
\item{m}{a 3x2 \code{matrix} of 3 points (rows) and (x; y) coordinates}

\item{method}{one of 'atan2' or 'acos' for a signed or not angle.}
}
\value{
\code{numeric} the angle in radians.
}
\description{
Returns the angle (in radians) defined by a triplet of points
either signed ('atan2') or not ('acos').
}
\examples{
data(bot)
b <- coo_sample(bot[1], 64)
b <- b[c(1, 14, 24), ]
coo_plot(b)
coo_theta3(b)
coo_theta3(b, method='acos')
}
\keyword{ShapeDescriptors}

