% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_trans}
\alias{coo_trans}
\title{Translates coordinates}
\usage{
coo_trans(coo, x = 0, y = 0)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{x}{\code{numeric} translation along the x-axis.}

\item{y}{\code{numeric} translation along the y-axis.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Translates the coordinates by a 'x' and 'y' value
}
\examples{
coo_plot(bot[1])
coo_plot(coo_trans(bot[1], 50, 100))
# on Coo
stack(bot)
stack(coo_trans(bot, 50, 100))
}

