% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_trim}
\alias{coo_trim}
\title{Trims coordinates from shape}
\usage{
coo_trim(coo, trim = 1)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{trim}{\code{numeric}, the number of coordinates to trim}
}
\description{
Removes \code{trim} coordinates at both ends of a shape, ie
from top and bottom of the shape matrix.
}
\examples{
olea[1] \%>\% coo_sample(12) \%T>\%
   print() \%T>\% ldk_plot() \%>\%
   coo_trim(1) \%T>\% print() \%>\% points(col="red")
}

