% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_unclose}
\alias{coo_unclose}
\title{'Uncloses' shapes}
\usage{
coo_unclose(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo}
object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a unclosed shape from (un)closed shapes. See also \link{coo_close}.
}
\examples{
x <- (matrix(1:10, ncol=2))
x2 <- coo_close(x)
x3 <- coo_unclose(x2)
x
is.closed(x)
x2
is.closed(x2)
x3
is.closed(x3)
}
\seealso{
\link{coo_close}, \link{is.closed}
}
\keyword{ShapeUtilities}

