% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_up}
\alias{coo_up}
\title{Retains coordinates with positive y-coordinates}
\usage{
coo_up(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
Useful when shapes are aligned along the x-axis (e.g. because of a
bilateral symmetry) #' and when one wants to retain just the upper side.
}
\examples{
data(bot)
b <- coo_alignxax(bot[1])
coo_plot(b)
coo_draw(coo_up(b), border='red')
}
\keyword{ShapeUtilities}

