% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_width}
\alias{coo_width}
\title{Calculates the width of a shape}
\usage{
coo_width(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
the width (in pixels) of the shape
}
\description{
Nothing more than \code{coo_lw(coo)[2]}.
}
\examples{
data(bot)
coo_width(bot[1])
}
\seealso{
\link{coo_lw}, \link{coo_length}.
}
\keyword{ShapeDescriptors}

