% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-premodern.R
\name{d}
\alias{d}
\title{A wrapper to calculates euclidean distances between two points}
\usage{
d(x, id1, id2)
}
\arguments{
\item{x}{a Ldk (typically), an Out or a matrix}

\item{id1}{id of the 1st row}

\item{id2}{id of the 2nd row}
}
\value{
numeric
}
\description{
The main advantage over \link{ed} is that it is a method that can
be passed to different objects and used in combination with \link{measure}.
See examples.
}
\note{
On Out objects, we first \link{get_ldk}.
}
\examples{
# single shape
d(wings[1], 1, 4)
# Ldk object
d(wings, 1, 4)
# Out object
d(hearts, 2, 4)
}
\seealso{
if you want all pairwise combinations, see \link{coo_truss}
}
