% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{d2m}
\alias{d2m}
\title{Converts a data.frame of coordinates to a matrix of coordinates.}
\usage{
d2m(d)
}
\arguments{
\item{d}{\code{data.frame} with two columns.}
}
\value{
\code{matrix} of (x; y) coordinates.
}
\description{
Converts a \code{data.frame} with two columns
(colnamed) \code{matrix} of coordinates.
}
\examples{
data(wings)
d <- d2m(wings[1])
d
m <- m2d(d)
m
}
\seealso{
Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{l2a}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2ll}}, \code{\link{m2l}}

Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{l2a}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2ll}}, \code{\link{m2l}}
}
