% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{chaff}
\alias{chaff}
\title{Data: Landmark and semilandmark coordinates on cereal glumes}
\format{An \link{Ldk} object with 50 configuration of landmarks and semi-landmarks (4 partitions)
sampled on cereal glumes}
\source{
Research support was provided by the European Research Council
(Evolutionary Origins of Agriculture (grant no. 269830-EOA)
PI: Glynis Jones, Dept of Archaeology, Sheffield, UK. Data collected by Emily Forster after some discussion
with Andreas Pares.
}
\description{
Data: Landmark and semilandmark coordinates on cereal glumes
}

