% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{charring}
\alias{charring}
\title{Data: Outline coordinates from an experimental charring on cereal grains}
\format{
An \link{Out} object with 18 grains, 3 views on each, for 2 cereal species, charred at
different temperatures for 6 hours (0C (no charring), 230C and 260C).
}
\source{
Research support was provided by the European Research Council
(Evolutionary Origins of Agriculture (grant no. 269830-EOA)
PI: Glynis Jones, Dept of Archaeology, Sheffield, UK. Data collected by Emily Forster.
}
\description{
Data: Outline coordinates from an experimental charring on cereal grains
}
\seealso{
Other datasets: 
\code{\link{apodemus}},
\code{\link{bot}},
\code{\link{chaff}},
\code{\link{flower}},
\code{\link{hearts}},
\code{\link{molars}},
\code{\link{mosquito}},
\code{\link{mouse}},
\code{\link{nsfishes}},
\code{\link{oak}},
\code{\link{olea}},
\code{\link{shapes}},
\code{\link{trilo}},
\code{\link{wings}}
}
\concept{datasets}
