% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-datasets.R
\docType{data}
\name{flower}
\alias{flower}
\title{Data: Measurement of iris flowers}
\format{A \code{TraCoe} object with 150 measurements of 4 variables
(petal + sepal) x (length x width) on 3 species of iris. This dataset is
the classical \link{iris} formatted for Momocs.}
\source{
see link{iris}
}
\description{
Data: Measurement of iris flowers
}
\keyword{Datasets}

