% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{molars}
\alias{molars}
\title{Data: Outline coordinates of 360 molars}
\format{A \link{Out} object containing 79 (x; y) coordinates for 360 molars,
along with their type, their number, and the individual they come from.}
\source{
Corny, J., & Detroit, F. (2014). Technical Note: Anatomic
identification of isolated modern human molars: testing Procrustes aligned
outlines as a standardization procedure for elliptic fourier analysis.
\emph{American Journal of Physical Anthropology}, 153(2), 314-22. doi:10.1002/ajpa.22428
\url{http://onlinelibrary.wiley.com/doi/10.1002/ajpa.22428/abstract}
}
\description{
Courtesy of Julien Corny and Florent Detroit.
}
\keyword{Datasets}

