\docType{data}
\name{mosquito}
\alias{mosquito}
\title{Data: Outline coordinates of 126 mosquito wings.}
\format{An Out object with the 126 mosquito wing outlines outlines
used Rohlf and Archie (1984).}
\source{
Rohlf F, Archie J. 1984. A comparison of Fourier methods
for the description of wing shape in mosquitoes (Diptera:
Culicidae). \emph{Systematic Biology}: 302-317. Arranged
from:
\url{http://life.bio.sunysb.edu/morph/data/RohlfArchieWingOutlines.nts}.
}
\description{
Data: Outline coordinates of 126 mosquito wings.
}
\keyword{datasets}

