% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{Data: Outline coordinates of mouse molars}
\format{A \link{Out} object 64 coordinates of 30 wood molar outlines.}
\source{
Renaud S, Dufour AB, Hardouin EA, Ledevin R, Auffray JC (2015):
Once upon multivariate analyses: When they tell several stories about biological evolution.
\emph{PLoS One} 10:1-18 \url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0132801}
}
\description{
Data: Outline coordinates of mouse molars
}
\seealso{
Other datasets: \code{\link{apodemus}}, \code{\link{bot}},
  \code{\link{chaff}}, \code{\link{charring}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{molars}}, \code{\link{mosquito}},
  \code{\link{nsfishes}}, \code{\link{oak}},
  \code{\link{olea}}, \code{\link{shapes}},
  \code{\link{trilo}}, \code{\link{wings}}
}
