% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{nsfishes}
\alias{nsfishes}
\title{Data: Outline coordinates of North Sea fishes}
\format{A \link{Out} object containing the outlines coordinates for 218 fishes from
the North Sea along with taxonomical cofactors.}
\source{
Caillon F, Frelat R, Mollmann C, Bonhomme V (submitted)
}
\description{
Data: Outline coordinates of North Sea fishes
}
\seealso{
Other datasets: \code{\link{apodemus}}, \code{\link{bot}},
  \code{\link{chaff}}, \code{\link{charring}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{molars}}, \code{\link{mosquito}},
  \code{\link{mouse}}, \code{\link{oak}},
  \code{\link{olea}}, \code{\link{shapes}},
  \code{\link{trilo}}, \code{\link{wings}}
}
