% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{oak}
\alias{oak}
\title{Data: Configuration of landmarks of oak leaves}
\format{A \link{Ldk} object containing 11 (x; y) landmarks  from 176 oak leaves wings, from}
\source{
Viscosi, V., & Cardini, A. (2011). Leaf morphology, taxonomy and
geometric morphometrics: a simplified protocol for beginners.
PloS One, 6(10), e25630. doi:10.1371/journal.pone.0025630
}
\description{
From Viscosi and Cardini (2001).
}
\seealso{
Other datasets: \code{\link{apodemus}}, \code{\link{bot}},
  \code{\link{chaff}}, \code{\link{charring}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{molars}}, \code{\link{mosquito}},
  \code{\link{mouse}}, \code{\link{nsfishes}},
  \code{\link{olea}}, \code{\link{shapes}},
  \code{\link{trilo}}, \code{\link{wings}}
}
