% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{shapes}
\alias{shapes}
\title{Data: Outline coordinates of 210 olive seeds oopen outlines.}
\format{An \link{Out} object with the outline coordinates of some various shapes.}
\source{
Borrowed from default shapes from (c) Adobe Photoshop.
}
\description{
Data: Outline coordinates of 210 olive seeds oopen outlines.
}
\keyword{datasets}

