% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-datasets.R
\docType{data}
\name{trilo}
\alias{trilo}
\title{Data: Outline coordinates of cephalic outlines of trilobite}
\format{A \link{Out} object 64 coordinates of 50 cephalic outlines from different
ontogenetic stages of trilobite.}
\source{
Arranged from: \url{http://folk.uio.no/ohammer/past/outlines.dat}.
The original data included 51 outlines and 5 ontogenetic stages,
but one of them has just a single outline thas has been removed.
}
\description{
Data: Outline coordinates of cephalic outlines of trilobite
}
\keyword{Datasets}

