% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dct.i}
\alias{dct.i}
\title{Investe discrete cosinus transform}
\usage{
dct.i(df, nb.pts = 60)
}
\arguments{
\item{df}{a list with \code{$A} and \code{$B} components, containing harmonic coefficients.}

\item{nb.pts}{numeric the number of pts for the shape reconstruction}
}
\value{
a matrix of (x; y) coordinates
}
\description{
Calculates inverse discrete cosine transforms (see \link{dct}), given a list of A and B harmonic coefficients,
typically such as those produced by \link{dct}.
}
\note{
Only the core functions so far. Will be implemented as an \link{Opn} method soon.
}
\examples{
# dct and inverse dct
data(olea)
o <- olea[1]
o <- coo.bookstein(o)
coo.plot(o)
o.dct <- dct(o, nb.h=12)
o.dct
o.i <- dct.i(o.dct)
o.i <- coo.bookstein(o.i)
coo.draw(o.i, border='red')

o <- olea[1]
h.range <- 2:13
coo <- list()
for (i in seq(along=h.range)){
coo[[i]] <- dct.i(dct(o, nb.h=h.range[i]))}
names(coo) <- paste0('h', h.range)
panel(Opn(coo), borders=col.india(12), names=TRUE)
title('Discrete Cosine Transforms')
}
\references{
\itemize{
 \item Dommergues, C. H., Dommergues, J.-L., & Verrecchia, E. P. (2007).
 The Discrete Cosine Transform, a Fourier-related Method for Morphometric Analysis of Open Contours.
 \emph{Mathematical Geology}, 39(8), 749-763. doi:10.1007/s11004-007-9124-6
 \item Many thanks to Remi Laffont for the translation in R).
}
}
\keyword{discreteCosine}

