% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dct}
\alias{dct}
\title{Discrete cosinus transform}
\usage{
dct(coo, nb.h)
}
\arguments{
\item{coo}{a matrix (or a list) of (x; y) coordinates}

\item{nb.h}{numeric the number of harmonics to calculate}
}
\value{
a list with the following components:
\itemize{
\item A the A harmonic coefficients
\item B the B harmonic coefficients
\item mod the modules of the points
\item arg the arguments of the points
}
}
\description{
Calculates discrete cosine transforms, as introduced by Dommergues and colleagues, on
a shape (mainly open outlines).
}
\note{
Only the core functions so far. Will be implemented as an \link{Opn} method soon.
}
\examples{
# dct and inverse dct
data(olea)
o <- olea[1]
o <- coo.bookstein(o)
coo.plot(o)
o.dct <- dct(o, nb.h=12)
o.dct
o.i <- dct.i(o.dct)
o.i <- coo.bookstein(o.i)
coo.draw(o.i, border="red")

#future hqual
o <- olea[1]
h.range <- 2:13
coo <- list()
for (i in seq(along=h.range)){
coo[[i]] <- dct.i(dct(o, nb.h=h.range[i]))}
names(coo) <- paste0("h", h.range)
panel(Opn(coo), borders=col.india(12), names=TRUE)
title("Discrete Cosine Transforms")
}
\references{
\itemize{
 \item Dommergues, C. H., Dommergues, J.-L., & Verrecchia, E. P. (2007).
 The Discrete Cosine Transform, a Fourier-related Method for Morphometric Analysis of Open Contours.
 \emph{Mathematical Geology}, 39(8), 749-763. doi:10.1007/s11004-007-9124-6
 \item Many thanks to Remi Laffont for the translation in R).
}
}
\keyword{discreteCosine}

