% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dct_shape}
\alias{dct_shape}
\title{Calculates and draws 'dct' shapes}
\usage{
dct_shape(A, B, nb.h, nb.pts = 60, alpha = 2, plot = TRUE)
}
\arguments{
\item{A}{vector of harmonic coefficients}

\item{B}{vector of harmonic coefficients}

\item{nb.h}{if \code{A} and/or \code{B} are not provided,
the number of harmonics to generate}

\item{nb.pts}{if \code{A} and/or \code{B} are not provided,
the number of points to use to reconstruct the shapes}

\item{alpha}{tThe power coefficient associated with the (usually decreasing)
amplitude of the harmonic coefficients (see \link{efourier_shape})}

\item{plot}{logical whether to plot the shape}
}
\description{
Calculates shapes based on 'Discrete cosine transforms' given harmonic coefficients
(see \link{dct}) or can generate some random 'dct' shapes.
Mainly intended to generate shapes and/or to understand how dct works.
}
\examples{
# some signatures
panel(coo_align(Opn(replicate(48, dct_shape(alpha=0.5, nb.h=6)))))
# some worms
panel(coo_align(Opn(replicate(48, dct_shape(alpha=2, nb.h=6)))))
}

