% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-ldk-utilities.R
\name{def_ldk_tips}
\alias{def_ldk_tips}
\title{Define tips as new landmarks}
\usage{
def_ldk_tips(coo)
}
\arguments{
\item{coo}{\code{Opn} object}
}
\value{
a Momocs object of same class
}
\description{
On \link{Opn} objects, this can be used before \link{coo_slice}. See examples.
}
\note{
any existing ldk will be preserved.
}
\examples{
is_ldk(olea) # no ldk for olea
olea \%>\%
slice(1:3) \%>\% #for the sake of speed
def_ldk_tips \%>\%
def_ldk_angle(3*pi/4) \%>\% def_ldk_angle(pi/4) \%T>\% stack \%>\%
coo_slice(ldk=1:4) -> oleas
stack(oleas[[1]])
stack(oleas[[2]]) # etc.
}
