% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-ldk-utilities.R
\name{def_links}
\alias{def_links}
\title{An utility to define links between landmarks}
\usage{
def_links(x, nb.ldk)
}
\arguments{
\item{x}{Ldk, matric or array}

\item{nb.ldk}{numeric the iterative procedure is stopped when the
user click on the top of the graphical window.}
}
\description{
Works on Ldk objects, on 2cols matrices, 3dim arrays (msshapes turns it into a matrix).
}
\examples{
\dontrun{
data(wings)
wm <- mshapes(wings)
links <- def_links(wm, 3) # click to define pairs of landmarks
ldk_links(wm, links)
}
}

