% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dev.segments}
\alias{dev.segments}
\title{Draws colored segments from a matrix of coordinates.}
\usage{
dev.segments(coo, cols, lwd = 1)
}
\arguments{
\item{coo}{A matrix of coordinates.}

\item{cols}{A vector of color of \code{length = nrow(coo)}.}

\item{lwd}{The \code{lwd} to use for drawing segments.}
}
\description{
Given a matrix of (x; y) coordinates, draws segments between every points
defined by the row of the matrix and uses a color to display an information.
}
\examples{
# we load some data
data(bot)
guinness <- coo.sample(bot[9], 100)

# we calculate the diff between 48 harm and one with 6 harm.
out.6    <- efourier.i(efourier(guinness, nb.h=6), nb.pts=120)

# we calculate deviations, you can also try 'edm'
dev <- edm.nearest(out.6, guinness) / coo.centsize(out.6)

# we prepare the color scale
d.cut <- cut(dev, breaks=20, labels=FALSE, include.lowest=TRUE)
cols  <- paste0(col.summer(20)[d.cut], "CC")

# we draw the results
coo.plot(guinness, main="Guiness fitted with 6 harm.", points=FALSE)
par(xpd=NA)
dev.segments(out.6, cols=cols, lwd=4)
coo.draw(out.6, lty=2, points=FALSE, col=NA)
par(xpd=FALSE)
}
\keyword{graphics}

