% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{dissolve}
\alias{dissolve}
\title{Uncombine Coe objects}
\usage{
dissolve(x, retain)
}
\arguments{
\item{x}{a Coe object}

\item{retain}{the partition id to retain. Or their name if the partitions are named
(see x$method) eg after a chop}
}
\description{
After a combine for instance. But note that the \code{$fac} slot may be wrong since
combine...well combines... this \code{$fac}
}
\examples{
data(bot)
w <- filter(bot, type=="whisky")
b <- filter(bot, type=="beer")
wf <- efourier(w, 10)
bf <- efourier(b, 10)
wbf <- combine(wf, bf)
dissolve(wbf, 1)
dissolve(wbf, 2)
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{chop}}, \code{\link{combine}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rename}}, \code{\link{sample_frac}},
  \code{\link{sample_n}}, \code{\link{select}},
  \code{\link{slice}}, \code{\link{transmute}}
}

