% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fgProcrustes}
\alias{fgProcrustes}
\title{Full Generalized Procrustes alignment between shapes}
\usage{
fgProcrustes(x, tol, verbose)
}
\arguments{
\item{x}{an array, a list of configurations, or an Out or on Ldk object}

\item{tol}{numeric when to stop iterations}

\item{verbose}{logical whether to print outputs}
}
\value{
a list with components
\enumerate{
\item rotated: Array of superimposed configurations.
\item iterationnumber: Number of iterations.
\item Q: Convergence criterion.
\item Qi: Full list of Q.
\item Qd: Difference between succesive Q.
\item interproc.dist: Minimal sum of squared norms of pairwise differences between
all shapes in the superimposed sample.
\item mshape: Mean shape configuration.
\item cent.size: Vector of centroid sizes.
} or and Ldk or and Out object.
}
\description{
Directly borrowed from Claude 2008 (fgpa2).
}
\note{
Slightly less optimized than procGPA in the shapes package (~20% on my machine).
}
\references{
Claude, J. (2008). Morphometrics with R. Analysis (p. 316). Springer.
}

