% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter}
\alias{filter}
\alias{filter.Coe}
\alias{filter.Coo}
\alias{filter.PCA}
\alias{filter.default}
\title{Filter (ala dplyr) rows with matching conditions, using the $fac slot}
\usage{
filter(.data, ...)

\method{filter}{default}(.data, ...)

\method{filter}{Coo}(.data, ...)

\method{filter}{Coe}(.data, ...)

\method{filter}{PCA}(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo} objects}

\item{...}{Logical conditions}
}
\value{
a Coo object
}
\description{
Extends this dplyr verb to \link{Coo} objects. See examples.
}
\details{
dplyr verbs are maintained.
}
\note{
this method is quite experimental, check the result and do not hesitate to report a bug.
}
\examples{
data(bot)
filter(bot, type=="whisky")
}
\seealso{
\link{select.Coo}
}

