% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{get_cur_binded}
\alias{get_cur_binded}
\title{Retrieves semi-landmarks coordinates from Ldk objects}
\usage{
get_cur_binded(Ldk)
}
\arguments{
\item{Ldk}{an \link{Ldk} object}
}
\value{
a list of matrices
}
\description{
Binds groups (if any) of landmarks from the \code{$cur}
on \link{Ldk} objects
}

