% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{get_curcoo_binded}
\alias{get_curcoo_binded}
\title{Binds landmarks and semi-landmarks coordinates from Ldk objects}
\usage{
get_curcoo_binded(x)
}
\arguments{
\item{Ldk}{an \link{Ldk} object}
}
\value{
a list of matrices
}
\description{
Binds \code{$coo} with groups (if any) of landmarks from the \code{$cur}
on \link{Ldk} objects
}
\seealso{
Other get_cur functions: \code{\link{get_cur_binded}}
}

