% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{get_slidings}
\alias{get_slidings}
\title{Extract sliding landmarks coordinates}
\usage{
get_slidings(Coo, partition)
}
\arguments{
\item{Coo}{an Ldk object}

\item{partition}{numeric which one(s) to get.}
}
\description{
From an \link{Ldk} object.
}
\examples{
#waiting for chaff todo

}

