% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getLandmarks}
\alias{getLandmarks}
\title{Retrieve landmarks coordinates from Opn and Out objects}
\usage{
getLandmarks(Coo)
}
\arguments{
\item{Coo}{a Coo object, either Out or Opn}
}
\value{
an array of coordinates X (x; y) coordinates X number of shapes.
}
\description{
In \link{Out} and \link{Opn} classes, landmarks (if any) are stored as
row indices. This methods allows to retrieve the corresponding (x; y) coordinates.
}
\examples{
data(hearts)
ldk.h <- getLandmarks(hearts)
stack(Ldk(a2l(ldk.h)))
ldk.h
}
\seealso{
\link{defLandmarks}, \link{gProcrustes}
}
\keyword{Ldk}
\keyword{Opn}
\keyword{Out}

