% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gProcrustes}
\alias{gProcrustes}
\title{Performs a general Procrustes alignment on Opn and Out objects}
\usage{
gProcrustes(Coo, ...)
}
\arguments{
\item{Coo}{a Coo object (only Out so far)}

\item{...}{additional arguments to be passed to procGPA}
}
\value{
a procGPA-aligned Coo object
}
\description{
It relies on the $ldk slot and on the procGPA function in the shapes package.
}
\examples{
\dontrun{
data(hearts)
stack(hearts) # raw hearts
hearts2 <- gProcrustes(hearts)
stack(hearts2) #procGPA-aligned hearts
hearts3 <- coo.bookstein(hearts, 2, 4)
stack(hearts3) # bookstein baseline
hearts4 <- coo.slide(hearts3, 2)
stack(hearts4) #bookstein baseline + first point on the second landmarks
}
}
\seealso{
\link{defLandmarks} to define landmarks on your shapes and
\link{coo.bookstein} for another "alignment" approach. \link{procGPA} for the
core function. Thanks to Dryden et al. !
}
\keyword{Ldk}
\keyword{Opn}
\keyword{Out}

