% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coe.R
\name{hcontrib.OutCoe}
\alias{hcontrib.OutCoe}
\title{@rdname harm.contrib
 @export}
\usage{
\method{hcontrib}{OutCoe}(Coe, id, harm.r, amp.r = c(0, 0.5, 1, 2, 5, 10),
  main = "Harmonic contribution to shape", xlab = "Harmonic rank",
  ylab = "Amplification factor", ...)
}
\description{
@rdname harm.contrib
 @export
}

