\docType{data}
\name{hearts}
\alias{hearts}
\title{Outline coordinates of 240 hand-drawn hearts}
\format{An Out object with the outline coordinates of 240 hand-drawn hearts
by 8 different persons, with 4 landmarks.}
\source{
We thank the fellows of the Ecology Department of the
French Institute of Pondicherry that drawn the hearts, that
then have been smoothed, scaled, centered, and reduced to
80 coordinates per outline.
}
\description{
Outline coordinates of 240 hand-drawn hearts
}
\keyword{datasets}

