% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/babel-import.R
\name{img_plot}
\alias{img_plot}
\alias{img_plot0}
\title{Plots a .jpg image}
\usage{
img_plot(img)

img_plot0(img)
}
\arguments{
\item{img}{a matrix of an image, such as those obtained with \link{readJPEG}.}
}
\description{
A very simple image plotter. If provided with a path,
read the .jpg and plots it. If not provided with an imagematrix, will
ask you to choose interactively a \code{.jpeg} image.
}
\details{
\code{img_plot} is used in import functions such as \link{import_jpg1};
\code{img_plot0} does the same job but preserves the \code{par} and plots axes.
}
\keyword{Import}

