% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/babel-import.R
\name{import_StereoMorph_curve1}
\alias{import_StereoMorph_curve}
\alias{import_StereoMorph_curve1}
\alias{import_StereoMorph_ldk}
\alias{import_StereoMorph_ldk1}
\title{Import files creates by StereoMorph into Momocs}
\usage{
import_StereoMorph_curve1(path)

import_StereoMorph_curve(path, names)

import_StereoMorph_ldk1(path)

import_StereoMorph_ldk(path, names)
}
\arguments{
\item{path}{toward a single file or a folder containing \code{.txt} files produced by StereoMorph}

\item{names}{to feed \link{lf_structure}}
}
\description{
Helps to read \code{.txt} files created by StereoMorph into (x; y) coordinates
or Momocs objects. Can be applied to 'curves' or 'ldk' text files.
}
\details{
*1 functions import a single \code{.txt} file. Their counterpart (no '1')
work when path indicates the folder, i.e. 'curves' or 'ldk'. They then return a list
of \link{Opn} or \link{Ldk} objects, respectively. Please do not hesitate to contact me
should you have a particular case or need something.
}

