% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coo.R
\name{inspect}
\alias{inspect}
\title{Graphical inspection of shapes}
\usage{
inspect(x, id, ...)
}
\arguments{
\item{x}{the \link{Coo} object}

\item{id}{the id of the shape to plot, if not provided a
random shape is plotted. If passed with \code{'all'} all shapes are plotted,
one by one.}

\item{...}{further arguments to be passed to \link{coo_plot}}
}
\description{
Allows to plot shapes, individually, for \link{Coo} (\link{Out}, \link{Opn} or \link{Ldk}) objects.
}
\examples{
\dontrun{
inspect(bot, 5)
inspect(bot)
inspect(bot, 5, pch=3, points=TRUE) # an example of '...' use
}
}
\seealso{
Other Coo_graphics: \code{\link{panel}},
  \code{\link{stack.Coo}}
}
