% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{is.Coo}
\alias{is.Coo}
\title{Various class/component testers}
\usage{
is.Coo(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE/FALSE
}
\description{
Class testers test if any of the classes of an object is of a given class. For instance
is.PCA on a PCA object (both 'PCA' and 'prcomp') will return TRUE.
Component testers check if a particular component (eg $fac, etc.) is present.
}
\examples{
data(bot)
is.Coo(bot)
is.Out(bot)
is.Ldk(bot)
}

