% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{is.Coo}
\alias{is.Coe}
\alias{is.Coo}
\alias{is.LDA}
\alias{is.Ldk}
\alias{is.LdkCoe}
\alias{is.Opn}
\alias{is.OpnCoe}
\alias{is.Out}
\alias{is.OutCoe}
\alias{is.PCA}
\alias{is.cur}
\alias{is.fac}
\alias{is.ldk}
\alias{is.links}
\title{Tests if an object is of a given class or has a particular component}
\usage{
is.Coo(x)

is.PCA(x)

is.LDA(x)

is.Out(x)

is.Opn(x)

is.Ldk(x)

is.Coe(x)

is.OutCoe(x)

is.OpnCoe(x)

is.LdkCoe(x)

is.fac(x)

is.ldk(x)

is.cur(x)

is.links(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE/FALSE
}
\description{
Class testers test if any of the classes of an object is of a given class. For instance
is.PCA on a PCA object (both 'PCA' and 'prcomp') will return TRUE.
Component testers check if a particular component (eg $fac, etc.) is present.
}
\examples{
data(bot)
is.Coo(bot)
is.Out(bot)
is.Ldk(bot)
}

