% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-helpers.R
\name{is}
\alias{is}
\alias{is_Coo}
\alias{is_PCA}
\alias{is_LDA}
\alias{is_Out}
\alias{is_Opn}
\alias{is_Ldk}
\alias{is_Coe}
\alias{is_OutCoe}
\alias{is_OpnCoe}
\alias{is_LdkCoe}
\alias{is_TraCoe}
\alias{is_shp}
\alias{is_fac}
\alias{is_ldk}
\alias{is_slidings}
\alias{is_links}
\title{Class and component testers}
\usage{
is_Coo(x)

is_PCA(x)

is_LDA(x)

is_Out(x)

is_Opn(x)

is_Ldk(x)

is_Coe(x)

is_OutCoe(x)

is_OpnCoe(x)

is_LdkCoe(x)

is_TraCoe(x)

is_shp(x)

is_fac(x)

is_ldk(x)

is_slidings(x)

is_links(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
\code{logical}
}
\description{
Class testers test if any of the classes of an object \emph{is} of a given class. For instance
\code{is.PCA} on a \link{PCA} object (of classes \code{PCA} and \code{prcomp}) will return \code{TRUE}.
Component testers check if \emph{there_is} a particular component (eg \code{$fac}, etc.) in an object.
}
\examples{
is_Coo(bot)
is_Out(bot)
is_Ldk(bot)
is_ldk(hearts) # mind the capitals!
}
