% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mult-KMEANS.R
\name{KMEANS}
\alias{KMEANS}
\title{k-means}
\usage{
KMEANS(x, ...)
}
\arguments{
\item{x}{PCA object}

\item{...}{additional arguments to be passed to \link{kmeans}}

\item{centers}{numeric number of centers}

\item{nax}{numeric the range of PC components to use (1:2 by default)}
}
\value{
the same thing as \link{kmeans}
}
\description{
A very basic and prototypic implementation of k-means
}
\examples{
data(bot)
bp <- PCA(efourier(bot, 10))
KMEANS(bp, 2)
}

