% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grindr-layers.R
\name{layerize}
\alias{layerize}
\title{Prepare for layers}
\usage{
layerize(x, ...)
}
\arguments{
\item{x}{multivariate object}

\item{f}{to feed \link{fac_dispatcher}}

\item{axes}{to show (default: \code{c(1, 2)})}

\item{palette}{a \link{palette} (default: \code{pal_qual})}
}
\value{
a list to send to \link{layers}
}
\description{
Typically on multivariate object (eg \link{PCA}) before layers.
Used internally by \link{plot_PCA} and friends
}
\seealso{
Other grindr: \code{\link{drawers}}, \code{\link{layers}},
  \code{\link{papers}}, \code{\link{pile}},
  \code{\link{plot_PCA}}
}
