% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lda2shp.efourier}
\alias{lda2shp.efourier}
\title{Calculates shapes from LD plane: efourier}
\usage{
lda2shp.efourier(pos, rot, mshape, amp.shp = 1, pts.shp = 60)
}
\arguments{
\item{pos}{the position on two PC axis}

\item{rot}{the (unstardized) loadings}

\item{mshape}{the meanshape}

\item{amp.shp}{amplification factor for the shape deformation}

\item{pts.shp}{number of points to reconstruct the shape}
}
\description{
Calculates shapes from LD plane: efourier
}

