% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ldk.confell}
\alias{ldk.confell}
\title{Draws confidence ellipses for landmark positions}
\usage{
ldk.confell(ldk, conf = 0.5, col = "grey40", ell.lty = 3, ax = TRUE,
  ax.lty = 2)
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{conf}{the confidence level (normal quantile)}

\item{col}{the color for the ellipse}

\item{ell.lty}{an lty for the ellipse}

\item{ax}{logical whether to draw ellipses axes}

\item{ax.lty}{an lty for ellipses axes}
}
\description{
todo
}

