% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ldk.contour}
\alias{ldk.contour}
\title{Draws kernel density contours around landmark positions}
\usage{
ldk.contour(ldk, nlevels = 5, grid.nb = 50, col = "grey60")
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{nlevels}{the number of contour lines}

\item{grid.nb}{the grid.nb}

\item{col}{a color for drawing the contour lines}
}
\description{
using \link{kde2d} #todo
}
\seealso{
\link{kde2d}
}

