% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ldk.labels}
\alias{ldk.labels}
\title{Add landmarks labels}
\usage{
ldk.labels(ldk, d = 0.1, cex = 2/3, ...)
}
\arguments{
\item{ldk}{a matrix of (x; y) coordinates: where to plot the labels}

\item{d}{how far from the coordinates, on a (centroid-landmark) segment}

\item{cex}{the cex for the label}

\item{...}{additional parameters to fed \link{text}}
}
\description{
Add landmarks labels
}
\examples{
data(wings)
coo.plot(wings[1])
ldk.labels(wings[1])
# closer and smaller
coo.plot(wings[1])
ldk.labels(wings[1], d=0.05, cex=0.5)
}
\keyword{Graphics}

