% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Ldk}
\alias{Ldk}
\title{Builds an Ldk object}
\usage{
Ldk(coo.list, links = matrix(), fac = data.frame())
}
\arguments{
\item{coo.list}{\code{list} of matrices of \eqn{(x; y)} coordinates}

\item{links}{a matrix of "links" between landmarks, mainly for plotting}

\item{fac}{(optionnal) a \code{data.frame} of factors,
specifying the grouping structure}
}
\value{
an \code{Ldk} object
}
\description{
In Momocs, \code{Ldk} classes objects are wrapping around
lists of configurations of landmarks, along with other informations,
on which generic methods such as plotting methods (e.g. \link{stack})
and specific methods (e.g. todo-Procrustes can be applied.
 \code{Ldk} objects are primarily \code{\link{Coo}} objects.
}
\details{
These methods can be applied on \code{Ldk} objects:
\enumerate{
\item Handling: \code{subset, coo.} family;
\item Graphics: \code{plot, stack, panel};
\item Morpho: todo
}
}
\seealso{
\link{Coo}, \link{Out}, link{Opn}.
}
\keyword{Ldk}

