% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{ldk_chull}
\alias{ldk_chull}
\title{Draws convex hulls around landmark positions}
\usage{
ldk_chull(ldk, col = "grey40", lty = 1)
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{col}{a color for drawing the convex hull}

\item{lty}{an lty for drawing the convex hulls}
}
\description{
A wrapper that uses \link{coo_chull}
}
\examples{
data(wings)
coo_plot(mshapes(wings))
ldk_chull(wings$coo)
}
\seealso{
\link{coo_chull}, \link{chull}, \link{ldk_confell}, \link{ldk_contour}
}
\keyword{Graphics}

