% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{ldk_contour}
\alias{ldk_contour}
\title{Draws kernel density contours around landmark}
\usage{
ldk_contour(ldk, nlevels = 5, grid.nb = 50, col = "grey60")
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{nlevels}{the number of contour lines}

\item{grid.nb}{the grid.nb}

\item{col}{a color for drawing the contour lines}
}
\description{
Using \link{kde2d} in the MASS package.
}
\examples{
data(wings)
coo_plot(mshapes(wings))
ldk_contour(wings$coo)
}
\seealso{
\link{kde2d}, \link{ldk_confell}, \link{ldk_chull}
}
\keyword{Graphics}

