% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ldk_labels}
\alias{ldk_labels}
\title{Add landmarks labels}
\usage{
ldk_labels(ldk, d = 0.05, cex = 2/3, ...)
}
\arguments{
\item{ldk}{a matrix of (x; y) coordinates: where to plot the labels}

\item{d}{how far from the coordinates, on a (centroid-landmark) segment}

\item{cex}{the cex for the label}

\item{...}{additional parameters to fed \link{text}}
}
\description{
Add landmarks labels
}
\examples{
data(wings)
coo_plot(wings[1])
ldk_labels(wings[1])
# closer and smaller
coo_plot(wings[1])
ldk_labels(wings[1], d=0.05, cex=0.5)
}
\keyword{Graphics}

