% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{ldk_links}
\alias{ldk_links}
\title{Draws links between landmarks}
\usage{
ldk_links(ldk, links, ...)
}
\arguments{
\item{ldk}{a matrix of (x; y) coordinates}

\item{links}{a matrix of links. On the first column the starting-id,
on the second column the ending-id (id= the number of the coordinate)}

\item{...}{additional parameters to fed \link{segments}}
}
\description{
Cosmetics only but useful to visualize shape variation.
}
\seealso{
Other ldk plotters: \code{\link{ldk_chull}},
  \code{\link{ldk_confell}}, \code{\link{ldk_contour}},
  \code{\link{ldk_labels}}

Other plotting functions: \code{\link{Ntable}},
  \code{\link{conf_ell}}, \code{\link{coo_arrows}},
  \code{\link{coo_draw}}, \code{\link{coo_listpanel}},
  \code{\link{coo_lolli}}, \code{\link{coo_plot}},
  \code{\link{coo_ruban}}, \code{\link{ldk_chull}},
  \code{\link{ldk_confell}}, \code{\link{ldk_contour}},
  \code{\link{ldk_labels}}, \code{\link{plot_devsegments}}
}

