% Generated by roxygen2 (4.0.2): do not edit by hand
\name{links.all}
\alias{links.all}
\title{Create links (all pariwise combinations) between landmarks}
\usage{
links.all(coo)
}
\arguments{
\item{coo}{a matrix (or a list) of (x; y) coordinates}
}
\value{
a matrix that can be passed to \link{ldk.links}, etc. The columns
are the row ids of the original shape.
}
\description{
Create links (all pariwise combinations) between landmarks
}
\examples{
data(wings)
w <- wings[1]
coo_plot(w)
links <- links.all(w)
ldk.links(w, links)
}

