% Generated by roxygen2 (4.0.2): do not edit by hand
\name{links.delaunay}
\alias{links.delaunay}
\title{Create links (Delaunay triangulation) between landmarks}
\usage{
links.delaunay(coo)
}
\arguments{
\item{coo}{a matrix (or a list) of (x; y) coordinates}
}
\value{
a matrix that can be passed to \link{ldk.links}, etc. The columns
are the row ids of the original shape.
}
\description{
Create links (Delaunay triangulation) between landmarks
}
\details{
uses \link{delaunayn} in the \code{geometry} package.
}
\examples{
data(wings)
w <- wings[1]
coo_plot(w, poly=FALSE)
links <- links.delaunay(w)
ldk.links(w, links)
}

