% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-misc.R
\name{list_methods_for_this_class}
\alias{list_methods_for_this_class}
\alias{list_classes_for_this_method}
\title{List available methods for any object and other classes for a method}
\usage{
list_methods_for_this_class(x)

list_classes_for_this_method(x)
}
\arguments{
\item{x}{any object}
}
\description{
Given any object
returns all available methods for its class, or given a method
list all other supported classes
}
\examples{
#methods for data.frame
list_methods_for_this_class(iris)

#methods for Coo objects
list_methods_for_this_class(bot)

#methods for Coe objects
list_methods_for_this_class(bot \%>\% efourier)

#methods for PCA objects
list_methods_for_this_class(bot \%>\% efourier \%>\% PCA)

# classes supported for
list_classes_for_this_method(PCAAA) # nothing
list_classes_for_this_method(PCA)
list_classes_for_this_method(sum) # nothing
}
