% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{m2d}
\alias{m2d}
\title{Converts a matrix of coordinates to a two-columns data.frame}
\usage{
m2d(m)
}
\arguments{
\item{m}{\code{matrix} of (x; y) coordinates.}
}
\value{
\code{data.frame} with two columns.
}
\description{
Converts a \code{matrix} of coordinates to a (colnamed) \code{data.frame}
with two columns.
}
\examples{
data(wings)
d <- d2m(wings[1])
d
m <- m2d(d)
m
}
\seealso{
Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{d2m}}, \code{\link{l2a}}, \code{\link{l2m}},
  \code{\link{m2a}}, \code{\link{m2ll}}, \code{\link{m2l}}
}
