% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{m2ll}
\alias{m2ll}
\title{Converts a matrix of coordinates into a list of matrices}
\usage{
m2ll(m, index = NULL)
}
\arguments{
\item{m}{\code{matrix}, typically of (x; y) coordinates}

\item{index}{\code{numeric}, the number of coordinates for every slice.}
}
\description{
Used internally to hanle coo and cur in \code{Ldk} objects but may be
useful elsewhere
}
\examples{
m2ll(wings[1], c(6, 4, 3, 5))
}
\seealso{
Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{d2m}}, \code{\link{l2a}}, \code{\link{l2m}},
  \code{\link{m2a}}, \code{\link{m2d}}, \code{\link{m2l}}
}
