% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Manova}
\alias{Manova}
\alias{Manova.OpnCoe}
\alias{Manova.OutCoe}
\title{Multivariate analysis of variance on Coe objects}
\usage{
Manova(x, fac, test = "Hotelling", retain, drop)

\method{Manova}{OpnCoe}(x, fac, test = "Hotelling", retain, drop)

\method{Manova}{OutCoe}(x, fac, test = "Hotelling", retain, drop)
}
\arguments{
\item{x}{a \link{Coe} object}

\item{fac}{a name of a colum in the \code{$fac} slot, or its id}

\item{test}{a test for \link{manova} (\code{"Hotelling"} by default)}

\item{retain}{how many harmonics (or polynomials) to retain}

\item{drop}{how many harmonics (or polynomials) to drop}
}
\value{
a list of matrices of (x,y) coordinates.
}
\description{
Performs multivariate analysis of variance on \link{Coe} objects.
}
\details{
For outlines, checks if the matrix of coefficients is of full rank, and if not removes the
higher order harmonics (for Out objects).

If OutCoe objects have been normalized, the first harmonic will be removed with a message.

If \link{removeAsymmetric} or \link{removeSymmetric}
have been used on OutCoe object, the zero-ed harmonics will be removed with a message.
}
\note{
Needs a review #todo.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
Manova(bot.f, "type")

data(olea)
op <- rawPolynomials(olea, 5)
Manova(op, "domes")
}
\keyword{Multivariate}

