% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ManovaPW}
\alias{ManovaPW}
\alias{ManovaPW.Coe}
\title{Pairwise Multitlivariate analyses of variance}
\usage{
ManovaPW(Coe, fac, verbose, ...)

\method{ManovaPW}{Coe}(Coe, fac, verbose = FALSE, ...)
}
\arguments{
\item{Coe}{a Coe object}

\item{fac}{a name (or its id) of a grouping factor in \code{$fac} or a factor}

\item{verbose}{to feed \link{Manova}}

\item{...}{more arguments to feed \link{Manova}}
}
\value{
a list with the following components is returned (invisibly because $manovas
may be very long, see examples):
\itemize{
 \item manovas a list containing all the raw manovas
 \item summary a matrix with all important statists
 \item stars.tab a table with 'significance star', discutable but useful:
 '***' if Pr(>F) < 0.001; '**' of < 0.01; '*' if < 0.05; "." if < 0.10 and "-" if above.
}
}
\description{
A wrapper for pairwise \link{Manova}s on \link{Coe} object.
}
\note{
If the fac passed has only two levels, there is only pair and it is
equivalent to \link{Manova}.
}
\examples{
data(bot)
# we create a fake factor with 4 levels
bot$fac$fake <- factor(rep(letters[1:4], each=10))
bot.f <- eFourier(bot, 8)
ManovaPW(bot.f, "fake") # or ManovaPW(bot.f, 2)

# an example on open outlines
data(olea)
op <- rawPolynomials(olea)
ManovaPW(op, "domes")
# to get the results
res <- ManovaPW(op, "domes")
res$manovas
res$stars.tab
res$summary
}
\seealso{
\link{Manova}
}
\keyword{Multivariate}

