% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs-package.R
\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{Momocs}
\description{
A complete toolkit for 2D morphometrics,
from data extraction to multivariate
analyses. Most common morphometrics
approaches are included: outlines, curves, landmarks,
"traditional" morphometrics, and facilities for
data preparation, manipulation and
visualization with a consistent grammar
throughout. Momocs allows reproducible and
complex morphometric analyses, other approaches to
be plugged in or developped from those existing, and
paves the way for a pure open-source workflow in R.
}
\details{
To cite Momocs in publications: \code{citation("Momocs")}.
}
\section{Cheers}{

We are very grateful to (in alphabetical order): Sean Asselin, Laurent Bouby, Matt Bulbert, Simon Crameri, Julia Cooke, April Dinwiddie,
Carl Lipo, Cedric Gaucherel, Catherine Girard, Christian Steven Hoggard,
Sarah Ivorra, Glynis Jones, Nathalie Keller, Ricardo Kriebel, Remi Laffont, Fabien Lafuma,
Matthias Mace, Stas Malavin, Neus Martinez, Sabrina Renaud, Marcelo Reginato, Evan Saitta, David Siddons, Eleanor Stillman,
Theodore Stammer, Norbert Telmon, Jean-Frederic Terral, Bill Venables, Daniele Ventura, Michael Wallace,
Asher Wishkerman, John Wood for their helpful ideas and bug reports.
}

\references{
\itemize{
\item Bonhomme V, Picq S, Gaucherel C, Claude J. 2014. Momocs: Outline Analysis Using R.
\emph{Journal of Statistical Software} \bold{56}. \url{http://www.jstatsoft.org/v56/i13}.
\item Claude J. 2008. \emph{Morphometrics with R}. Springer-Verlag, New-York.
}
}
\seealso{
\itemize{
\item \bold{Homepage}: \url{https://github.com/vbonhomme/Momocs}
\item \bold{Issues}: \url{https://github.com/vbonhomme/Momocs/issues}
\item \bold{Tutorial}: \code{browseVignettes("Momocs")}
\item \bold{Email}: \code{bonhomme.vincent@gmail.com} to contribute to dev, ask for something, share your data, etc.
}
}
