% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-misc.R
\name{Momocs_help}
\alias{Momocs_help}
\title{Browse Momocs online doc}
\usage{
Momocs_help(topic = NULL)
}
\arguments{
\item{topic}{the function name to access. If not specified the homepage of
the online manual is accessed.}
}
\description{
Launch a browser to an online version of the manual
}

