% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-misc.R
\name{Momocs_installedversion}
\alias{Momocs_installedversion}
\alias{Momocs_installlastversion}
\alias{Momocs_currentGitHubversion}
\alias{Momocs_currentCRANversion}
\alias{Momocs_help}
\alias{Momocs_website}
\title{Momocs helpers}
\usage{
Momocs_installedversion()

Momocs_installlastversion()

Momocs_currentGitHubversion()

Momocs_currentCRANversion()

Momocs_help(topic = NULL)

Momocs_website()
}
\arguments{
\item{topic}{\code{character} quoted function name for which you want to browse online help}
}
\description{
Help access Momocs website (\code{Momocs_website})
and online help (\code{Momocs_help}),
last released function (\code{Momocs_currentCRANversion}),
last development version on GitHub (\code{Momocs_currentGitHubversion})\code{, install the latter (}Momocs_installlastversion\code{) , know the currently installed version (}Momocs_installedversion`).
}
\examples{
\dontrun{
Momocs_website()
Momocs_help("efourier")
Momocs_currentCRANversion()
Momocs_currentGitHubversion()
Momocs_installlastversion()
Momocs_installedversion
}
}
