% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-misc.R
\name{Momocs_lastversion}
\alias{Momocs_currentCRANversion}
\alias{Momocs_currentGitHubversion}
\alias{Momocs_lastversion}
\alias{Momocs_version}
\title{Install and load the last version of Momocs}
\usage{
Momocs_lastversion()

Momocs_currentGitHubversion()

Momocs_currentCRANversion()

Momocs_version()
}
\description{
Download the last version of Momocs from its GitHub account
from \code{http://www.github.com/vbonhomme/Momocs}), install it and load it (\code{library(Momocs)}).
You need devtools, but it is checked anyway.

Simple wrappers to check installed and available Momocs versions.
}
\examples{
\dontrun{
Momocs_currentGitHubversion()
Momocs_currentCRANversion()
}
}

