% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mshape}
\alias{mshape}
\title{Mean shape calculation}
\usage{
mshape(x)
}
\arguments{
\item{x}{an array (of dim=3), a list of coordinates or an Ldk object}
}
\value{
a matrix of (x; y) coordinates
}
\description{
Returns the mean shape, usually for procrustes/bookstein/baseline - aligned
Ldk objects, matrices (just return the matrix) and dim-3 arrays.
}
\examples{
data(wings)
mshape(wings)
mshape(wings$coo)
data(bot)
mshape(coo.sample(bot, 24)$coo)
mshape(wings[1])
}

