% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mshapes}
\alias{mshapes}
\alias{mshapes.OpnCoe}
\alias{mshapes.OutCoe}
\alias{mshapes.default}
\title{Mean shape calculation from matrices of coefficients}
\usage{
mshapes(Coe, fac, nb.pts)

\method{mshapes}{default}(Coe, fac, nb.pts)

\method{mshapes}{OutCoe}(Coe, fac, nb.pts = 120)

\method{mshapes}{OpnCoe}(Coe, fac, nb.pts = 120)
}
\arguments{
\item{Coe}{a \link{Coe} object}

\item{fac}{factor from the $fac slot. See examples below.}

\item{nb.pts}{numeric the number of points for calculated shapes}
}
\value{
a list of matrices of (x,y) coordinates.
}
\description{
Calculates mean shapes on matrices of coefficients by groups (if passed with
a "fac") or globally (if not).
}
\keyword{Multivariate}

