% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mutate}
\alias{mutate}
\alias{mutate.Coe}
\alias{mutate.default}
\title{Mutate Coe (and others) objects}
\usage{
mutate(x, ...)

\method{mutate}{default}(x, margin = 2, size = dim(x)[ifelse(margin == 1, 2,
  1)], ...)

\method{mutate}{Coe}(x, size = nrow(x$coe), ...)
}
\arguments{
\item{x}{the object to permute}

\item{...}{useless here}

\item{margin}{numeric whether 1 or 2 (rows or columns)}

\item{size}{numeric the required size for the final object}
}
\description{
This methods applies column-wise on the \code{coe} of any
\link{Coe} object but relies on a function that can be used on any matrix. It
simply uses \link{rnorm} with the mean and sd calculated for every column (or row).
}
\examples{
m <- matrix(1:12, nrow=3)
mutate(m, margin=2, size=4)
mutate(m, margin=1, size=10)

data(bot)
bot.f <- eFourier(bot, 12)
bot.m <- mutate(bot.f, 80)
bot.m
panel(bot.m)
}
\seealso{
\link{perm}
}
\keyword{Coe}

