% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nts2Coo}
\alias{nts2Coo}
\title{From .nts to Coo objects}
\usage{
nts2Coo(nts.path, sep = "\\t")
}
\arguments{
\item{nts.path}{the path to the .nts file}

\item{sep}{the separator between data}
}
\description{
Useful to convert .nts files into Coo objects.
It returns a list of matrices of coordinates that can be passed to \link{Coo}.
}
\details{
I'm not very familiar to other morphometric formats.
So if you have troubles importing your datasets, please contact me.
}
\keyword{babel}

