% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel2}
\alias{panel2}
\title{A family picture of a Coo object: all shapes side by side}
\usage{
panel2(Coo)
}
\arguments{
\item{Coo}{a Coo object}
}
\value{
a ggplot2 object
}
\description{
Will replace panel soon.
}
\examples{
data(shapes)
panel2(shapes)
}

