\name{pca2shp.efourier}
\alias{pca2shp.efourier}
\title{Returns shape for a point on a PC plan}
\usage{
pca2shp.efourier(pos, rot, mshape, amp = 1, nb.pts = 60, trans = TRUE)
}
\arguments{
  \item{pos}{the position on the plan}

  \item{rot}{the loadings}

  \item{mshape}{the mean shape}

  \item{amp}{the amplification factor}

  \item{nb.pts}{the number of point for drawing the shapes}

  \item{trans}{logical}
}
\description{
todo
}
\keyword{graphics}

