\name{pca2shp.polynomials}
\alias{pca2shp.polynomials}
\title{Calculates shapes from PC plane: polynomials}
\usage{
pca2shp.polynomials(pos, rot, mshape, amp.shp = 1, pts.shp = 60, mod)
}
\arguments{
  \item{pos}{the position on two PC axis}

  \item{rot}{the corresponding loadings}

  \item{mshape}{the meanshape}

  \item{amp.shp}{amplification factor for the shape
  deformation}

  \item{pts.shp}{number of points to reconstruct the shape}

  \item{mod}{a lm model}
}
\description{
Calculates shapes from PC plane: polynomials
}

