% Generated by roxygen2 (4.0.0): do not edit by hand
\name{PCA2shp.procrustes}
\alias{PCA2shp.procrustes}
\title{Calculates shapes from PC plane: (aligned) landmarks}
\usage{
PCA2shp.procrustes(pos, rot, mshape, amp.shp = 1)
}
\arguments{
\item{pos}{the position on two PC axis}

\item{rot}{the corresponding loadings}

\item{mshape}{the meanshape}

\item{amp.shp}{amplification factor for the shape deformation}
}
\description{
Calculates shapes from PC plane: (aligned) landmarks
}
\keyword{Graphics}

