% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pca2shp.rfourier}
\alias{pca2shp.rfourier}
\title{Calculates shapes from PC plane: rfourier}
\usage{
pca2shp.rfourier(pos, rot, mshape, amp.shp = 1, pts.shp = 60)
}
\arguments{
\item{pos}{the position on two PC axis}

\item{rot}{the corresponding loadings}

\item{mshape}{the meanshape}

\item{amp.shp}{amplification factor for the shape deformation}

\item{pts.shp}{number of points to reconstruct the shape}
}
\description{
Calculates shapes from PC plane: rfourier
}

