% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grindr-pile.R
\name{pile}
\alias{pile}
\alias{pile}
\alias{pile.default}
\alias{pile}
\alias{pile.Out}
\alias{pile}
\alias{pile.Opn}
\alias{pile}
\alias{pile.Ldk}
\title{Graphical pile of shapes}
\usage{
pile(coo, ...)

\method{pile}{default}(coo, ...)

\method{pile}{Out}(coo, ...)

\method{pile}{Opn}(coo, ...)

\method{pile}{Ldk}(coo, ...)
}
\arguments{
\item{coo}{a single shape or any  \link{Coo} object}

\item{...}{more arguments to feed the core drawer, depending on the object}
}
\description{
Pile all shapes in the same graphical window. Useful to check
their normalization in terms of size, position, rotation, first point, etc.
It is, essentially, a shortcut around \code{paper + drawers} of the grindr family.
}
\note{
A variation of this plot was called \code{stack} before \code{Momocs 1.2.5}
}
\examples{
# all Coo are supported wtih sensible defaults
pile(bot)    # outlines
pile(olea)   # curves
pile(wings)  # landmarks

# you can continue the pipe with compatible drawers
pile(bot, trans=0.9) \%>\% draw_centroid

# if you are not happy with this, build your own !
# eg see Momocs::pile.Out

my_pile <- function(x, transp=0.5){
    x \%>\% paper_chess(n=100) \%>\%
          draw_outlines(transp=transp) \%>\%
          draw_centroid(col="red", pch=20, trans=transp) \%>\%
          draw_firstpoint(transp=transp)
}

bot \%>\% my_pile(transp=1/4)

}
\seealso{
Other grindr: \code{\link{drawers}}, \code{\link{layers}},
  \code{\link{papers}}, \code{\link{plot_PCA}}
}
