% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.Coo}
\alias{plot.Coo}
\title{Plots Coo objects: quick review}
\usage{
\method{plot}{Coo}(x, id, ...)
}
\arguments{
\item{x}{the \link{Coo} object}

\item{id}{the id of the shape to plot, if not provided a
random shape is plotted. If passed with \code{"all"} all shapes are plotted,
one by one.}

\item{...}{further arguments to be passed to \link{coo.plot}}
}
\description{
Allows to plot shapes, individually, for \link{Coo} (\link{Out}, \link{Opn} or \link{Ldk}) objects.
}
\examples{
\dontrun{
data(bot)
plot(bot, 5)
plot(bot)
plot(bot, 5, pch=3, points=TRUE) # an example of '...' use
}
}
\seealso{
\link{panel.Coo}, \link{stack.Coo}.
}
\keyword{Graphics}

