% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-LDA.R
\name{plot.LDA}
\alias{plot.LDA}
\title{Plots Linear Discriminant Analysis}
\usage{
\method{plot}{LDA}(x, fac = x$fac, xax = 1, yax = 2, points = TRUE,
  col = "#000000", pch = 20, cex = 0.5, palette = col_solarized,
  center.origin = FALSE, zoom = 1, bg = par("bg"), grid = TRUE,
  nb.grids = 3, morphospace = FALSE, pos.shp = c("range", "full",
  "circle", "xy", "range_axes", "full_axes")[1], amp.shp = 1, size.shp = 1,
  nb.shp = 12, nr.shp = 6, nc.shp = 5, rotate.shp = 0,
  flipx.shp = FALSE, flipy.shp = FALSE, pts.shp = 60,
  border.shp = col_alpha("#000000", 0.5), lwd.shp = 1,
  col.shp = col_alpha("#000000", 0.95), stars = FALSE, ellipses = FALSE,
  conf.ellipses = 0.5, ellipsesax = TRUE, conf.ellipsesax = c(0.5, 0.9),
  lty.ellipsesax = 1, lwd.ellipsesax = sqrt(2), chull = FALSE,
  chull.lty = 1, chull.filled = FALSE, chull.filled.alpha = 0.92,
  density = FALSE, lev.density = 20, contour = FALSE, lev.contour = 3,
  n.kde2d = 100, delaunay = FALSE, loadings = FALSE,
  labelspoints = FALSE, col.labelspoints = par("fg"),
  cex.labelspoints = 0.6, abbreviate.labelspoints = TRUE,
  labelsgroups = TRUE, cex.labelsgroups = 0.8, rect.labelsgroups = FALSE,
  abbreviate.labelsgroups = FALSE, color.legend = FALSE, axisnames = TRUE,
  axisvar = TRUE, unit = FALSE, eigen = TRUE, rug = TRUE,
  title = substitute(x), box = TRUE, old.par = TRUE, ...)
}
\arguments{
\item{x}{an object of class "LDA", typically obtained with \link{LDA}}

\item{fac}{name or the column id from the $fac slot, or a formula combining colum names
from the $fac slot (cf. examples). A factor or a numeric of the same length
can also be passed on the fly.}

\item{xax}{the first PC axis}

\item{yax}{the second PC axis}

\item{points}{logical whether to plot points}

\item{col}{a color for the points (either global, for every level of the fac
or for every individual, see examples)}

\item{pch}{a pch for the points (either global, for every level of the fac
or for every individual, see examples)}

\item{cex}{the size of the points}

\item{palette}{a \link{palette}}

\item{center.origin}{logical whether to center the plot onto the origin}

\item{zoom}{to keep your distances}

\item{bg}{color for the background}

\item{grid}{logical whether to draw a grid}

\item{nb.grids}{and how many of them}

\item{morphospace}{logical whether to add the morphological space}

\item{pos.shp}{passed to \link{pos.shapes}, one of
\code{"range", "full", "circle", "xy", "range_axes", "full_axes"}. Or directly
a matrix of positions. See \link{pos.shapes}}

\item{amp.shp}{amplification factor for shape deformation}

\item{size.shp}{the size of the shapes}

\item{nb.shp}{(pos.shp="circle") the number of shapes on the compass}

\item{nr.shp}{(pos.shp="full" or "range) the number of shapes per row}

\item{nc.shp}{(pos.shp="full" or "range) the number of shapes per column}

\item{rotate.shp}{angle in radians to rotate shapes (if several methods, a vector of angles)}

\item{flipx.shp}{same as above, whether to apply coo_flipx}

\item{flipy.shp}{same as above, whether to apply coo_flipy}

\item{pts.shp}{the number of points fro drawing shapes}

\item{border.shp}{the border color of the shapes}

\item{lwd.shp}{the line width for these shapes}

\item{col.shp}{the color of the shapes}

\item{stars}{logical whether to draw "stars"}

\item{ellipses}{logical whether to draw confidence ellipses}

\item{conf.ellipses}{numeric the quantile for the (bivariate gaussian) confidence ellipses}

\item{ellipsesax}{logical whether to draw ellipse axes}

\item{conf.ellipsesax}{one or more numeric, the quantiles for the (bivariate gaussian) ellipses axes}

\item{lty.ellipsesax}{if yes, the lty with which to draw these axes}

\item{lwd.ellipsesax}{if yes, one or more numeric for the line widths}

\item{chull}{logical whether to draw a convex hull}

\item{chull.lty}{if yes, its linetype}

\item{chull.filled}{logical whether to add filled convex hulls}

\item{chull.filled.alpha}{numeric alpha transparency}

\item{density}{whether to add a 2d density kernel estimation (based on \link{kde2d})}

\item{lev.density}{if yes, the number of levels to plot (through \link{image})}

\item{contour}{whether to add contour lines based on 2d density kernel}

\item{lev.contour}{if yes, the (approximate) number of lines to draw}

\item{n.kde2d}{the number of bins for \link{kde2d}, ie the 'smoothness' of density kernel}

\item{delaunay}{logical whether to add a delaunay 'mesh' between points}

\item{loadings}{logical whether to add loadings for every variables}

\item{labelspoints}{if TRUE rownames are used as labels, a colname from $fac can also be passed}

\item{col.labelspoints}{a color for these labels, otherwise inherited from fac}

\item{cex.labelspoints}{a cex for these labels}

\item{abbreviate.labelspoints}{logical whether to abbreviate}

\item{labelsgroups}{logical whether to add labels for groups}

\item{cex.labelsgroups}{ifyes, a numeric for the size of the labels}

\item{rect.labelsgroups}{logical whether to add a rectangle behind groups names}

\item{abbreviate.labelsgroups}{logical, whether to abbreviate group names}

\item{color.legend}{logical whether to add a (cheap) color legend for numeric fac}

\item{axisnames}{logical whether to add PC names}

\item{axisvar}{logical whether to draw the variance they explain}

\item{unit}{logical whether to add plane unit}

\item{eigen}{logical whether to draw a plot of the eigen values}

\item{rug}{logical whether to add rug to margins}

\item{title}{character a name for the plot}

\item{box}{whether to draw a box around the plotting region}

\item{old.par}{whether to restore the old \link{par}. Set it to \code{FALSE} if you want to reuse the graphical window.}

\item{...}{useless here, just to fit the generic plot}
}
\description{
The Momocs' \code{\link{LDA}} plotter with many graphical options.
}
\details{
Widely inspired by the "layers" philosophy behind graphical functions
of the ade4 R package.
}
\note{
Morphospaces are deprecated so far. 99% of the code
is shared with \link{plot.PCA} waiting for a general rewriting of a multivariate plotter.
See https://github.com/vbonhomme/Momocs/issues/121
}
\examples{
data(bot)
bot.f <- efourier(bot, 24)
bot.l <- LDA(PCA(bot.f), "type")
plot(bot.l)

bot.f$fac$fake <- factor(rep(letters[1:4], each=10))
bot.l <- LDA(PCA(bot.f), "fake")
plot(bot.l)
}
\seealso{
\link{LDA}, \link{plot_CV}, \link{plot_CV2}, \link{plot.PCA}.
}

