\name{plot.Out}
\alias{plot.Out}
\title{Plot on Out objects: quick review}
\usage{
\method{plot}{Out}(x, id, ...)
}
\arguments{
  \item{x}{the Out object ('x' and not Out since this
  methods extends plot)}

  \item{id}{the id of the shape to plot, if not provided a
  random shape is plotted}

  \item{...}{further arguments to be passed to
  \link{coo.plot}}
}
\description{
Allows to plot shapes from Out objects todo
}
\keyword{Out}

