\name{plot}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY,ANY-method}
\alias{plot,Coo,ANY-method}
\title{Plots coordinate outlines.}
\description{
Plots one or a range of the outline(s) contained in \code{Coo}-object.
}
\examples{
data(bottles.cont)
\dontrun{
plot(bottles.cont)
plot(bottles.cont, range=21:40)
}
}
\keyword{Coo methods}

