% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot3}
\alias{plot3}
\alias{plot3.PCA}
\title{Plots a combination of the three first PCs}
\usage{
plot3(PCA, ...)

\method{plot3}{PCA}(PCA, ...)
}
\arguments{
\item{PCA}{a \link{PCA} object}

\item{...}{additional arguments to fed \link{plot.PCA}}
}
\description{
Creates a 2 x 3 layout with, from top to bottom and form left to right: PC1-PC2,
PC1-PC3, PC2-3, and the barplot of eigenvalues percentages.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
bot.p <- PCA(bot.f)
plot3(bot.p) # no groups
plot3(bot.p, 1) # groups
plot3(bot.p, "type", pos.shp="circle") # all plot.PCA args should work
}
\keyword{Graphics}

