% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult-KMEDOIDS.R
\name{plot_silhouette}
\alias{plot_silhouette}
\title{Silhouette plot}
\usage{
plot_silhouette(x, palette = pal_qual)
}
\arguments{
\item{x}{object returned by \link{KMEDOIDS}}

\item{palette}{one of \link{palettes}}
}
\value{
a \code{ggplot} plot
}
\description{
Only used, so far, after \link{KMEDOIDS}.
}
\examples{

olea \%>\% opoly(5) \%>\%
    KMEDOIDS(4) \%>\%
    plot_silhouette(pal_qual_solarized)
}
