% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polynomials_i}
\alias{polynomials_i}
\title{Calculates shape from a polynomial model}
\usage{
polynomials_i(pol, nb.pts = 120, reregister = TRUE)
}
\arguments{
\item{pol}{a pol list such as created by \link{polynomials}}

\item{nb.pts}{the number of points to predict. By default (and cannot be higher)
the number of points in the original shape.}

\item{reregister}{logical whether to reregister the shape with the original baseline.}
}
\value{
a matrix of (x; y) coordinates.
}
\description{
Returns a matrix of (x; y) coordinates when passed with a list obtained with
\link{polynomials}.
}
\examples{
data(olea)
o <- olea[5]
coo_plot(o)
for (i in 2:7){
x <- polynomials_i(polynomials(o, i, ortho=TRUE))
coo_draw(x, border=col_summer(7)[i], points=FALSE)  }
data(olea)
o <- olea[1]
op <- polynomials(o, degree=4)
op
# shape reconstruction
opi <- polynomials_i(op)
lines(opi, col='red')
# R2 for degree 1 to 10
r <- numeric()
for (i in 1:10) { r[i] <- polynomials(o, degree=i)$r2 }
plot(2:10, r[2:10], type='b', pch=20, col='red', main='R2 / degree')
}
\keyword{Polynomials}

