% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pProcrustes}
\alias{pProcrustes}
\title{Partial Procrustes alignment between two shapes}
\usage{
pProcrustes(coo1, coo2)
}
\arguments{
\item{coo1}{Configuration matrix to be superimposed onto the centered preshape of coo2.}

\item{coo2}{Reference configuration matrix.}
}
\value{
a list with components
\enumerate{
\item coo1: Superimposed centered preshape of coo1 onto the centered preshape of coo2
\item coo2: Centered preshape of coo2
\item rotation: Rotation matrix
\item DP: Partial Procrustes distance between coo1 and coo2.
\item rho: Trigonometric Procrustes distance.
}
}
\description{
Directly borrowed from Claude 2008.
}
\references{
Claude, J. (2008). Morphometrics with R. Analysis (p. 316). Springer.
}

