% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rawPolynomials}
\alias{rawPolynomials}
\title{Calculates raw (natural) polynomials on Opn}
\usage{
rawPolynomials(Opn, degree, baseline1, baseline2, nb.pts)
}
\arguments{
\item{Opn}{an \link{Opn} object}

\item{degree}{of the polynomial}

\item{baseline1}{numeric the \eqn{(x; y)} coordinates of the first baseline
by default \eqn{(x= -1; y=0)}}

\item{baseline2}{numeric the \eqn{(x; y)} coordinates of the second baseline
by default \eqn{(x= 1; y=0)}}

\item{nb.pts}{number of points to sample and on which to calculate polynomials}
}
\value{
a \code{OpnCoe} object.
}
\description{
Calculates raw (natural) polynomials on Opn
}
\examples{
data(olea)
rawPolynomials(olea, 5)
}
\keyword{Opn}

