% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs-package.R
\docType{import}
\name{reexports}
\alias{\%$\%}
\alias{\%<>\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%T>\%}}}
}}
\keyword{internal}

