% Generated by roxygen2 (4.0.0): do not edit by hand
\name{removeAsymmetric}
\alias{removeAsymmetric}
\alias{removeAsymmetric.OutCoe}
\alias{removeAsymmetric.default}
\alias{removeSymmetric}
\alias{removeSymmetric.OutCoe}
\alias{removeSymmetric.default}
\title{Removes asymmetric and symmetric variation on OutCoe objects}
\usage{
removeAsymmetric(OutCoe)

\method{removeAsymmetric}{default}(OutCoe)

\method{removeAsymmetric}{OutCoe}(OutCoe)

removeSymmetric(OutCoe)

\method{removeSymmetric}{default}(OutCoe)

\method{removeSymmetric}{OutCoe}(OutCoe)
}
\arguments{
\item{OutCoe}{an OutCoe object}
}
\value{
an OutCoe object
}
\description{
Only for those obtained with \link{eFourier}, otherwise a message is returned.
\code{removeAsymmetric} sets all B and C coefficients to 0; \code{removeSymmetric} sets
all A and D coefficients to 0.
}
\examples{
data(bot)
botf <- eFourier(bot, 12)
botSym <- removeAsymmetric(botf)
boxplot(botSym)
botSymp <- PCA(botSym)
plot(botSymp)
plot(botSymp, amp.shp=5)

# Asymmetric only
botAsym <- removeSymmetric(botf)
boxplot(botAsym)
botAsymp <- PCA(botAsym)
plot(botAsymp)
# strange shapes because the original shape was mainly symmetric and would need its
# symmetric (eg its average) for a proper reconstruction. Should only be used like that:
plot(botAsymp, morpho=FALSE)
}
\references{
Below: the first mention, and two applications.
\itemize{
#' \item Iwata, H., Niikura, S., Matsuura, S., Takano, Y., & Ukai, Y. (1998).
Evaluation of variation of root shape of Japanese radish (Raphanus sativus L.)
based on image analysis using elliptic Fourier descriptors. Euphytica, 102, 143-149.
\item Iwata, H., Nesumi, H., Ninomiya, S., Takano, Y., & Ukai, Y. (2002).
The Evaluation of Genotype x Environment Interactions of Citrus Leaf Morphology
Using Image Analysis and Elliptic Fourier Descriptors. Breeding Science, 52(2),
89-94. doi:10.1270/jsbbs.52.89
\item Yoshioka, Y., Iwata, H., Ohsawa, R., & Ninomiya, S. (2004).
Analysis of petal shape variation of Primula sieboldii by elliptic fourier descriptors
and principal component analysis. Annals of Botany, 94(5), 657-64. doi:10.1093/aob/mch190
}
}

