% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{rename}
\alias{rename}
\title{Rename columns by name}
\usage{
rename(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{comma separated list of unquoted expressions}
}
\value{
a Momocs object of the same class.
}
\description{
Rename variables, from the \code{$fac}. See examples and \link[dplyr:rename]{dplyr::rename}.
}
\details{
dplyr verbs are maintained.
}
\examples{
olea
rename(olea, variety=var, domesticated=domes) # rename var column

}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{at_least}}, \code{\link{chop}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{fac_dispatcher}}, \code{\link{filter}},
  \code{\link{mutate}}, \code{\link{rescale}},
  \code{\link{rm_harm}}, \code{\link{rm_missing}},
  \code{\link{rm_uncomplete}}, \code{\link{rw_fac}},
  \code{\link{sample_frac}}, \code{\link{sample_n}},
  \code{\link{select}}, \code{\link{slice}},
  \code{\link{subsetize}}
}
