% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-out-rfourier.R
\name{rfourier}
\alias{rfourier}
\alias{rfourier.Out}
\alias{rfourier.default}
\title{Radii variation Fourier transform}
\usage{
rfourier(x, ...)

\method{rfourier}{default}(x, nb.h, smooth.it = 0, norm = FALSE,
  verbose = TRUE, ...)

\method{rfourier}{Out}(x, nb.h = 40, smooth.it = 0, norm = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{list} or \code{matrix} of coordinates or an \code{Out} object}

\item{...}{useless here}

\item{nb.h}{\code{integer}. The number of harmonics to use. If missing 99pc harmonic power is used.}

\item{smooth.it}{\code{integer}. The number of smoothing iterations to
perform.}

\item{norm}{\code{logical}. Whether to scale the outlines so that the mean
length of the radii used equals 1.}

\item{verbose}{\code{logical}. Whether to display diagnosis messages.}
}
\value{
A list with following components:
\itemize{
 \item \code{an} vector of \eqn{a_{1->n}} harmonic coefficients
 \item \code{bn} vector of \eqn{b_{1->n}} harmonic coefficients
 \item \code{ao} ao harmonic coefficient.
 \item \code{r} vector of radii lengths.
 }
}
\description{
\code{rfourier} computes radii variation Fourier analysis from a matrix or a
list of coordinates.
}
\details{
Given a closed outline, the radius \eqn{r}, taken as the distance from the
outline barycentre and a given point of the outline, can be expressed as a
periodic function of the angle \eqn{\theta}. Harmonics from \eqn{0} to
\eqn{k} approximate the function \eqn{r(\theta)}:

\eqn{r(\theta)= \frac{1}{2}a_0 + \sum\limits_{n=1}^{k}a_k\cos(w_k\theta +
b_k\sin(w_k\theta)} with: \eqn{ a_n =
\frac{2}{p}\sum\limits_{n=1}^{p}r_i\cos(n\theta_i) } \eqn{ b_n =
\frac{2}{p}\sum\limits_{n=1}^{p}r_i\sin(n\theta_i) } with \eqn{ a_0 =
\sqrt{\frac{2}{p}}\sum\limits_{n=1}^{p}r_i }

The \eqn{a_n} and \eqn{b_n} harmonic coefficients, extracted for every
individual shape, are then used for multivariate analyses.
}
\note{
Directly borrowed for Claude (2008), and called \code{fourier1} there.
}
\examples{
data(bot)
coo <- coo_center(bot[1]) # centering is almost mandatory for rfourier family
coo_plot(coo)
rf  <- rfourier(coo, 12)
rf
rfi <- rfourier_i(rf)
coo_draw(rfi, border='red', col=NA)
}
\references{
Claude, J. (2008) \emph{Morphometrics with R}, Use R! series,
Springer 316 pp.
}
\seealso{
\link{rfourier} for rfourier on \link{Out} objects.
\link{rfourier_i} for the inverse operation, \link{rfourier_shape} to play around
with this approach.
\link{efourier}, \link{tfourier} for the other members of the Fourier's
family.
}

