% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{rm_harm}
\alias{rm_harm}
\title{Removes harmonics}
\usage{
rm_harm(x, drop = 1)
}
\arguments{
\item{x}{Coe object}

\item{drop}{numeric number of harmonics to drop}
}
\description{
Useful to drop harmonics on Coe objects. Should only work for
Fourier-based approached since it looks for \code{[A-D][1-drop]} pattern.
}
\examples{
data(bot)
bf <- efourier(bot)
colnames(rm_harm(bf, 1)$coe)
}

