% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{rw_rule}
\alias{rw_rule}
\title{Rename levels with a rewriting rule
rw_rule stands for 'rewriting rule'. Typically useful to correct typos
at the import, or merge some levels within covariates. Drops levels silently.}
\usage{
rw_rule(x, fac, from, to)
}
\arguments{
\item{x}{any Momocs object}

\item{fac}{the id of the name of the $fac column to look for}

\item{from}{which level should be renamed}

\item{to}{which name ?}
}
\value{
a Momocs object of the same type
}
\description{
Rename levels with a rewriting rule
rw_rule stands for 'rewriting rule'. Typically useful to correct typos
at the import, or merge some levels within covariates. Drops levels silently.
}
\examples{
data(bot)
rw_rule(bot, "type", "whisky", "agua_de_fuego") # 1 instead of "type" is fine too
}

