% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{sample_frac}
\alias{sample_frac}
\alias{sample_frac.Coe}
\alias{sample_frac.Coo}
\alias{sample_frac.default}
\title{Samples a fraction of shapes in Momocs objects}
\usage{
sample_frac(tbl, size, replace, ...)

\method{sample_frac}{default}(tbl, size = 1, replace = FALSE, ...)

\method{sample_frac}{Coo}(tbl, size = 1, replace = FALSE, fac = NULL, ...)

\method{sample_frac}{Coe}(tbl, size = 1, replace = FALSE, fac = NULL, ...)
}
\arguments{
\item{tbl}{a Momocs object (Coo, Coe)}

\item{size}{numeric (0 < numeric <= 1) the fraction of shapes to select}

\item{replace}{logical whether sample should be done with ot without replacement}

\item{...}{additional arguments to dplyr::sample_frac and to maintain generic compatibility}

\item{fac}{a column name if a $fac is defined; size is then applied within levels of this factor}
}
\description{
Uses (and maintain) dplyr syntax and verb.
}
\note{
the resulting fraction is rounded with \link{ceiling}.
}
\examples{
data(bot)
bot
# samples 50\% of the bottles no matter their type
sample_frac(bot, 0.5)
# 80\% bottles of beer and of whisky
table(sample_frac(bot, 0.8, fac="type"))
# bootstrap the same number of bootles of each type but with replacement
table(names(sample_frac(bot, 1, replace=TRUE)))
}
\seealso{
\link{sample_n}
}

