% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{select}
\alias{select}
\alias{select.Coe}
\alias{select.Coo}
\alias{select.PCA}
\alias{select.default}
\title{Select/rename (ala dplyr) variables by name, using the $fac slot}
\usage{
select(.data, ...)

\method{select}{default}(.data, ...)

\method{select}{Coo}(.data, ...)

\method{select}{Coe}(.data, ...)

\method{select}{PCA}(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo} objects}

\item{...}{comma separated list of unquoted expressions}
}
\value{
a Coo object
}
\description{
Extends this dplyr verb to \link{Coo} objects. See examples.
}
\details{
dplyr verbs should be maintained.
}
\note{
this method is quite experimental, check the result and do not hesitate to report a bug.
}
\examples{
data(olea)
olea
select(olea, cep, view) # drops domes and ind
select(olea, cepage=cep, domesticated_status=domes, view)
# combine with filter with magrittr pipes
\dontrun{
library(magrittr)
# only dorsal views, and 'cep' and 'domes' columns
filter(olea, view=="VD") \%>\% select(cep, domes)
}
}
\seealso{
\link{filter.Coo}
}

