% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core-premodern.R
\name{shape_descriptors}
\alias{shape_descriptors}
\title{Calculates shape descriptors from Coo objects}
\usage{
shape_descriptors(Coo, ...)
}
\arguments{
\item{Coo}{object on which to calculate descriptors}

\item{...}{functions that returns a scalar (a single value), see examples}
}
\description{
Extracts descriptors from a list of functions and returns a TraCoe object
}
\details{
any function, including custom ones, that return a scalar can be passed.
To access the list of 'coo_' functions - not all are suitable, see \code{apropos("coo_")}
}
\examples{
data(bot)
desc <- shape_descriptors(bot, coo_circularity, coo_rectangularity, coo_area)
desc
desc$coe
# a custom function the product of length*width
pseudo_area <- function(x) prod(coo_lw(x))
shape_descriptors(bot, pseudo_area)
}

