\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,Coo-method}
\alias{show,Nef-method}
\title{show methods for Momocs' objects}
\description{
Momocs objects have \code{show} methods. They will be expanded in further versions.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Coo")}}{show a \code{Coo}-object}

\item{\code{signature(object = "Nef")}}{show a \code{Coo}-object}
}}
\examples{
data(bottles.cont)
bottles.cont

data(bottles.nef)
bottles.nef
}
\keyword{methods}
