% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{slice}
\alias{slice}
\title{Slices (ala dplyr) on Momocs objects}
\usage{
slice(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{logical conditions}
}
\value{
a Momocs object of the same class.
}
\description{
Select rows by position, based on \code{$fac}. See examples and \code{?dplyr::slice}.
}
\details{
dplyr verbs are maintained.
}
\examples{
olea
slice(olea, 1) # if you only want the coordinates, try bot[1]
slice(olea, 1:20)
slice(olea, 21:30)
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{at_least}}, \code{\link{chop}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rename}}, \code{\link{rm_uncomplete}},
  \code{\link{rw_fac}}, \code{\link{sample_frac}},
  \code{\link{sample_n}}, \code{\link{select}},
  \code{\link{subset.Coo}}
}
