% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{slidings_scheme}
\alias{slidings_scheme}
\title{Extract partition of sliding coordinates}
\usage{
slidings_scheme(Coo)
}
\arguments{
\item{Coo}{an Ldk object}
}
\value{
a list with two components: \code{n} the number of partition; \code{id}
their position.
}
\description{
Helper function that deduces (likely to be a reminder)
partition scheme from \code{$slidings} of \code{Ldk} objects.
}
\examples{
# waiting for chaff, todo
}

