% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stack.Coo}
\alias{stack.Coo}
\alias{stack.Ldk}
\title{Plots Coo objects: stacks all shapes}
\usage{
\method{stack}{Coo}(x, cols, borders, points = FALSE, first.point = TRUE,
  centroid = TRUE, ldk = TRUE, ldk.pch = 3, ldk.col = "#FF000055",
  ldk.cex = 0.5, ldk.links = FALSE, ldk.confell = FALSE,
  ldk.contour = FALSE, ldk.chull = FALSE, ldk.labels = FALSE,
  xy.axis = TRUE, ...)

\method{stack}{Ldk}(x, cols, borders, first.point = TRUE, centroid = TRUE,
  ldk = TRUE, ldk.pch = 20, ldk.col = "#33333333", ldk.cex = 0.3,
  ldk.links = FALSE, ldk.confell = FALSE, ldk.contour = FALSE,
  ldk.chull = FALSE, ldk.labels = FALSE, xy.axis = TRUE, ...)
}
\arguments{
\item{x}{The \code{Coo} object to plot.}

\item{cols}{A \code{vector} of colors for drawing the outlines.
Either a single value or of length exactly equals to the number of coordinates.}

\item{borders}{A \code{vector} of colors for drawing the borders.
Either a single value or of length exactly equals to the number of coordinates.}

\item{points}{logical whether to draw or not points}

\item{first.point}{logical whether to draw or not the first point}

\item{centroid}{logical whether to draw or not the centroid}

\item{ldk}{\code{logical}. Whether to display landmarks (if any).}

\item{ldk.pch}{A \code{pch} for these landmarks}

\item{ldk.col}{A color for these landmarks}

\item{ldk.cex}{A \code{cex} fro these landmarks}

\item{ldk.links}{logical whether to draw links (of the mean shape)}

\item{ldk.confell}{logical whether to draw conf ellipses}

\item{ldk.contour}{logical whether to draw contour lines}

\item{ldk.chull}{logical whether to draw convex hull}

\item{ldk.labels}{logical whether to draw landmark labels}

\item{xy.axis}{whether to draw or not the x and y axes}

\item{...}{further arguments to be passed to \link{coo.plot}}
}
\description{
Plots all the outlines, on the same graph, from a \link{Coo} (\link{Out}, \link{Opn} or \link{Ldk})
object.
}
\examples{
\dontrun{
data(mosquito)
stack(mosquito, borders='#1A1A1A22', first.point=FALSE)
data(hearts)
stack(hearts)
stack(hearts, ldk=FALSE)
stack(hearts, borders='#1A1A1A22', ldk=TRUE, ldk.col=col.summer(4), ldk.pch=20)
}
}
\seealso{
\link{panel.Coo}, \link{plot.Coo}.
}
\keyword{Graphics}
\keyword{Ldk}
\keyword{Opn}
\keyword{Out}

