% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stack2}
\alias{stack2}
\title{A family picture of a Coo object: all shapes in the same plot}
\usage{
stack2(Coo)
}
\arguments{
\item{Coo}{a Coo object}
}
\value{
a ggplot2 object
}
\description{
Will replace stack soon.
}
\examples{
data(bot)
stack2(bot)
}

