\name{subset.Out}
\alias{subset.Out}
\title{Create subsets of Out objects}
\usage{
\method{subset}{Out}(x, subset, ...)
}
\arguments{
  \item{x}{and Out object}

  \item{subset}{logical from the fac or indices}

  \item{...}{(to preserve consistence with subset generic)}
}
\description{
todo
}
\keyword{Out}

