% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{table}
\alias{table}
\alias{table.default}
\alias{table.Coo}
\alias{table.Coe}
\alias{table.PCA}
\alias{table.LDA}
\title{Cross-tabulates objects}
\usage{
table(...)

\method{table}{default}(...)

\method{table}{Coo}(...)

\method{table}{Coe}(...)

\method{table}{PCA}(...)

\method{table}{LDA}(...)
}
\arguments{
\item{...}{a list of, first, a Momocs object (Coo, Coe, PCA, etc.), then, column names in the $fac slot. If not specified,
returns a table on the entire $fac data.frame}
}
\description{
Simply extends base \link{table} for a more convenient use on $fac slot.
}
\examples{
table(olea, "var", "domes")
table(olea)
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{at_least}}, \code{\link{chop}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rename}}, \code{\link{rescale}},
  \code{\link{rm_harm}}, \code{\link{rm_uncomplete}},
  \code{\link{rw_fac}}, \code{\link{sample_frac}},
  \code{\link{sample_n}}, \code{\link{select}},
  \code{\link{slice}}
}
