% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tFourier}
\alias{tFourier}
\title{Calculates tangent angle Fourier analysis on Out objects}
\usage{
tFourier(Out, nb.h, smooth.it, norm)
}
\arguments{
\item{Out}{the Out object on which to calculate eft}

\item{nb.h}{the number of harmonics to calculate}

\item{smooth.it}{the number of smoothing iterations to perform}

\item{norm}{whether to normalize the matrix of coefficients. See Details.}
}
\description{
A wrapper for \link{tfourier} to be applied on Out objects.
}
\examples{
data(bot)
tFourier(bot, 12)
}
\seealso{
\link{tfourier}
}

