% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{tie_jpg_txt}
\alias{tie_jpg_txt}
\title{Binds .jpg outlines from .txt landmarks taken on them}
\usage{
tie_jpg_txt(lf)
}
\arguments{
\item{lf}{a list of filenames}
}
\description{
Given a list of files (lf) that includes matching filenames with .jpg (black masks)
and .txt (landmark positions on them as .txt), returns an Out with $ldk defined.
Typically be useful if you use ImageJ to define landmarks on your outlines.
}
\note{
Not optimized (images are read twice). Please do not hesitate to contact me
should you have a particular case or need something.
}
\seealso{
Other babel functions: \code{\link{lf_structure}}
}
