% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-def-TraCoe.R
\name{TraCoe}
\alias{TraCoe}
\title{Traditional morphometrics class}
\usage{
TraCoe(coe = matrix(), fac = dplyr::tibble())
}
\arguments{
\item{coe}{a matrix of measurements}

\item{fac}{a data.frame for covariates}
}
\value{
a list of class TraCoe
}
\description{
Defines the builder for traditional measurement class in Momocs.
Is is intended to ease calculations, data handling and multivariate statistics
just ad the other Momocs' classes
}
\examples{
# let's (more or less) rebuild the flower dataset
fl <- TraCoe(iris[, 1:4], dplyr::tibble(sp=iris$Species))
fl \%>\% PCA() \%>\% plot("sp")
}
\seealso{
Other classes: 
\code{\link{Coe}()},
\code{\link{Coo}()},
\code{\link{Ldk}()},
\code{\link{OpnCoe}()},
\code{\link{Opn}()},
\code{\link{OutCoe}()},
\code{\link{Out}()}
}
\concept{classes}
