% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{transmute}
\alias{transmute}
\alias{transmute.Coe}
\alias{transmute.Coo}
\alias{transmute.PCA}
\alias{transmute.default}
\title{transmute (ala dplyr) variables by name, using the $fac slot on Momocs classes}
\usage{
transmute(.data, ...)

\method{transmute}{default}(.data, ...)

\method{transmute}{Coo}(.data, ...)

\method{transmute}{Coe}(.data, ...)

\method{transmute}{PCA}(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{comma separated list of unquoted expressions}
}
\value{
a Coo object
}
\description{
Extends this dplyr verb to \link{Coo} objects. See examples.
}
\details{
dplyr verbs should be maintained.
}
\note{
this method is quite experimental, check the result and do not hesitate to report a bug.
}
\examples{
data(olea)
olea
select(olea, var, view) # drops domes and ind
select(olea, variety=var, domesticated_status=domes, view)
# combine with filter with magrittr pipes
\dontrun{
library(magrittr)
# only dorsal views, and 'var' and 'domes' columns
filter(olea, view=="VD") \%>\% select(var, domes)
}
}
\seealso{
\link{select}, \link{filter}, \link{slice}, \link{chop}, \link{combine}.
}

