% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core-premodern.R
\name{truss}
\alias{truss}
\title{Truss measurement}
\usage{
truss(x)
}
\arguments{
\item{x}{a shape or an Ldk object}
}
\value{
a named numeric or matrix
}
\description{
A method to calculate on shapes or on Coo truss measurements,
which is all pairwise combinations of euclidean distances
}
\note{
Mainly implemented for historical/didactical reasons.
}
\examples{
# example on a single shape
data(shapes)
cat <- coo_sample(shapes[4], 6)
truss(cat)

# example on wings dataset
data(wings)
tx <- truss(wings)
dim(tx)
# we normalize and plot an heatmap
txn <- apply(tx, 2, .normalize)
# heatmap(txn)

txp <- PCA(tx, scale. = TRUE, center=TRUE, fac=wings$fac)
plot(txp, 1)
}

