% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-package.R
\name{update_Momocs}
\alias{update_Momocs}
\title{Install and load the last version of Momocs}
\usage{
update_Momocs()
}
\description{
Download the last version of Momocs from its GitHub account
from \code{http://www.github.com/vbonhomme/Momocs}), install it and load it (\code{library(Momocs)})
}
\details{
You need devtools, but it is checked anyway.
}
\examples{
\dontrun{
# use it with:
update_Momocs()
}
}

