% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-def-CooCoe.R
\name{validate}
\alias{validate}
\title{Validates Coo objects
No validation for S3 objects, so this method is a (very tolerant) attempt at checking
\link{Coo} objects, \link{Out}, \link{Opn} and \link{Ldk} objects.}
\usage{
validate(Coo)
}
\arguments{
\item{Coo}{any Coo object}
}
\value{
a Coo object.
}
\description{
This methods checks \enumerate{
   \item if coordinates do all pass \link{coo_check}
   \item that \code{$fac}, if any, is a {data.frame}, with the same number of rows and coordinates.
   Copies the name of \code{$coo} to th rownames
   \item that \code{$ldk}, if any, has a the same number of landmarks, for all shapes
} and finally returns the \code{Coo}. This method is used in \code{Coo} constructor, and before applying
any method on thos object.
}
\examples{
validate(bot)
bot[12] <- NA
validate(bot)

validate(hearts)
hearts$ldk[[4]] <- c(1, 2)
validate(hearts)
}

