% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a2m}
\alias{a2m}
\title{Converts an array of coordinates to a matrix}
\usage{
a2m(a)
}
\arguments{
\item{a}{\code{array} of (x; y) coordinates.}
}
\value{
matrix (see above).
}
\description{
All the individuals (the 3rd dimension of the array) becomes rows, and
columns are (all the) x coordinates and (all the) y coordinates, so that we have
x1, x2, ..., xn, y1, y2, ..., yn columns. Rows and colums are named anyway.
}
\details{
Used in landmarks methods, e.g. for multivariate analysis after a Procrustes alignment.
}
\examples{
data(wings)
a <- l2a(wings$coo)
a
a2m(a)
}
\seealso{
\link{m2a} the reverse function.
}
\keyword{Babel}

