\name{Momocs Package}
\alias{Momocs-package}
\alias{Momocs}
\docType{package}
\title{
Outline Analysis using R.
}
\description{
Momocs is intended to ease and popularize shape analysis of outlines (especially using elliptical Fourier analysis). It mostly hinges on the core functions developped in \emph{Morphometrics with R} (Claude, 2008). From outline extraction of images and elliptical analysis, radii variation and tangent angle Fourier calculation to multivariate analysis and calculation of the morphological space, Momocs provides a complete and convenient toolkit to specialists within every field that are, or may be, interested in morphological comparisons of outlines.

Have a look to Momocs' website: \url{http://www.vincentbonhomme.fr/Momocs}. You will find a slideshow plus some vignettes and other useful links.
}

\author{
\enumerate{
\item Vincent Bonhomme, French Institute of Pondicherry, India. <bonhomme.vincent@gmail.com>
\item Sandrine Picq, UMR CBAE, Montpellier, France.
\item Julien Claude UMR, ISEM, Universite of Montpellier II, France.
}}

\references{
Please cite the paper below in your great papers. It's freely available in JSS website.

Vincent Bonhomme, Sandrine Picq, Cedric Gaucherel, Julien Claude (2014). Momocs: Outline Analysis Using R. Journal of Statistical Software, 56(13), 1-24. \url{ http://www.jstatsoft.org/v56/i13/.}.
  
Have a look to Julien Claude's book that greatly inspired Momocs: Claude, J. (2008) \emph{Morphometrics with R}, Use R! series, Springer 316 pp.
}

\keyword{Package}