% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as_df}
\alias{as_df}
\title{Convert Momocs objects to data.frames}
\usage{
as_df(x)
}
\arguments{
\item{x}{an object, typically a Momocs class}
}
\value{
a data.frame
}
\description{
Used in particular for ggplot2 compatibility
}
\examples{
data(bot)
head(as_df(bot))
bot.f <- efourier(bot, 10)
head(as_df(bot.f))
bot.p <- PCA(bot.f)
head(as_df(bot.p))
bot.l <- LDA(bot.p, "type")
head(as_df(bot.l))
}

