% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{at_least}
\alias{at_least}
\title{Retains group with at least a certain number of individuals within}
\usage{
at_least(x, fac, N)
}
\arguments{
\item{x}{any Momocs object}

\item{N}{minimal number of individuals to retain the group}

\item{y}{the id of name of the $fac column}
}
\description{
Title and examples are self-speaking.
}
\note{
if N is too ambitious the original object is returned with a message
}
\examples{
data(trilo)
table(trilo, "onto")
at_least(trilo, "onto", 9)
at_least(trilo, "onto", 16)
at_least(trilo, "onto", 2000) # too ambitious !
}

