% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{at_least}
\alias{at_least}
\title{Retain groups with at least n shapes}
\usage{
at_least(x, fac, N)
}
\arguments{
\item{x}{any Momocs object}

\item{fac}{the id of name of the $fac column}

\item{N}{minimal number of individuals to retain the group}
}
\description{
Examples are self-speaking.
}
\note{
if N is too ambitious the original object is returned with a message
}
\examples{
table(trilo$onto)
at_least(trilo, "onto", 9)
at_least(trilo, "onto", 16)
at_least(trilo, "onto", 2000) # too ambitious !
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{chop}}, \code{\link{combine}},
  \code{\link{dissolve}}, \code{\link{fac_dispatcher}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rescale}}, \code{\link{rm_harm}},
  \code{\link{rm_uncomplete}}, \code{\link{rw_fac}},
  \code{\link{sample_frac}}, \code{\link{sample_n}},
  \code{\link{select}}, \code{\link{slice}}
}
