% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bezier}
\alias{bezier}
\title{Calculates Bezier coefficients from a shape}
\usage{
bezier(coo, n)
}
\arguments{
\item{coo}{a matrix or a list of (x; y) coordinates}

\item{n}{the degree, by default the number of coordinates.}
}
\value{
a list with components:
\itemize{
\item \code{$J} matrix of Bezier coefficients
\item \code{$B} matrix of Bezier vertices.
}
}
\description{
Directly borrowed for Claude (2008), and also called \code{bezier} there.
}
\note{
Not implemented for open outlines but may be useful for other purposes.
}
\examples{
set.seed(34)
x <- coo.sample(efourier.shape(), 5)
plot(x, ylim=c(-3, 3), asp=1, type="b", pch=20)
b <- bezier(x)
bi <- bezier.i(b$B)
lines(bi, col="red")
}
\keyword{Bezier}

