% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplot.PCA}
\alias{boxplot.PCA}
\title{Boxplot on PCA objects}
\usage{
\method{boxplot}{PCA}(x, fac, nax = 1:4, cols, palette = col_qual,
  fixed.axes = TRUE, center.origin = TRUE, cex.legend = 1, ...)
}
\arguments{
\item{x}{an object of class "PCA", typically obtained with \link{PCA}}

\item{fac}{factor, or a name or the column id from the $fac slot}

\item{nax}{the range of PC axis}

\item{cols}{a vector of colors is palette is not provided}

\item{palette}{a color palette}

\item{fixed.axes}{logical whether axes shoudl have the same scaled}

\item{center.origin}{if TRUE before, whether to center the origin}

\item{cex.legend}{a numeric to specify cex for the legend (if any)}

\item{...}{further arguments to feed \link{boxplot}}
}
\description{
Boxplot on PCA objects
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
bot.p <- PCA(bot.f)
boxplot(bot.p, 1)
}

