% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-PCA.R
\name{boxplot.PCA}
\alias{boxplot.PCA}
\title{Boxplot on PCA objects}
\usage{
\method{boxplot}{PCA}(x, fac = NULL, nax = 1:3, ...)
}
\arguments{
\item{x}{an object of class "PCA", typically obtained with \link{PCA}}

\item{fac}{factor, or a name or the column id from the $fac slot}

\item{nax}{the range of PC to plot}

\item{...}{useless here}
}
\value{
a ggplot object
}
\description{
Boxplot on PCA objects
}
\examples{
bot.f <- efourier(bot, 12)
bot.p <- PCA(bot.f)
boxplot(bot.p)
p <- boxplot(bot.p, 1)
#p +  theme_minimal() + scale_fill_grey()
#p + facet_wrap(~PC, scales = "free")
}
