% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-outopn-calibrate.R
\name{calibrate_r2}
\alias{calibrate_r2}
\alias{calibrate_r2.Opn}
\title{Quantitative calibration, through r2, for Opn objects}
\usage{
calibrate_r2(x, ...)

\method{calibrate_r2}{Opn}(x, method = "opoly", id = 1:length(Opn),
  degree.range = 1:8, thresh = c(0.9, 0.95, 0.99, 0.999), plot = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{and Opn object}

\item{...}{useless here}

\item{method}{one of 'npoly' or 'opoly'}

\item{id}{the ids of shapes on which to calculate r2 (all by default)}

\item{degree.range}{on which to calculate r2}

\item{thresh}{the threshold to return diagnostic}

\item{plot}{logical whether to print the plot}

\item{verbose}{logical whether to print messages}
}
\description{
Estimates the r2 to calibrate the degree for \code{npoly} and \code{opoly} methods.
Also returns a plot
}
\details{
May be long, so you can estimate it on a sample.
}
\section{Methods (by class)}{
\itemize{
\item \code{Opn}: Method for Opn objects
}}
\examples{
\dontrun{
data(olea)
calibrate_r2(olea, "opoly", degree.range=1:12, thresh=c(0.9, 0.99)

}
}
\seealso{
\link{calibrate_harmonicpower}
}

