% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calibrate_reconstructions}
\alias{calibrate_reconstructions}
\alias{calibrate_reconstructions.Opn}
\alias{calibrate_reconstructions.Out}
\title{Calibrate using reconstructed shapes}
\usage{
calibrate_reconstructions(x, ...)

\method{calibrate_reconstructions}{Out}(x, method = c("efourier", "rfourier",
  "tfourier"), id, range = 1:9, ...)

\method{calibrate_reconstructions}{Opn}(x, method = c("npoly", "opoly"), id,
  range = 2:10, baseline1 = c(-1, 0), baseline2 = c(1, 0), ...)
}
\arguments{
\item{x}{the \code{Coo} object on which to calibrate_reconstructions}

\item{...}{only used for the generic}

\item{method}{any method from \code{c('efourier', 'rfourier', 'tfourier')}
for \code{Out}, or from \code{c('opoly', 'npoly')} for \code{Opn}}

\item{id}{the shape on which to perform calibrate_reconstructions}

\item{range}{vector of harmonics on which to perform calibrate_reconstructions}

\item{baseline1}{\eqn{(x; y)} coordinates for the first point of the baseline}

\item{baseline2}{\eqn{(x; y)} coordinates for the second point of the baseline}
}
\description{
Calculate and displays reconstructed shapes using a
range of harmonic number.
}
\examples{
data(bot)
calibrate_reconstructions(bot)
}
\keyword{Out}

