% Generated by roxygen2 (4.0.0): do not edit by hand
\name{chc2pix}
\alias{chc2pix}
\title{Convert chain-coded coordinates to (x; y) coordinates}
\usage{
chc2pix(chc)
}
\arguments{
\item{chc}{chain-coded coordinates}
}
\description{
May be useful to convert (prehistoric?) chain-coded coordinates
to (x; y) coordinates. The first point is set at the origin.
}
\examples{
data(bot)
x <- pix2chc(bot[1])
coo.plot(chc2pix(x))
}
\references{
Kuhl, F. P., & Giardina, C. R. (1982).
Elliptic Fourier features of a closed contour.
Computer Graphics and Image Processing, 18(3), 236-258.
}
\seealso{
\link{pix2chc}
}
\keyword{babel}

