% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-cheapbabi-paper_and_draw.R
\name{cheapbabi_drawers}
\alias{cheapbabi_drawers}
\alias{draw_polygon}
\alias{draw_outline}
\alias{draw_points}
\alias{draw_landmarks}
\alias{draw_lines}
\alias{draw_centroid}
\alias{draw_curve}
\alias{draw_firstpoint}
\alias{draw_axes}
\alias{draw_labels}
\alias{draw_links}
\title{Cheapbabi drawers for shape plots}
\usage{
draw_polygon(coo, f, border = par("fg"), col = NA, lwd = 1, lty = 1,
  ...)

draw_outline(coo, f, border = par("fg"), col = NA, lwd = 1, lty = 1,
  ...)

draw_points(coo, f, col = par("fg"), cex = 1, pch = 20, ...)

draw_landmarks(coo, f, col = par("fg"), cex = 1, pch = 20, ...)

draw_lines(coo, f, col = par("fg"), lwd = 1, lty = 1, ...)

draw_centroid(coo, f, col = par("fg"), pch = 3, cex = 0.5, ...)

draw_curve(coo, f, col = par("fg"), lwd = 1, lty = 1, ...)

draw_firstpoint(coo, f, label = "^", col = par("fg"), cex = 1, ...)

draw_axes(coo, col = "#333333", cex = 3/4, lwd = 3/4, ...)

draw_labels(coo, labels = 1:nrow(coo), cex = 1/2, d = 1/20, ...)

draw_links(coo, f, links, col = "#99999955", lwd = 1/2, lty = 1, ...)
}
\arguments{
\item{coo}{\code{matrix} of 2 columns for (x, y) coordinates}

\item{border}{color (hexadecimal) to draw components}

\item{col}{color (hexadecimal) to draw components}

\item{lwd}{to draw components}

\item{...}{additional options to feed core functions for each drawer}

\item{cex}{to draw components}

\item{pch}{to draw components}

\item{label}{to indicate first point}

\item{labels}{\code{character} name of labels to draw (defaut to \code{1:nrow(coo)})}

\item{d}{`numeric` proportion of `d(centroid-each_point)` to add when centrifugating landmarks}

\item{links}{`matrix` of links to use to draw segments between landmarks. See `wings$ldk` for an example}

\item{d}{\code{numeric} how far away on a segment should the labels be displayed
(on a centroid-point) segment.}

\item{links}{\code{matrix} of links}
}
\description{
Useful drawers for building custom
shape plots using the cheapbabi approach. See examples and vignettes.
}
\note{
This approach will (soon) replace [coo_plot] and friends in further versions.
All comments are welcome.
}
\examples{
bot[1] \%>\% paper_grid() \%>\% draw_polygon()
olea \%>\% paper_chess \%>\% draw_lines(~var)

hearts[240] \%>\% paper_white() \%>\% draw_outline() \%>\%
  coo_sample(24) \%>\% draw_landmarks \%>\% draw_labels() \%>\%
  draw_links(replicate(2, sample(1:24, 8)))
}
\seealso{
cheapbabi_layers

Other graphics cheapbabi: \code{\link{cheapbabi_layers}}
}
