% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{chop}
\alias{chop}
\alias{chop.Coe}
\alias{chop.Coo}
\alias{chop.default}
\title{Rougher slicing that accepts a classifier ie a column name from the $fac on Momocs classes}
\usage{
chop(.data, fac)

\method{chop}{default}(.data, fac)

\method{chop}{Coo}(.data, fac)

\method{chop}{Coe}(.data, fac)
}
\arguments{
\item{.data}{a \code{Coo} or \code{Coe} object}

\item{fac}{a column name from the $fac}
}
\value{
a named list of \code{Coo} or \code{Coe} objects
}
\description{
Returns a named (after every level) list that can be lapply-ed and combined. See examples.
}
\examples{
data(bot)
}
\seealso{
\link{select}, \link{filter}, \link{slice}, \link{chop}, \link{combine}.
}

