% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mult-clust.R
\name{CLUST}
\alias{CLUST}
\title{Hierarchical clustering}
\usage{
CLUST(x, fac, method, type, palette, ...)
}
\arguments{
\item{x}{an Coe object}

\item{fac}{a column or an id from $fac}

\item{method}{the method to use among \code{'euclidean'}, \code{'maximum'},
\code{'manhattan'}, \code{'canberra'}, \code{'binary'} or \code{'minkowski'} to feed \link{dist}}

\item{type}{the type of plot among \code{'phylogram'} , \code{'cladogram'},
\code{'fan'}, \code{'unrooted'} and \code{'radial'} (the default) to feed \link{plot.phylo}}

\item{palette}{a color palette to use}

\item{...}{additional arguments to fed \link{plot.phylo}}
}
\value{
a (invisible) list with the following components:
\itemize{
\item dist.mat the distance matrix
\item hclust the hclust object
}
}
\description{
Performs hierarchical clustering through \link{dist} and \link{hclust}. So far it is mainly
a wrapper around these two functions, plus plotting using \link{plot.phylo} from the
package ape.
}
\examples{
data(bot)
bot.f <- efourier(bot, 12)
CLUST(bot.f)
bot.c <- CLUST(bot.f, 'type', cex=0.7)
#bot.c

data(olea)
olea.VL <- filter(olea, view=='VL')
op <- npoly(olea.VL, 5)
CLUST(op, 1, type='cladogram', palette=col_autumn, cex=0.5)

data(wings)
wp <- fgProcrustes(wings)
CLUST(wp, 1, cex=0.5)
}
\keyword{Graphics}
\keyword{Multivariate}

