% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clust}
\alias{clust}
\title{Hierarchical clustering #todo}
\usage{
clust(OutCoe, fac, method, type, palette, ...)
}
\arguments{
\item{OutCoe}{an OutCoe object}

\item{fac}{a colum or an id from $fac}

\item{method}{the method to use}

\item{type}{the type of plot}

\item{palette}{a color palette to use}

\item{...}{additional arguments to fed plot.phylo}
}
\description{
Hierarchical clustering #todo
}
\keyword{Multivariate}

