% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clust}
\alias{clust}
\title{Hierarchical clustering}
\usage{
clust(x, fac, method, type, palette, ...)
}
\arguments{
\item{x}{an Coe object}

\item{fac}{a column or an id from $fac}

\item{method}{the method to use among \code{'euclidean'}, \code{'maximum'},
\code{'manhattan'}, \code{'canberra'}, \code{'binary'} or \code{'minkowski'} to feed \link{dist}}

\item{type}{the type of plot among \code{'phylogram'} (the default), \code{'cladogram'},
\code{'fan'}, \code{'unrooted'} and \code{'radial'} to feed \link{plot.phylo}}

\item{palette}{a color palette to use}

\item{...}{additional arguments to fed \link{plot.phylo}}
}
\value{
a (invisible) list with the following components:
\itemize{
\item dist.mat the distance matrix
\item hclust the hclust object
}
}
\description{
Performs hierarchical clustering through \link{dist} and \link{hclust}. So far it is mainly
a wrapper around these two functions, plus plotting using \link{plot.phylo} from the
package ape.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
clust(bot.f)
bot.c <- clust(bot.f, 'type', cex=0.7)
#bot.c

data(olea)
olea.VL <- subset(olea, view=='VL')
op <- rawPolynomials(olea.VL, 5)
clust(op, 1, type='cladogram', palette=col.autumn, cex=0.5)

data(wings)
wp <- fgProcrustes(wings)
clust(wp, 1)
}
\keyword{Graphics}
\keyword{Multivariate}

