% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Coe}
\alias{Coe}
\title{Coe class}
\usage{
Coe(...)
}
\arguments{
\item{...}{anything and, anyway, this function will simply returns a message.}
}
\description{
\code{Coe} class is a super class for \link{OutCoe} and \link{OpnCoe} classes,
matrices of coefficients, along with other informations, obtained with
morphometrics methods on \code{\link{Out}} and \code{\link{Opn}} objects.
}
\details{
It allows to recycle most of the methods, since both of them apply on
matrices of coefficients. In other words, \code{OutCoe} and \code{OpnCoe}
classes are all, primarily, \code{Coe} objects on which we define generic
\emph{and} specific methods.

More, generally, see \link{Opn} and \link{Out} for documentation on these classes.
}
\keyword{Coe}

