% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Coe}
\alias{Coe}
\title{Coe class}
\usage{
Coe(...)
}
\arguments{
\item{...}{anything and, anyway, this function will simply returns a message.}
}
\description{
\code{Coe} class is the 'parent' class of
\code{\link{OutCoe}}, \code{\link{OpnCoe}} and \code{\link{LdkCoe}} classes.
}
\details{
\code{Coe} class is the 'parent' class of the following 'child' classes \itemize{
\item \code{\link{OutCoe}} for coefficients from closed \bold{out}lines morphometrics
\item \code{\link{OpnCoe}} for coefficients from \bold{op}e\bold{n} outlines morphometrics
\item \code{\link{LdkCoe}} for coefficients from configuration of \bold{l}an\bold{d}mar\bold{k}s morphometrics.
}

In other words, \code{\link{OutCoe}}, \code{\link{OpnCoe}} and \code{\link{LdkCoe}} classes
are all, primarily, \code{Coe} objects on which we define generic \emph{and}
specific methods. See their respective help pages for more help.

You can access all the methods available for \code{Coe} objects
with \code{methods(class=Coe)}.
}
\examples{
# to see all methods for Coo objects.
methods(class='Coe')

data(bot)
bot.f<- efourier(bot, 12)
bot.f
class(bot.f)
inherits(bot.f, "Coe")

# if you want to work directly on the matrix of coefficients
bot.f$coe

data(olea)
op <- opoly(olea, 5)
op
class(op)
op$coe # same thing

data(wings)
wp <- fgProcrustes(wings, tol=1e-4)
wp
class(wp) # for Ldk methods, LdkCoe objects can also be considered as Coo objects
# so you can apply all Ldk methods available.
wp$coe # Procrustes aligned coordinates
}

