% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col.summer}
\alias{col.autumn}
\alias{col.blackgallus}
\alias{col.bw}
\alias{col.cold}
\alias{col.gallus}
\alias{col.hot}
\alias{col.india}
\alias{col.sari}
\alias{col.solarized}
\alias{col.spring}
\alias{col.summer}
\title{Some color palettes.}
\usage{
col.summer(n)

col.spring(n)

col.autumn(n)

col.solarized(n)

col.gallus(n)

col.blackgallus(n)

col.hot(n)

col.cold(n)

col.sari(n)

col.india(n)

col.bw(n)
}
\arguments{
\item{n}{the number of colors to generate from the color palette}
}
\value{
colors (hexadecimal format)
}
\description{
Colors, colors, colors.
}
\examples{
barplot(1:10, col=col.summer(10), main="col.summer")
barplot(1:10, col=col.spring(10), main="col.spring")
barplot(1:10, col=col.autumn(10), main="col.autumn")
barplot(1:10, col=col.solarized(10), main="col.solarized")
barplot(1:10, col=col.gallus(10), main="col.gallus")
barplot(1:10, col=col.blackgallus(10), main="col.blackgallus")
barplot(1:10, col=col.hot(10), main="col.hot")
barplot(1:10, col=col.cold(10), main="col.cold")
barplot(1:10, col=col.sari(10), main="col.sari")
barplot(1:10, col=col.india(10), main="col.india")
barplot(1:10, col=col.bw(10), main="col.bw")
}
\keyword{Graphics}

