% Generated by roxygen2 (4.0.2): do not edit by hand
\name{col.summer}
\alias{Palettes}
\alias{col.autumn}
\alias{col.black}
\alias{col.bw}
\alias{col.cold}
\alias{col.gallus}
\alias{col.grey}
\alias{col.heat}
\alias{col.hot}
\alias{col.india}
\alias{col.qual}
\alias{col.sari}
\alias{col.solarized}
\alias{col.spring}
\alias{col.summer}
\alias{col.summer2}
\title{Some color palettes.}
\usage{
col.summer(n)

col.summer2(n)

col.spring(n)

col.autumn(n)

col.black(n)

col.solarized(n)

col.gallus(n)

col.qual(n)

col.heat(n)

col.hot(n)

col.cold(n)

col.sari(n)

col.india(n)

col.bw(n)

col.grey(n)
}
\arguments{
\item{n}{the number of colors to generate from the color palette}
}
\value{
colors (hexadecimal format)
}
\description{
Colors, colors, colors.
}
\note{
\itemize{
\item \code{col.solarized} is based on Solarized: \url{http://ethanschoonover.com/solarized};
\item \code{col.div} on ColorBrewer2: \url{http://colorbrewer2.org/};
\item \code{col.qual}, \code{col.heat}, \code{col.cold} and \code{col.gallus} use the colorspace package.
}
}
\examples{
wheel <- function(palette, n=10){
 op <- par(mar=rep(0, 4)) ; on.exit(par(op))
 pie(rep(1, n), col=palette(n), labels=NA, clockwise=TRUE)}

 # Qualitative
 wheel(col.qual)
 wheel(col.solarized)
 wheel(col.summer)
 wheel(col.summer2)
 wheel(col.spring)
 wheel(col.autumn)

 # Divergent
 wheel(col.gallus)
 wheel(col.india)

 # Sequential
 wheel(col.heat)
 wheel(col.hot)
 wheel(col.cold)
 wheel(col.sari)
 wheel(col.bw)
 wheel(col.grey)

 # Black only for pubs
 wheel(col.black)
}
\keyword{Graphics}

