% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col.transp}
\alias{.transp}
\alias{col.transp}
\title{Transparency helpers and palettes}
\usage{
col.transp(n, col = "#000000", ceiling = 1)

.transp(cols, transp = 0)
}
\arguments{
\item{n}{the number of colors to generate}

\item{col}{a color in hexadecimal format on which to generate levels of transparency}

\item{ceiling}{the maximal opacity (from 0 to 1)}

\item{cols}{on or more colors, provided as hexadecimal values}

\item{transp}{numeric between 0 and 1, the value of the transparency to obtain}
}
\description{
To ease transparency handling.
}
\examples{
x <- col.transp(10, col='#000000')
x
barplot(1:10, col=x, main='a transparent black is grey')

summer10 <- col.summer(10)
summer10
summer10.transp8 <- .transp(summer10, 0.8)
summer10.transp8
summer10.transp2 <- .transp(summer10, 0.8)
summer10.transp2
x <- 1:10
barplot(x, col=summer10.transp8)
barplot(x/2, col=summer10.transp2, add=TRUE)
}
\keyword{Graphics}

