% Generated by roxygen2 (4.0.0): do not edit by hand
\name{combine}
\alias{combine}
\alias{combine.Ldk}
\alias{combine.Opn}
\alias{combine.OpnCoe}
\alias{combine.Out}
\alias{combine.OutCoe}
\title{Combine Out objects}
\usage{
combine(...)

\method{combine}{Out}(...)

\method{combine}{Opn}(...)

\method{combine}{Ldk}(...)

\method{combine}{OutCoe}(...)

\method{combine}{OpnCoe}(...)
}
\arguments{
\item{...}{a list of Out objects}
}
\description{
Combine Out objects
}
\seealso{
\link{subset.Coo}
}

