\name{conf.ell}
\alias{conf.ell}
\title{Confidence ellipses}
\usage{
conf.ell(x, y, conf = 0.95, nb.pts = 60)
}
\arguments{
  \item{x}{numeric values on the x axis}

  \item{y}{numeric values on the y axis}

  \item{conf}{the level of confidence}

  \item{nb.pts}{the number of points to return, to draw the
  ellipsis}
}
\value{
a matrix of (x; y) coordinates to draw the ellipsis
}
\description{
Draw (gaussian) confidence ellipses
}
\keyword{graphics}

