% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.align}
\alias{coo.align}
\title{Aligns coordinates}
\usage{
coo.align(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Aligns the coordinates along their longer axis using var-cov matrix and eigen values.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.align(bot))
coo.plot(b)
coo.plot(coo.align(b))
}
\seealso{
\link{coo.aligncalliper}, \link{coo.alignxax}
}
\keyword{ShapeUtilities}

