% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.aligncalliper}
\alias{coo.aligncalliper}
\title{Aligns shapes along their 'calliper length'}
\usage{
coo.aligncalliper(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
And returns them registered on bookstein coordinates.
See \link{coo.bookstein}.
}
\examples{
\dontrun{
data(bot)
b <- bot[1]
coo.plot(b)
coo.plot(coo.aligncalliper(b))
bot.al <- coo.aligncalliper(bot)
stack(bot.al)
}
}
\seealso{
\link{coo.align}, \link{coo.alignxax}, \link{coo.calliper}
}
\keyword{ShapeUtilities}

