% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.alignxax}
\alias{coo.alignxax}
\title{Aligns shapes along the x-axis}
\usage{
coo.alignxax(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Align the longest axis of a shape along the x-axis.
}
\examples{
\dontrun{
data(bot)
b <- bot[1]
coo.plot(b)
coo.plot(coo.alignxax(b))
}
}
\seealso{
\link{coo.align}, \link{coo.aligncalliper}
}
\keyword{ShapeUtilities}

