% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.area}
\alias{coo.area}
\title{Calculates the area of a shape}
\usage{
coo.area(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric}, the area.
}
\description{
Calculates the area for a shape using \link{area.poly} in gpc package.
}
\examples{
data(bot)
coo.area(bot[1])
# for the distribution of the area of the bottles dataset
hist(sapply(bot$coo, coo.area), breaks=10)
}
\keyword{ShapeDescriptors}

