% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.calliper}
\alias{coo.calliper}
\title{Calculates the calliper length}
\usage{
coo.calliper(coo, arr.ind = FALSE)
}
\arguments{
\item{coo}{a \code{matrix} of \eqn{(x; y)} coordinates.}

\item{arr.ind}{a boolean, if provided returns}
}
\value{
\code{numeric}, the centroid size. If arr.ind=TRUE, a list with the calliper length ($length)
and the two points ($arr.ind)
}
\description{
Also called the Feret's diameter, the longest distance between two points of
the shape provided.
}
\examples{
data(bot)
b <- bot[1]
coo.calliper(b)
p <- coo.calliper(b, arr.ind=TRUE)
p$length
ids <- p$arr.ind
coo.plot(b)
segments(b[ids[1], 1], b[ids[1], 2], b[ids[2], 1], b[ids[2], 2], lty=2)
}
\keyword{coo_descriptors}

