% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.center}
\alias{coo.center}
\title{Centers coordinates}
\usage{
coo.center(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a shape centered on the origin.
}
\examples{
data(bot)
b <- bot[1]
}
\keyword{ShapeUtilities}

