% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.centpos}
\alias{coo.centpos}
\title{Returns the position of the centroid}
\usage{
coo.centpos(coo)
}
\arguments{
\item{coo}{a \code{matrix} of \eqn{(x; y)} coordinates.}
}
\value{
\eqn{(x; y)} coordinates as \code{numeric}.
}
\description{
Returns the \eqn{(x; y)} centroid coordinates of a shape.
}
\examples{
data(bot)
b <- bot[1]
coo.plot(b)
xy <- coo.centpos(b)
points(xy[1], xy[2], cex=2, col="blue")
xy
}
\keyword{coo_utilities}

