% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.centsize}
\alias{coo.centsize}
\title{Calculates the centroid size}
\usage{
coo.centsize(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric}, the centroid size.
}
\description{
Calculates the centroid size
}
\examples{
data(bot)
coo.centsize(bot[1])
cs <- sapply(bot$coo, coo.centsize)
hist(cs, breaks=10)
}
\keyword{ShapeDescriptors}

