% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.check}
\alias{coo.check}
\title{Checks "coo" shapes}
\usage{
coo.check(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a list.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
A simple utility, used internally, mostly in the coo functions and methods.
Returns a matrix of coordinates, when passed with either a list or a \code{matrix} of coordinates.
}
\examples{
#coo.check("Not a shape")
#coo.check(matrix(1:10, ncol=2))
#coo.check(list(x=1:5, y=6:10))
}
\seealso{
\link{ldk.check}
}
\keyword{ShapeUtilities}

