% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.circularity.haralick}
\alias{coo.circularity.haralick}
\title{Calculates the Haralick's circularity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.circularity.haralick(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the Haralick's circularity.
}
\description{
Returns Haralick's circularity which is less sensible
to digitalization noise than coo.circularity.
}
\examples{
data(bot)
coo.circularity.haralick(bot[1])
}
\keyword{Global}
\keyword{descriptors}
\keyword{shape}

