% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.circularity.norm}
\alias{coo.circularity.norm}
\title{Calculates the "normalized" circularity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.circularity.norm(coo)
}
\arguments{
\item{coo}{a \code{matrix} of \eqn{(x; y)} coordinates.}
}
\value{
numeric, the circularity normalized to the unit circle.
}
\description{
Returns the "circularity", also called compactness
}
\examples{
data(bot)
coo.circularity.norm(bot[1])
}
\keyword{coo_descriptors}

