% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.circularity}
\alias{coo.circularity}
\alias{coo.compactness}
\alias{coo.shapefactor}
\title{Calculates the circularity of a shape}
\usage{
coo.circularity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the circularity.
}
\description{
Returns the "circularity measure". Also called "compactness"
and "shape factor" sometimes.
}
\examples{
data(bot)
coo.circularity(bot[1])
}
\keyword{ShapeDescriptors}

