% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.close}
\alias{coo.close}
\title{Closes/'Uncloses' shapes}
\usage{
coo.close(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a closed shape from (un)closed shapes. See also \link{coo.unclose}.
}
\examples{
x <- (matrix(1:10, ncol=2))
x2 <- coo.close(x)
x3 <- coo.unclose(x2)
x
is.closed(x)
x2
is.closed(x2)
x3
is.closed(x3)
}
\seealso{
\link{coo.unclose}, \link{is.closed}
}
\keyword{ShapeUtilities}

