% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.draw}
\alias{coo.draw}
\title{Adds a shape to the current plot}
\usage{
coo.draw(coo, ...)
}
\arguments{
\item{coo}{a \code{list} or a \code{matrix} of coordinates.}

\item{...}{optional parameters for \link{coo.plot}}
}
\description{
\code{coo.draw} is simply a \link{coo.plot} with \code{plot.new=FALSE}, ie
that adds a shape on the active plot.
}
\examples{
data(bot)
b1 <- bot[4]
b2 <- bot[5]
coo.plot(b1)
coo.draw(b2, border="red") # all coo.plot arguments will work for coo.draw
}
\keyword{Graphics}

