% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.eccentricity.eigen}
\alias{coo.eccentricity.eigen}
\title{Calculates the eccentricity (using eigenvalues) of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.eccentricity.eigen(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the eccentricity (eigenvalues)
}
\description{
Calculated using a ratio of the eigen values (inertia axes oof coordinates.)
}
\examples{
data(bot)
coo.eccentricity.eigen(bot[1])
}
\seealso{
\link{coo.eccentricity.boundingbox}
}
\keyword{Global}
\keyword{descriptors}
\keyword{shape}

