% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.elongation}
\alias{coo.elongation}
\title{Calculates the elongation of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.elongation(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the circularity normalized to the unit circle.
}
\description{
Calculates the elongation of a shape
}
\examples{
data(bot)
coo.elongation(bot[1])
}
\keyword{ShapeDescriptors}

