% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.force2close}
\alias{coo.force2close}
\title{Forces shapes to close}
\usage{
coo.force2close(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of \eqn{(x; y)} coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of \eqn{(x; y)} coordinates, or a \link{Coo} object.
}
\description{
An exotic function that distribute the distance between the first and the last points
of unclosed shapes, so that they become closed. May be useful (?) e.g. for t/rfourier methods
where reconstructed shapes may not be closed.
}
\examples{
data(bot)
b <- coo.sample(bot[1], 64)
b <- b[1:40,]
coo.plot(b)
coo.draw(coo.force2close(b), border="red")
}
\keyword{coo_utilities}

