% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.list.panel}
\alias{coo.list.panel}
\title{Plots sets of shapes.}
\usage{
coo.list.panel(coo.list, dim, byrow = TRUE, fromtop = TRUE, mar = rep(0,
  4), cols, borders, reorder = NULL, poly = TRUE, points = FALSE,
  points.pch = 3, points.cex = 0.2, points.col = "#333333")
}
\arguments{
\item{coo.list}{A \code{list} of coordinates}

\item{dim}{A \code{vector} of the form \code{(nb.row, nb.cols)} to specify
the panel display. If missing, shapes are arranged in a square.}

\item{byrow}{\code{logical}. Whether to succesive shape by row or by col.}

\item{fromtop}{\code{logical}. Whether to display shapes from the top of the
plotting region.}

\item{mar}{A \code{vector} to define margins.}

\item{cols}{A \code{vector} of colors to fill shapes.}

\item{borders}{A \code{vector} of colors to draw shape borders.}

\item{reorder}{a factor or a numeric to reorder shapes, colors and borders.}

\item{poly}{logical whether to use polygon or lines to draw shapes.
mainly for use for outlines and open outlines.}

\item{points}{logical if poly is set to FALSE whether to add points}

\item{points.pch}{if points is TRUE, a pch for these points}

\item{points.cex}{if points is TRUE, a cex for these points}

\item{points.col}{if points is TRUE, a col  for these points}
}
\value{
Returns (invisibly) a \code{data.frame} with position of shapes that
can be used for other sophisticated plotting design.
}
\description{
\code{coo.list.panel} plots a list of shapes if passed with a list of
coordinates. Outlines are templated and on the same graphical window with
the help of \link{coo.template}.
}
\examples{
data(bot)
coo.list.panel(bot$coo)
x <- coo.list.panel(bot$coo)
x # positions of shapes returned invisibly
# axis(1) ; axis(2) # that's a single graphical window

data(bot)
coo <- bot$coo
ord <- sapply(coo, coo.eccentricity.eigen)
pos <- coo.list.panel(coo, reorder=ord)
text(pos, labels=signif(ord[order(ord)], 3))
}
\seealso{
\link{coo.plot} and \link{coo.template}.
}
\keyword{Graphics}

