% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.perim.pts}
\alias{coo.perim.pts}
\title{Calculates the chordal distance along a shape.}
\usage{
coo.perim.pts(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric} the distance between every point.
}
\description{
Calculates the euclidean distance between every points of a shape for coo.perim.pts.
The cumulative sum for coo.perim.cum
}
\examples{
data(bot)
b <- coo.sample(bot[1], 24)
coo.perim.pts(b)
}
\keyword{ShapeDescriptors}

