% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.perim}
\alias{coo.perim}
\title{Calculates the perimeter}
\usage{
coo.perim(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric}, the perimeter.
}
\description{
Calculates the perimeter
}
\examples{
data(bot)
coo.perim(bot[1])
hist(sapply(bot$coo, coo.perim), breaks=10)
}
\keyword{coo_descriptors}

