% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.rectangularity}
\alias{coo.rectangularity}
\title{Calculates the rectangularity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.rectangularity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the rectangularity.
}
\description{
Calculates the rectangularity of a shape
}
\examples{
data(bot)
coo.rectangularity(bot[1])
}
\keyword{ShapeDescriptors}

