% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.rev}
\alias{coo.rev}
\title{Reverses coordinates}
\usage{
coo.rev(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
Returns the reverse suite of coordinates, i.e. change shape's orientation
}
\examples{
data(bot)
b <- coo.sample(bot[1], 4)
b
coo.rev(b)
}
\keyword{ShapeUtilities}

