% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.rotate}
\alias{coo.rotate}
\title{Rotates coordinates}
\usage{
coo.rotate(coo, theta = 0)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{theta}{\code{numeric}the angle (in radians) to rotate shapes.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Rotates the coordinates by a 'theta' angle (in radians) If not provided,
assumed to be the centroid size. It involves three steps: centering from current position,
dividing coordinates by 'scale', pusing back to the original position.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.rotate(bot, pi))
coo.plot(b)
coo.plot(coo.rotate(b, pi))
}
\seealso{
\link{coo.rotatecenter}
}
\keyword{ShapeUtilities}

