% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.rotatecenter}
\alias{coo.rotatecenter}
\title{Rotates shapes with a custom center}
\usage{
coo.rotatecenter(coo, theta, center = c(0, 0))
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{theta}{\code{numeric} the angle (in radians) to rotate shapes.}

\item{center}{the (x; y) position of the center}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
rotates a shape of "theta" angles (in radians) and with a (x; y) "center".
}
\examples{
data(bot)
b <- bot[1]
coo.plot(b)
coo.draw(coo.rotatecenter(b, -pi/2, c(200, 200)), border="red")
}
\seealso{
\link{coo.rotate}
}
\keyword{ShapeUtilities}

