% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.scale}
\alias{coo.scale}
\title{Scales coordinates}
\usage{
coo.scale(coo, scale)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{scale}{the scaling factor, by default, the centroid size.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Scales the coordinates by a 'scale' factor. If not provided,
assumed to be the centroid size. It involves three steps: centering from current position,
dividing coordinates by 'scale', pusing back to the original position.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.scale(bot))
coo.plot(b)
coo.plot(coo.scale(b))
}
\keyword{ShapeUtilities}

