% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.smooth}
\alias{coo.smooth}
\title{Smoothes coordinates (closed outlines)}
\usage{
coo.smooth(coo, n)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{n}{an (integer) \code{numeric} to specify the number of smoothing iterations}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Smoothes coordinates using a simple moving average.
May be useful to remove digitization noise.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.smooth(bot, 10))
coo.plot(bot[1])
coo.plot(coo.smooth(bot[1], 30))
}
\seealso{
\link{coo.smoothcurve}
}
\keyword{ShapeUtilities}

