% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.solidity}
\alias{coo.solidity}
\title{Calculates the solidity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo.solidity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the solidity of a shape.
}
\description{
Returns the ids of points that define the convex hull of a shape.
}
\examples{
data(bot)
coo.solidity(bot[1])
}
\keyword{Global}
\keyword{descriptors}
\keyword{shape}

