% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo.trans}
\alias{coo.trans}
\title{Translates coordinates}
\usage{
coo.trans(coo, x = 0, y = 0)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{x}{\code{numeric}translation along the x-axis.}

\item{y}{\code{numeric}translation along the y-axis.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Translates the coordinatesby a 'x' and 'y' value
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo.trans(bot, 50, 100))
coo.plot(b)
coo.plot(coo.trans(b, 50, 100))
}
\keyword{ShapeUtilities}

