% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coo.unclose}
\alias{coo.unclose}
\title{'Uncloses' shapes}
\usage{
coo.unclose(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of \eqn{(x; y)} coordinates, or a \link{Coo}
object.}
}
\value{
a \code{matrix} of \eqn{(x; y)} coordinates, or a \link{Coo} object.
}
\description{
Returns a unclosed shape from (un)closed shapes. See also \link{coo.close}.
}
\examples{
x <- (matrix(1:10, ncol=2))
x2 <- coo.close(x)
x3 <- coo.unclose(x2)
x
is.closed(x)
x2
is.closed(x2)
x3
is.closed(x3)
}
\seealso{
\link{coo.close}, \link{is.closed}
}
\keyword{coo_utilities}

