% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_align}
\alias{coo_align}
\title{Aligns coordinates}
\usage{
coo_align(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Aligns the coordinates along their longer axis using var-cov matrix and eigen values.
}
\examples{
coo_plot(bot[1])
coo_plot(coo_align(bot[1]))
# on a Coo
stack(bot)
stack(coo_align(bot))
}
\seealso{
\link{coo_aligncalliper}, \link{coo_alignxax}

Other aligning functions: \code{\link{coo_aligncalliper}},
  \code{\link{coo_alignminradius}},
  \code{\link{coo_alignxax}}
}

