% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_alignminradius}
\alias{coo_alignminradius}
\title{Aligns shapes using their shortest radius}
\usage{
coo_alignminradius(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
And returns them slided with the first coordinate on the east.
May be used as an aligning strategy on shapes with a clear 'invaginate' part.
}
\examples{
\dontrun{
stack(coo_alignminradius(hearts))
}
}
\seealso{
Other aligning functions: \code{\link{coo_aligncalliper}},
  \code{\link{coo_alignxax}}, \code{\link{coo_align}}
}

