% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_alignxax}
\alias{coo_alignxax}
\title{Aligns shapes along the x-axis}
\usage{
coo_alignxax(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Align the longest axis of a shape along the x-axis.
}
\details{
If some shapes are upside-down
(or mirror of each others), try redefining a new starting point (eg with coo_slidedirection) before
the alignment step. This may solve your problem because coo_calliper orders the \code{$arr.ind} used by
coo_aligncalliper.
}
\examples{
\dontrun{
data(bot)
b <- bot[1]
coo_plot(b)
coo_plot(coo_alignxax(b))
}
}
\seealso{
\link{coo_align}, \link{coo_aligncalliper}
}
\keyword{ShapeUtilities}

