% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_bookstein}
\alias{coo_bookstein}
\title{Register Bookstein's coordinates}
\usage{
coo_bookstein(coo, ldk1, ldk2)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{ldk1}{the id of the first point of the new baseline (the first, by default)}

\item{ldk2}{the id of the second point of the new baseline (the last, by default)}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Registers a new baseline for the shape, with the \code{ldk1}-th
and \code{ldk2}-th points being set on \eqn{(x= -0.5; y=0)} and \eqn{(x= 0.5; y=0)}, respectively.
}
\details{
For \link{Out}, it tries to do it using \code{$ldk} slot. Also the case for \link{Opn}, but if
no landmark is defined, it will do it on the first and the last point of the shape.

For \code{Out} and \code{Opn} defines the first landmark as the first point of the
new shapes with \link{coo_slide}.
}
\examples{
data(hearts)
stack(hearts)
stack(coo_bookstein(hearts, 2, 4))
h <- hearts[1]
coo_plot(h)
coo_plot(coo_bookstein(h, 20, 57), border='red')
}
\seealso{
\link{coo_baseline}
}
\keyword{ShapeUtilities}

