% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_centdist}
\alias{coo_centdist}
\title{Returns the distance between everypoints and the centroid}
\usage{
coo_centdist(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
For every point of the shape, returns the (centroid-points) distance.
}
\examples{
data(bot)
b <- coo_sample(bot[1], 64)
d <- coo_centdist(b)
barplot(d)
}
\keyword{ShapeDescriptors}

