% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_center}
\alias{coo_center}
\alias{coo_centre}
\title{Centers coordinates}
\usage{
coo_center(coo)

coo_centre(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a shape centered on the origin.

brittons...
}
\examples{
data(bot)
b <- bot[1]
}

