% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_center}
\alias{coo_center}
\title{Centers coordinates}
\usage{
coo_center(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Returns a shape centered on the origin.
}
\examples{
data(bot)
b <- bot[1]
}
\keyword{ShapeUtilities}

