% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_centpos}
\alias{coo_centpos}
\title{Calculate centroid coordinates}
\usage{
coo_centpos(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
(x; y) coordinates of the centroid as a vector or a matrix.
}
\description{
Returns the (x; y) centroid coordinates of a shape.
}
\examples{
b <- bot[1]
coo_plot(b)
xy <- coo_centpos(b)
points(xy[1], xy[2], cex=2, col='blue')
# on a Coo
coo_centpos(bot)
}
\seealso{
Other centroid functions: \code{\link{coo_centdist}},
  \code{\link{coo_centsize}}
}

