% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_centsize}
\alias{coo_centsize}
\title{Calculates the centroid size}
\usage{
coo_centsize(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a Coo object}
}
\value{
\code{numeric}, the centroid size.
}
\description{
Calculates the centroid size
}
\details{
This function can be used to integrate size - if meaningful -
to Coo objects. See also \link{coo_length} and \link{rescale}.
}
\examples{
data(bot)
coo_centsize(bot[1])
coo_centsize(bot)

mutate(bot, size=coo_centsize(bot))
}
\keyword{ShapeDescriptors}

