% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_centsize}
\alias{coo_centsize}
\title{Calculates the centroid size}
\usage{
coo_centsize(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric}, the centroid size.
}
\description{
Calculates the centroid size
}
\examples{
data(bot)
coo_centsize(bot[1])
cs <- sapply(bot$coo, coo_centsize)
hist(cs, breaks=10)
}
\keyword{ShapeDescriptors}

