% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_check}
\alias{coo_check}
\title{Checks shapes}
\usage{
coo_check(coo)
}
\arguments{
\item{coo}{\code{matrix} of \code{(x; y)} coordinates or any \link{Coo} object.}
}
\value{
\code{matrix} of \code{(x; y)} coordinates or a \link{Coo} object.
}
\description{
A simple utility, used internally, mostly in the coo functions and methods.
Returns a matrix of coordinates, when passed with either a list or a \code{matrix} of coordinates.
}
\examples{
#coo_check('Not a shape')
#coo_check(iris)
#coo_check(matrix(1:10, ncol=2))
#coo_check(list(x=1:5, y=6:10))
}
