% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_convexity}
\alias{coo_convexity}
\title{Calculates the convexity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_convexity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the convexity.
}
\description{
Calculated using a ratio of the eigen values (inertia axis)
}
\examples{
data(bot)
coo_convexity(bot[1])
}
\keyword{ShapeDescriptors}

