% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_down}
\alias{coo_down}
\title{coo_down -----------------
Retains coordinates with negative y-coordinates}
\usage{
coo_down(coo, slidegap = FALSE)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{slidegap}{\code{logical} whether to apply \link{coo_slidegap} after coo_down}
}
\value{
a \code{matrix} of (x; y) coordinates or a \link{Coo} object (\link{Out} are returned as \link{Opn})
}
\description{
Useful when shapes are aligned along the x-axis (e.g. because of a
bilateral symmetry) and when one wants to retain just the lower side.
}
\note{
When shapes are "sliced" along the x-axis, it usually results on open curves and thus to huge/artefactual
gaps between points neighboring this axis. This is usually solved with \link{coo_slidegap}. See examples there.

Also, when apply a coo_left/right/up/down on an \link{Out} object, you then obtain an \link{Opn} object, which is done
automatically.
}
\examples{
b <- coo_alignxax(bot[1])
coo_plot(b)
coo_draw(coo_down(b), border='red')
}
\seealso{
Other opening functions: \code{\link{coo_left}},
  \code{\link{coo_right}}, \code{\link{coo_up}}
}

