% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{coo_draw}
\alias{coo_draw}
\title{Adds a shape to the current plot}
\usage{
coo_draw(coo, ...)
}
\arguments{
\item{coo}{a \code{list} or a \code{matrix} of coordinates.}

\item{...}{optional parameters for \link{coo_plot}}
}
\description{
\code{coo_draw} is simply a \link{coo_plot} with \code{plot.new=FALSE}, ie
that adds a shape on the active plot.
}
\examples{
b1 <- bot[4]
b2 <- bot[5]
coo_plot(b1)
coo_draw(b2, border='red') # all coo_plot arguments will work for coo_draw
}
\seealso{
Other plotting functions: \code{\link{coo_arrows}},
  \code{\link{coo_listpanel}}, \code{\link{coo_lolli}},
  \code{\link{coo_plot}}, \code{\link{coo_ruban}},
  \code{\link{ldk_chull}}, \code{\link{ldk_confell}},
  \code{\link{ldk_contour}}, \code{\link{ldk_labels}},
  \code{\link{ldk_links}}, \code{\link{plot_devsegments}},
  \code{\link{plot_table}}
}
