% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_dxy}
\alias{coo_dxy}
\title{Calculate abscissa and ordinate on a shape}
\usage{
coo_dxy(coo)
}
\arguments{
\item{coo}{a matrix (or a list) of (x; y) coordinates}
}
\value{
a list with two components \code{dx} and \code{dy}
}
\description{
A simple wrapper to calculate dxi - dx1 and dyi - dx1.
}
\examples{
coo_dxy(bot[1])
}

