% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_eccentricityeigen}
\alias{coo_eccentricityeigen}
\title{Calculates the eccentricity (using eigenvalues) of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_eccentricityeigen(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the eccentricity (eigenvalues)
}
\description{
Calculated using a ratio of the eigen values (inertia axes oof coordinates.)
}
\examples{
data(bot)
coo_eccentricityeigen(bot[1])
}
\seealso{
\link{coo_eccentricityboundingbox}
}
\keyword{ShapeDescriptors}

