% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_flipx}
\alias{coo_flipx}
\alias{coo_flipy}
\title{Flips shapes}
\usage{
coo_flipx(coo)

coo_flipy(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
a \code{matrix} of (x; y) coordinates
}
\description{
\code{coo_flipx} flips shapes about the x-axis; \code{coo_flipy} about the y-axis.
}
\examples{
cat <- shapes[4]
cat <- coo_center(cat)
coo_plot(cat)
coo_draw(coo_flipx(cat), border="red")
coo_draw(coo_flipy(cat), border="blue")

#' # to flip an entire Coo:
shapes2 <- shapes
shapes$coo <- lapply(shapes2$coo, coo_flipx)
}

