% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_lw}
\alias{coo_lw}
\title{Calculates length and width of a shape}
\usage{
coo_lw(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
a vector of two \code{numeric}: the length and the width.
}
\description{
Returns the length and width of a shape based on their iniertia axis
i.e. alignment to the x-axis. The length is defined as
the range along the x-axis; the width as the range on the y-axis.
}
\examples{
data(bot)
coo_lw(bot[1])
}
\seealso{
\link{coo_length}, \link{coo_width}.
}
\keyword{ShapeDescriptors}

