% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_nb}
\alias{coo_nb}
\title{Counts coordinates}
\usage{
coo_nb(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}
}
\value{
either a single numeric or a vector of numeric
}
\description{
Returns the number of coordinates, for a single shape or a Coo object
}
\examples{
# single shape
coo_nb(bot[1])
# Coo object
coo_nb(bot)
}

