% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_perim}
\alias{coo_perim}
\title{Calculates the perimeter}
\usage{
coo_perim(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
\code{numeric}, the perimeter.
}
\description{
Calculates the perimeter
}
\examples{
data(bot)
coo_perim(bot[1])
hist(sapply(bot$coo, coo_perim), breaks=10)
}

