% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_rectangularity}
\alias{coo_rectangularity}
\title{Calculates the rectangularity of a shape}
\source{
Rosin PL. 2005. Computing global shape measures.
Handbook of Pattern Recognition and Computer Vision. 177-196.
}
\usage{
coo_rectangularity(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates.}
}
\value{
numeric, the rectangularity.
}
\description{
Calculates the rectangularity of a shape
}
\examples{
data(bot)
coo_rectangularity(bot[1])
}
\keyword{ShapeDescriptors}

