% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_rotate}
\alias{coo_rotate}
\title{Rotates coordinates}
\usage{
coo_rotate(coo, theta = 0)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or any \link{Coo} object.}

\item{theta}{\code{numeric}the angle (in radians) to rotate shapes.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Rotates the coordinates by a 'theta' angle (in radians) in
the trigonometric direction (anti-clockwise). If not provided,
assumed to be the centroid size. It involves three steps: centering from current position,
dividing coordinates by 'scale', translating to the original position.
}
\examples{
coo_plot(bot[1])
coo_plot(coo_rotate(bot[1], pi/2))
# on Coo
stack(bot)
stack(coo_rotate(bot, pi/2))
}
\seealso{
Other rotation functions: \code{\link{coo_rotatecenter}}
}

