% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_sample}
\alias{coo_sample}
\title{Sample coordinates (among points)}
\usage{
coo_sample(coo, n)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates or an \link{Out} or an \link{Opn} object.}

\item{n}{\code{integer}, the number fo points to sample.}
}
\value{
a \code{matrix} of (x; y) coordinates, or an \link{Out} or an \link{Opn} object.
}
\description{
Sample n coordinates among existing points.
}
\details{
For the \link{Out} an \link{Opn}
methods (pointless for \link{Ldk}), in an \code{$ldk} component is defined,
it is changed accordingly by multiplying the ids by n over the number of coordinates.
}
\examples{
b <- bot[1]
stack(bot)
stack(coo_sample(bot, 24))
coo_plot(b)
coo_plot(coo_sample(b, 24))
}

