% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_samplerr}
\alias{coo_samplerr}
\title{Samples coordinates (regular radius)}
\usage{
coo_samplerr(coo, n)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{n}{\code{integer}, the number of points to sample.}
}
\value{
a \code{matrix} of (x; y) coordinates or a Coo object.
}
\description{
Samples n coordinates with a regular angle.
}
\examples{
stack(bot)
bot <- coo_center(bot)
stack(coo_samplerr(bot, 12))
coo_plot(bot[1])
coo_plot(rr <- coo_samplerr(bot[1], 12))
cpos <- coo_centpos(bot[1])
segments(cpos[1], cpos[2], rr[, 1], rr[, 2])
}
\seealso{
Other sampling functions: \code{\link{coo_extract}},
  \code{\link{coo_interpolate}}, \code{\link{coo_sample}}
}

