% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_samplerr}
\alias{coo_samplerr}
\title{Samples coordinates (regular radius)}
\usage{
coo_samplerr(coo, n)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{n}{integer, the number of points to sample.}
}
\value{
a \code{matrix} of (x; y) coordinates or an Coo object.
}
\description{
Samples n coordinates with a regular angle.
}
\examples{
data(bot)
stack(bot)
bot <- coo_center(bot)
stack(coo_samplerr(bot, 12))
coo_plot(bot[1])
coo_plot(rr <- coo_samplerr(bot[1], 12))
cpos <- coo_centpos(bot[1])
segments(cpos[1], cpos[2], rr[, 1], rr[, 2])
}
\keyword{ShapeUtilities}

