% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_scale}
\alias{coo_scale}
\title{Scales coordinates}
\usage{
coo_scale(coo, scale)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{scale}{the scaling factor, by default, the centroid size.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Scales the coordinates by a 'scale' factor. If not provided,
assumed to be the centroid size. It involves three steps: centering from current position,
dividing coordinates by 'scale', pusing back to the original position.
}
\examples{
coo_plot(bot[1])
coo_plot(coo_scale(bot[1]))
# on Coo objects
stack(coo_center(bot))
stack(coo_scale(coo_center(bot)))
}
\seealso{
Other scaling functions: \code{\link{coo_scalex}},
  \code{\link{coo_template}}
}

