% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_scale}
\alias{coo_scale}
\title{Scales coordinates}
\usage{
coo_scale(coo, scale)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{scale}{the scaling factor, by default, the centroid size.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Scales the coordinates by a 'scale' factor. If not provided,
assumed to be the centroid size. It involves three steps: centering from current position,
dividing coordinates by 'scale', pusing back to the original position.
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo_scale(bot))
coo_plot(b)
coo_plot(coo_scale(b))
}
\keyword{ShapeUtilities}

