% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_shearx}
\alias{coo_shearx}
\alias{coo_sheary}
\title{Shears shapes}
\usage{
coo_shearx(coo, k)

coo_sheary(coo, k)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{k}{\code{numeric} shear factor}
}
\value{
a \code{matrix} of (x; y) coordinates.
}
\description{
\code{coo_shearx} applies a shear mapping on a matrix of (x; y) coordinates (or a list), parallel
to the x-axis (i.e. x' = x + ky; y' = y + kx). \code{coo_sheary} does it parallel to the y-axis.
}
\examples{
coo <- coo_template(shapes[11])
coo_plot(coo)
coo_draw(coo_shearx(coo, 0.5), border="blue")
coo_draw(coo_sheary(coo, 0.5), border="red")

}

