% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slice}
\alias{coo_slice}
\title{Slices shapes between successive coordinates}
\usage{
coo_slice(coo, ids)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{ids}{\code{numeric} of length >= 2, where to slice the shape(s)}
}
\value{
a list of shapes or a list of \link{Opn}
}
\description{
Takes a shape with n coordinates. When you pass this function with at least
two ids (<= n), the shape will be open on the corresponding coordinates and
slices returned as a list
}
\examples{
# single shape, a list of matrices is returned
sh <- coo_slice(hearts[1], c(12, 24, 36, 48))
coo_plot(sh[[1]])
panel(Opn(sh))
# on a Coo, a list of Opn is returned
# makes no sense if shapes are not normalized first
sh2 <- coo_slice(hearts, c(12, 24, 36, 48))
panel(sh2[[1]])
}

