% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slidedirection}
\alias{coo_slidedirection}
\title{Slides coordinates in a particular direction}
\usage{
coo_slidedirection(coo, direction, center, id)
}
\arguments{
\item{coo}{a matrix (or a list) or (x; y) coordinates or a Coo object}

\item{direction}{a character among \code{'N'} (by default), \code{'S'}, \code{'E'}, or \code{'W'}.}

\item{center}{logical whether to center or not before sliding}

\item{id}{whether to return the id of the point or the slided shapes}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Shapes are centered and then, according to direction, the point northwards, southwards,
eastwards or westwards the centroid, becomes the first point with \link{coo_slide}.
}
\examples{
data(bot)
b <- coo_rotate(bot[1], pi/6) # dummy example just to make it obvious
coo_plot(b) # not the first point
coo_plot(coo_slidedirection(b, 'N'))
coo_plot(coo_slidedirection(b, 'E'))
coo_plot(coo_slidedirection(b, 'W'))
coo_plot(coo_slidedirection(b, 'S'))

# on Coo objects
stack(bot)
stack(coo_slidedirection(bot, 'E'))
}
\keyword{ShapeUtilities}

