% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_slidegap}
\alias{coo_slidegap}
\title{Slides coordinates using the widest gap}
\usage{
coo_slidegap(coo, force)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or a \code{list}, or any \link{Coo} object.}

\item{force}{\code{logical} whether to use the widest gap, with no check, as the real gap}
}
\value{
a \code{matrix} of (x; y) coordinates or a \link{Coo} object.
}
\description{
When slicing a shape using two landmarks, or functions such as \link{coo_up},
an open curve is obtained and the rank of points make wrong/artefactual results.
If the widest gap is > 5 * median of other gaps, then the couple of coordinates
forming this widest gap is used as starting and ending points. This switch helps
to deal with open curves. Examples are self-speaking.
Use \code{force=TRUE} to bypass this check
}
\examples{
cat <- coo_center(shapes[4])
coo_plot(cat)

# we only retain the bottom of the cat
cat_down <- coo_down(cat, slidegap=FALSE)

# see? the segment on the x-axis coorespond to the widest gap.
coo_plot(cat_down)

# that's what we meant
coo_plot(coo_slidegap(cat_down))
}
\seealso{
Other sliding functions: \code{\link{coo_slidedirection}},
  \code{\link{coo_slide}}
}

