% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coo_template}
\alias{coo_template}
\title{'Templates' shapes}
\usage{
coo_template(coo, size)
}
\arguments{
\item{coo}{A \code{list} or a \code{matrix} of coordinates.}

\item{size}{\code{numeric}. Indicates the length of the side 'inscribing'
the shape.}
}
\value{
Returns a matrix of \code{(x; y)}coordinates.
}
\description{
\code{coo_template} returns shape centered on the origin and inscribed in a \code{size}-side square
}
\details{
See \link{coo_listpanel} for an illustration of this function. The morphospaces
functions also take profit of this function. May be useful to develop other graphical functions.
}
\examples{
data(bot)
coo <- bot[1]
coo_plot(coo_template(coo), xlim=c(-1, 1), ylim=c(-1, 1))
rect(-0.5, -0.5, 0.5, 0.5)

s <- 0.01
coo_plot(coo_template(coo, s))
rect(-s/2, -s/2, s/2, s/2)
}
\seealso{
\link{coo_listpanel}.
}
\keyword{Graphics}

