% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{coo_trans}
\alias{coo_trans}
\title{Translates coordinates}
\usage{
coo_trans(coo, x = 0, y = 0)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}

\item{x}{\code{numeric}translation along the x-axis.}

\item{y}{\code{numeric}translation along the y-axis.}
}
\value{
a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.
}
\description{
Translates the coordinatesby a 'x' and 'y' value
}
\examples{
data(bot)
b <- bot[1]
stack(bot)
stack(coo_trans(bot, 50, 100))
coo_plot(b)
coo_plot(coo_trans(b, 50, 100))
}
\keyword{ShapeUtilities}

