% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo-shapedescriptors.R
\name{coo_width}
\alias{coo_width}
\title{Calculates the width of a shape}
\usage{
coo_width(coo)
}
\arguments{
\item{coo}{a \code{matrix} of (x; y) coordinates or Coo object}
}
\value{
the width (in pixels) of the shape
}
\description{
Nothing more than \code{coo_lw(coo)[2]}.
}
\examples{
coo_width(bot[1])
}
\seealso{
\link{coo_lw}, \link{coo_length}.

Other coo_ descriptors: \code{\link{coo_angle_edges}},
  \code{\link{coo_angle_tangent}}, \code{\link{coo_area}},
  \code{\link{coo_boundingbox}}, \code{\link{coo_chull}},
  \code{\link{coo_circularity}},
  \code{\link{coo_convexity}},
  \code{\link{coo_eccentricity}},
  \code{\link{coo_elongation}}, \code{\link{coo_length}},
  \code{\link{coo_lw}}, \code{\link{coo_rectangularity}},
  \code{\link{coo_rectilinearity}},
  \code{\link{coo_scalars}}, \code{\link{coo_solidity}},
  \code{\link{coo_tac}}
}
