% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{cpx2coo}
\alias{cpx2coo}
\title{Converts complex to cartesian coordinates}
\usage{
cpx2coo(Z)
}
\arguments{
\item{Z}{coordinates expressed in the complex form}
}
\value{
coordinates expressed in the cartesian form
}
\description{
Converts complex to cartesian coordinates
}
\examples{
shapes[4] \%>\% coo_sample(24) \%>\% coo2cpx()
shapes[4] \%>\% coo_sample(24) \%>\% coo2cpx() \%>\% cpx2coo()
}

