% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Momocs-Datasets.R
\docType{data}
\name{bot}
\alias{bot}
\title{Data: Outline coordinates of beer and whisky bottles.}
\format{A \link{Out} object containing the outlines coordinates and a grouping factor
for 20 beer and 20 whisky bottles}
\source{
Images have been grabbed on the internet and prepared by the package's
authors. No particular choice has been made on the dimension of the original
images or the brands cited here.
}
\description{
Data: Outline coordinates of beer and whisky bottles.
}
\keyword{Datasets}

