% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{flower}
\alias{flower}
\title{Data: Measurement of iris flowers}
\format{A \code{TraCoe} object with 150 measurements of 4 variables
(petal + sepal) x (length x width) on 3 species of iris. This dataset is
the classical \link{iris} formatted for Momocs.}
\source{
see link{iris}
}
\description{
Data: Measurement of iris flowers
}
\seealso{
Other datasets: \code{\link{bot}}, \code{\link{chaff}},
  \code{\link{hearts}}, \code{\link{molars}},
  \code{\link{mosquito}}, \code{\link{oak}},
  \code{\link{olea}}, \code{\link{shapes}},
  \code{\link{trilo}}, \code{\link{wings}}
}

