% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{hearts}
\alias{hearts}
\title{Data: Outline coordinates of hand-drawn hearts}
\format{
A \link{Out} object with the outline coordinates of 240 hand-drawn hearts
by 8 different persons, with 4 landmarks.
}
\source{
We thank the fellows of the Ecology Department of the French Institute
of Pondicherry that drawn the hearts, that then have been smoothed, scaled, centered,
and downsampled to 80 coordinates per outline.
}
\description{
Data: Outline coordinates of hand-drawn hearts
}
\seealso{
Other datasets: 
\code{\link{apodemus}},
\code{\link{bot}},
\code{\link{chaff}},
\code{\link{charring}},
\code{\link{flower}},
\code{\link{molars}},
\code{\link{mosquito}},
\code{\link{mouse}},
\code{\link{nsfishes}},
\code{\link{oak}},
\code{\link{olea}},
\code{\link{shapes}},
\code{\link{trilo}},
\code{\link{wings}}
}
\concept{datasets}
