% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{molars}
\alias{molars}
\title{Data: Outline coordinates of 360 molars}
\format{A \link{Out} object containing 79 equilinearly spaced (x; y)
coordinates for 360 crown outlines, of modern human molars,
along with their type (\code{$type}) - 90 first upper molars (UM1), 90 second upper molars (UM2),
90 first lower molars (LM1), 90 second lower molars (LM2) -
and the individual (\code{ind}) they come from (the data of the 360 molars are taken
from 180 individuals).}
\source{
Corny, J., & Detroit, F. (2014). Technical Note: Anatomic
identification of isolated modern human molars: testing Procrustes aligned
outlines as a standardization procedure for elliptic fourier analysis.
\emph{American Journal of Physical Anthropology}, 153(2), 314-22. doi:10.1002/ajpa.22428
\url{http://onlinelibrary.wiley.com/doi/10.1002/ajpa.22428/abstract}
}
\description{
Courtesy of Julien Corny and Florent Detroit.
}
\seealso{
Other datasets: \code{\link{bot}}, \code{\link{chaff}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{mosquito}}, \code{\link{oak}},
  \code{\link{olea}}, \code{\link{shapes}},
  \code{\link{trilo}}, \code{\link{wings}}
}

