% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-datasets.R
\docType{data}
\name{oak}
\alias{oak}
\title{Data: Configuration of landmarks of oak leaves}
\format{A \link{Ldk} object containing 11 (x; y) landmarks  from 176 oak leaves wings, from}
\source{
Viscosi, V., & Cardini, A. (2011). Leaf morphology, taxonomy and
geometric morphometrics: a simplified protocol for beginners.
PloS One, 6(10), e25630. doi:10.1371/journal.pone.0025630
}
\description{
From Viscosi and Cardini (2001).
}
\keyword{Datasets}

