% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Momocs-Datasets.R
\docType{data}
\name{olea}
\alias{olea}
\title{Data: Outline coordinates of olive seeds open outlines.}
\format{An \link{Opn} object with the outline coordinates of olive seeds.}
\source{
We thank Jean-Frederic Terral and Sarah Ivorra (UMR CBAE, Montpellier, France)
from allowing us to share the data.

You can have a look to the original paper:
Terral J-F, Alonso N, Capdevila RB i, Chatti N, Fabre L, Fiorentino G,
Marinval P, Jorda GP, Pradat B, Rovira N, et al. 2004.
Historical biogeography of olive domestication (\emph{Olea europaea} L.)
as revealed by geometrical morphometry applied to biological and
archaeological material. \emph{Journal of Biogeography} \bold{31}: 63-77.
}
\description{
Data: Outline coordinates of olive seeds open outlines.
}
\keyword{Datasets}

