% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs-datasets.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Data: Outline coordinates of various shapes}
\format{An \link{Out} object with the outline coordinates of some various shapes.}
\source{
Borrowed from default shapes from (c) Adobe Photoshop.
}
\description{
Data: Outline coordinates of various shapes
}
\keyword{Datasets}

