% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Data: Outline coordinates of various shapes}
\format{An \link{Out} object with the outline coordinates of some various shapes.}
\source{
Borrowed default shapes from (c) Adobe Photoshop. Do not send me to jail.
}
\description{
Data: Outline coordinates of various shapes
}
\seealso{
Other datasets: \code{\link{apodemus}}, \code{\link{bot}},
  \code{\link{chaff}}, \code{\link{charring}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{molars}}, \code{\link{mosquito}},
  \code{\link{mouse}}, \code{\link{oak}},
  \code{\link{olea}}, \code{\link{trilo}},
  \code{\link{wings}}
}

