% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{shapes}
\alias{shapes}
\title{Data: Outline coordinates of various shapes}
\format{An \link{Out} object with the outline coordinates of some various shapes.}
\source{
Borrowed from default shapes from (c) Adobe Photoshop.
}
\description{
Data: Outline coordinates of various shapes
}
\keyword{Datasets}

