% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{trilo}
\alias{trilo}
\title{Data: Outline coordinates of cephalic outlines of trilobite}
\format{A \link{Out} object 64 coordinates of 50 cephalic outlines from different
ontogenetic stages of trilobite.}
\source{
Arranged from: \url{http://folk.uio.no} (used to be in \code{ohammer} website but
seems to be deprecated now).
The original data included 51 outlines and 5 ontogenetic stages,
but one of them has just a single outline thas has been removed.
}
\description{
Data: Outline coordinates of cephalic outlines of trilobite
}
\seealso{
Other datasets: 
\code{\link{apodemus}},
\code{\link{bot}},
\code{\link{chaff}},
\code{\link{charring}},
\code{\link{flower}},
\code{\link{hearts}},
\code{\link{molars}},
\code{\link{mosquito}},
\code{\link{mouse}},
\code{\link{nsfishes}},
\code{\link{oak}},
\code{\link{olea}},
\code{\link{shapes}},
\code{\link{wings}}
}
\concept{datasets}
