% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-datasets.R
\docType{data}
\name{wings}
\alias{wings}
\title{Data: Landmarks coordinates of mosquito wings}
\format{A \link{Ldk} object containing 18 (x; y) landmarks  from 127 mosquito wings, from}
\source{
Rohlf and Slice 1990 and \url{http://life.bio.sunysb.edu/morph/data/RohlfSlice1990Mosq.nts}
}
\description{
Data: Landmarks coordinates of mosquito wings
}
\seealso{
Other datasets: \code{\link{bot}}, \code{\link{chaff}},
  \code{\link{flower}}, \code{\link{hearts}},
  \code{\link{molars}}, \code{\link{mosquito}},
  \code{\link{oak}}, \code{\link{olea}},
  \code{\link{shapes}}, \code{\link{trilo}}
}

