% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{wings}
\alias{wings}
\title{Data: Outline coordinates of 20 beer and 20 whisky bottles.}
\format{A \link{Ldk} object containing 18 (x; y) landmarks  from 127 mosquito wings, from}
\source{
Rohlf and Slice 1990 and \url{http://life.bio.sunysb.edu/morph/data/RohlfSlice1990Mosq.nts}
}
\description{
Data: Outline coordinates of 20 beer and 20 whisky bottles.
}
\keyword{Datasets}

