% Generated by roxygen2 (4.0.0): do not edit by hand
\name{defLandmarks}
\alias{defLandmarks}
\title{Define landmarks on Out and Opn objects}
\usage{
defLandmarks(Coo, nb.ldk)
}
\arguments{
\item{Coo}{an Out or Opn object}

\item{nb.ldk}{the number of landmarks to define on every shape}
}
\value{
an Out or an Opn object with some landmarks defined
}
\description{
Helps to define landmarks on a \code{Coo} object.
The number of landmarks must be specified and rows indices that
correspond to the nearest points clicked on every outlines are
stored in the \code{$ldk} slot of the \code{Coo} object.
}
\examples{
\dontrun{
data(bot)
bot <- bot[1:5] # to make it shorter to try
# click on 3 points, 5 times.
# Don't forget to save the object returned by defLandmarks...
bot2 <- defLandmarks(bot, 3)
stack(bot2)
bot2$ldk
}
}
\keyword{Ldk}
\keyword{Opn}
\keyword{Out}

