% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-Misc.R
\name{ed}
\alias{ed}
\title{Calculates euclidean distance between two points.}
\usage{
ed(pt1, pt2)
}
\arguments{
\item{pt1}{(x; y) coordinates of the first point.}

\item{pt2}{(x; y) coordinates of the second point.}
}
\value{
Returns the euclidean distance between the two points.
}
\description{
\code{ed} simply calculates euclidean distance between two points defined by
their (x; y) coordinates.
}
\examples{
ed(c(0,1), c(1,0))
}
\seealso{
\link{edm}, \link{edm_nearest}, \link{dist}.
}
\keyword{Miscellaneous}

