% Generated by roxygen2 (4.0.2): do not edit by hand
\name{efourier}
\alias{efourier}
\alias{efourier.Out}
\alias{efourier.default}
\title{Elliptical Fourier transform}
\usage{
efourier(x, ...)

\method{efourier}{default}(x, nb.h, smooth.it = 0, verbose = TRUE, ...)

\method{efourier}{Out}(x, nb.h, smooth.it = 0, norm = TRUE, start = FALSE,
  ...)
}
\arguments{
\item{x}{A \code{list} or a \code{matrix} of coordinates or a \code{Out} object}

\item{...}{useless here}

\item{nb.h}{\code{integer}. The number of harmonics to use. If missing 99pc harmonic power is used.}

\item{smooth.it}{\code{integer}. The number of smoothing iterations to
perform.}

\item{verbose}{\code{logical}. Whether to print or not diagnosis messages.}

\item{norm}{whether to normalize the coefficients using \link{efourier_norm}}

\item{start}{logical whether to consider the first point as homologous}
}
\value{
A list with these components: \item{an }{\code{vector} of
\eqn{a_{1->n}} harmonic coefficients.} \item{bn }{\code{vector} of
\eqn{b_{1->n}} harmonic coefficients.} \item{cn }{\code{vector} of
\eqn{c_{1->n}} harmonic coefficients.} \item{dn }{\code{vector} of
\eqn{d_{1->n}} harmonic coefficients.} \item{ao }{\code{ao} Harmonic
coefficient.} \item{co }{\code{co} Harmonic coefficient.}
}
\description{
\code{efourier} computes Elliptical Fourier Analysis (or Transforms or EFT)
from a matrix (or a list) of (x; y) coordinates.
}
\details{
These functions and their mathematical background detailed below are here
detailed to ease their use in new methods but are used internally by methods
on \code{Out}-objects.

Elliptic Fourier analysis and normalization are calculated as follows. Let
\eqn{T} be the perimeter of a given closed outline, here considered as the
period of the signal. One sets \eqn{\omega = 2\pi/T} to be the pulse. Then,
the curvilinear abscissa, \eqn{t} varies from \eqn{0} to \eqn{T}. One can
express \eqn{x(t)} and \eqn{y(t)} as: \deqn{ x(t) =
\frac{a_0}{2}+\sum\limits_{n=1}^{+\infty} a_n\cos n\omega t + b_n\sin
n\omega t } with \deqn{ a_n = \frac{2}{T}+ \int\limits_{0}^{T} x(t)\cos
(n\omega t) \mathrm{d} t } \deqn{ b_n = \frac{2}{T}+ \int\limits_{0}^{T}
x(t)\sin (n\omega t) \mathrm{d} t }

similarly, \deqn{ y(t) = \frac{c_0}{2}+\sum\limits_{n=1}^{+\infty} c_n\cos
n\omega t + d_n\sin n\omega t } with \deqn{ c_n = \frac{2}{T}+
\int\limits_{0}^{T} y(t)\cos (n\omega t) \mathrm{d} t } \deqn{d_n =
\frac{2}{T}+ \int\limits_{0}^{T} y(t)\sin (n\omega t) \mathrm{d} t }

Since the outline contains a \eqn{k} finite number of points, one can
therefore calculate discrete estimators for every harmonic coefficient of
the \eqn{n^{th}} harmonics: \deqn{
a_n=\frac{T}{2\pi^2n^2}\sum\limits_{p=1}^k \frac{\Delta x_p}{\Delta
t_p}(\cos\frac{2\pi nt_p}{T}-\cos\frac{2\pi nt_{p-1}}{T}) } \deqn{
b_n=\frac{T}{2\pi^2n^2}\sum\limits_{p=1}^k \frac{\Delta x_p}{\Delta
t_p}(\sin\frac{2\pi nt_p}{T}-\sin\frac{2\pi nt_{p-1}}{T}) }

\eqn{\Delta x_1=x_1-x_k} and \eqn{c_n} and \eqn{d_n} are calculated
similarly. \eqn{a_0} and \eqn{c_0} correspond to the estimate of the
coordinates of the centroid of original outline and are estimated by: \deqn{
a_0=\frac{2}{T}\sum\limits_{i=1}^p x_i } and \deqn{
c_0=\frac{2}{T}\sum\limits_{i=1}^p y_i }

Intuitively, for all positive integers \eqn{n}, the sum of a cosine curve
and a sine curve represent the \eqn{n^{th}} harmonic content of the \eqn{x}
and \eqn{y} projections of the \eqn{k}-edged polygon, and for any \eqn{n},
these two curves define an ellipse in the plane. Ferson and colleagues
noticed that in the 'time' it takes the \eqn{n^{th}} harmonic to traverse
its ellipse \eqn{n} times, the \eqn{(n+1)^{th}} harmonic has traversed its
own ellipse \eqn{n+1} times. The reconstruction of the original polygon is
done by vector adding these ellipses for all harmonics, which echoes
astronomical Ptolemy's epicycles (see \link{Ptolemy}), and the
reconstruction obtained from \eqn{N} harmonics is the best possible fit in a
least-squares sense.

Normalization of coefficients has long been a matter of trouble,
and not only for newcomers. There are two ways of normalizing outlines: the first,
and by far the msot used, is to use a "numerical" alignment, directly on the
matrix of coefficients. The coefficients of the first harmonic are consumed
by this process but harmonics of higher rank are normalized in terms of size
and rotation. This is sometimes referred as using the "first ellipse", as the
harmonics define an ellipse in the plane, and the first one is the mother of all
ellipses, on which all others "roll" along. This approach is really convenient
as it is done easily by most software (if not the only option) and by Momocs too.
It is the default option of \code{efourier}.

But here is the pitfall: if your shapes are prone to bad aligments among all
the first ellipses, this will result in poorly (or even not at all) "homologous" coefficients.
The shapes prone to this are either (at least roughly) circular and/or with a strong
bilateral symmetry. You can try to use \code{\link{stack}} on the \code{\link{Coe}} object
 returned by \code{efourier}. Also, when plotting PCA using Momocs,
this will be strikingly clear though. This phenomenon will result in two clusters,
and more strikingly into upside-down (or 180 degrees rotated)
shapes on the morphospace. If this happen, you should seriously consider
aligning your shapes \emph{before} the \code{efourier} step,
and performing the latter with no normalization (\code{norm = FALSE}), since
it has been done before.

You have several options to align your shapes, using control points (or landmarks),
of Procrustes alignment (see \code{\link{fgProcrustes}}) through their calliper
length (see \code{\link{coo_aligncalliper}}), etc. You should also make the first
point homologous either with \code{\link{coo_slide}} or \code{\link{coo_slidedirection}}
to minimize any subsequent problems.

I will dedicate one vignette to this problem
asap (fall 2014). In the meantime, contact me should you think we could
solve this with two brains.
}
\note{
Directly borrowed for Claude (2008), and also called \code{efourier} there.
}
\examples{
data(bot)
coo <- bot[1]
coo_plot(coo)
ef <- efourier(coo, 12)
ef
efi <- efourier_i(ef)
coo_draw(efi, border='red', col=NA)
}
\references{
Claude, J. (2008) \emph{Morphometrics with R}, Use R! series,
Springer 316 pp.
Ferson S, Rohlf FJ, Koehn RK. 1985. Measuring shape variation of
two-dimensional outlines. \emph{Systematic Biology} \bold{34}: 59-68.
}
\seealso{
\link{efourier_i} for the reverse operation. \link{Ptolemy} for an implementation of the
Ptolemaic ellipses graph sometimes used to illustrate this approach.
\link{rfourier}, \link{tfourier} for the other members
of the Fourier's family.
}
\keyword{efourier}

