% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-ldk-procrustes.R
\name{fgsProcrustes}
\alias{fgsProcrustes}
\title{Full Generalized Procrustes alignment between shapes with sliding landmarks}
\source{
See \code{?gpagen} in \code{geomorph} package
}
\usage{
fgsProcrustes(x)
}
\arguments{
\item{x}{Ldk object with some \code{$slidings}}
}
\value{
a list
}
\description{
Directly wrapped around \code{geomorph::gpagen}.
}
\note{
Landmarks methods are the less tested in Momocs. Keep in mind that some features
are still experimental and that your help is welcome.
}
\examples{
ch <- chaff \%>\% slice(1:5) # for the sake of speed
chaffp <- fgsProcrustes(ch)
chaffp
chaffp \%>\% PCA() \%>\% plot("taxa")
}
\seealso{
Other procrustes functions: 
\code{\link{fProcrustes}()},
\code{\link{fgProcrustes}()},
\code{\link{pProcrustes}()}
}
\concept{procrustes functions}
