% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fProcrustes}
\alias{fProcrustes}
\title{Full Procrustes alignment between two shapes}
\usage{
fProcrustes(coo1, coo2)
}
\arguments{
\item{coo1}{Configuration matrix to be superimposed onto the centered preshape of coo2.}

\item{coo2}{Reference configuration matrix.}
}
\value{
a list with components
\enumerate{
\item coo1: Superimposed centered preshape of coo1 onto the centered preshape of coo2
\item coo2: Centered preshape of coo2
\item rotation: Rotation matrix
\item scale: Scale parameter
\item DF: Full Procrustes distance between coo1 and coo2.
}
}
\description{
Directly borrowed from Claude 2008 (fPsup)
}
\references{
Claude, J. (2008). Morphometrics with R. Analysis (p. 316). Springer.
}

