% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-def-Out.R
\name{get_ldk}
\alias{get_ldk}
\title{Retrieve landmarks coordinates from Opn and Out objects}
\usage{
get_ldk(Coo)
}
\arguments{
\item{Coo}{a Coo object, either Out or Opn}
}
\value{
an array of coordinates X (x; y) coordinates X number of shapes.
}
\description{
In \link{Out} and \link{Opn} classes, landmarks (if any) are stored as
row indices. This methods allows to retrieve the corresponding (x; y) coordinates.
}
\examples{
data(hearts)
ldk.h <- get_ldk(hearts)
stack(Ldk(a2l(ldk.h)))
ldk.h
}
\seealso{
\link{def_ldk}, \link{fgProcrustes}
}

