% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coe.R
\name{hist.OutCoe}
\alias{hist.Coe}
\alias{hist.OutCoe}
\title{Histogram of morphometric coefficients}
\usage{
\method{hist}{OutCoe}(x, retain = 4, drop = 0, bw = 20, ...)
}
\arguments{
\item{x}{the \link{Coe} object}

\item{retain}{numeric the number of harmonics to retain}

\item{drop}{numeric the number of harmonics to drop}

\item{bw}{the number of bins (range/bw) to display}

\item{...}{useless here but maintain the consistency with generic hist}
}
\value{
a ggplot2 object
}
\description{
Explores the distribution of coefficient values.
}
\examples{
data(bot)
bot.f <- efourier(bot, 24)
hist(bot.f)

data(olea)
op <- opoly(olea)
hist(op)
}
\seealso{
\link{boxplot.Coe}
}

