% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hpow}
\alias{hpow}
\title{Quantitative calibration, through harmonic power, for Out objects}
\usage{
hpow(Out, method = "efourier", id = 1:length(Out), nb.h = 16, drop = 1,
  smooth.it = 0, plot = TRUE,
  title = "Fourier coefficients power spectrum", lineat.y = c(0.9, 0.95,
  0.99, 0.999), bw = 0.1)
}
\arguments{
\item{Out}{the \code{Out} object on which to hpow}

\item{method}{any method from \code{c("efourier", "rfourier", "tfourier")}}

\item{id}{the shape on which to perform hpow. All by default}

\item{nb.h}{numeric the maximum number of harmonic}

\item{drop}{numeric the number of harmonics to drop for the cumulative sum}

\item{smooth.it}{numeric, number of smoothing iterations}

\item{plot}{logical whether to plot the result or simply return the matrix}

\item{title}{a title for the plot}

\item{lineat.y}{vector of numeric for drawing horizontal lines}

\item{bw}{width of horizontal segments drawn for each harmonic.}
}
\value{
a matrix containing cumulated harmonic power for each harmonic.
}
\description{
Estimates the number of harmonics required for the three Fourier methods
implemented so far in Momocs: elliptical Fourier analysis
(see \link{efourier}), radii variation analysis (see \link{tfourier})
and tangent angle analysis (see \link{tfourier}).
It returns and can plot cumulated harmonic power whether dropping
the first harmonic or not.
}
\details{
The power of a given harmonic \eqn{n} is calculated as follows for
elliptical Fourier analysis and the n-th harmonic:
\deqn{HarmonicPower_n \frac{A^2_n+B^2_n+C^2_n+D^2_n}{2}}
and as follows for radii variation and tangent angle:
\deqn{
HarmonicPower_n= \frac{A^2_n+B^2_n+C^2_n+D^2_n}{2}
}
}
\examples{
data(bot)
hpow(bot)
}
\keyword{Out}

