% Generated by roxygen2 (4.0.0): do not edit by hand
\name{img.plot}
\alias{img.plot}
\title{Extract structure from filenames}
\usage{
img.plot(img)
}
\arguments{
\item{img}{a matrix of an image, such as those obtained with \link{readJPEG}.}
}
\description{
A very simple (and fast) image plotter.
}
\keyword{Import}

