% Generated by roxygen2 (4.0.0): do not edit by hand
\name{img.plot}
\alias{img.plot}
\alias{img.plot0}
\title{Plots a .jpg image}
\usage{
img.plot(img)

img.plot0(img)
}
\arguments{
\item{img}{a matrix of an image, such as those obtained with \link{readJPEG}.}
}
\description{
A very simple image plotter. If provided with a path,
read the .jpg and plots it. If not provided with an imagematrix, will
ask you to choose interactively a \code{.jpeg} image.
}
\details{
\code{img.plot} is used in import functions such as \link{import.jpg1};
\code{img.plot0} does the same job but preserves the \code{par} and plots axes.
}
\keyword{Import}

