% Generated by roxygen2 (4.0.0): do not edit by hand
\name{import.jpg}
\alias{import.jpg}
\title{Extract outline coordinates from multiple .jpg files}
\usage{
import.jpg(jpg.paths = NULL, auto.notcentered = TRUE,
  fun.notcentered = NULL, threshold = 0.5, verbose = TRUE)
}
\arguments{
\item{jpg.paths}{a vector of paths corresponding to the .jpg files to import. If not
provided (or \code{NULL}), switches to the automatic version. See Details below.}

\item{auto.notcentered}{logical if TRUE random locations will be used until.
one of them is (assumed) to be within the shape (because of a black pixel);
if FALSE a \link{locator} will be called, and you will have to click on a
point within the shape.}

\item{fun.notcentered}{NULL by default. Is your shapes are not centered and if a random pick of
a black pixel is not satisfactory. See \link{import.jpg1} help and examples.}

\item{threshold}{the threshold value use to binarize the images. Above, pixels
are turned to 1, below to 0.}

\item{verbose}{whether to print which file is being treated. Useful to detect problems.}
}
\value{
a list of matrices of (x; y) coordinates that can be passed to \link{Out}
}
\description{
This function is used to import outline coordinates and is built around
\link{import.jpg1}.
}
\details{
see \link{import.jpg1} for important informations about how the outlines are extracted,
and \link{import.Conte} for the algorithm itself.

If \code{jpg.paths} is not provided (or \code{NULL}), you will have to select any \code{.jpg}
file in the folder taht contains all your files. All the outlines should be imported then.
}
\examples{
\dontrun{

# if your images are in the folder "/foo/jpgs/"
lf <- list.files("/foo/jpegs", full.names=TRUE)
coo <- import.jpg(lf)
Out(coo)

# "automatic" version
coo <- import.jpg()
}
}
\seealso{
\link{import.jpg1}, \link{import.Conte}, \link{import.txt}, \link{lf.structure}.
See also Momocs' vignettes for data import.
}
\keyword{Import}

