% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{import_tps}
\alias{import_tps}
\title{Imports a tps file}
\usage{
import_tps(tps.path, curves = TRUE)
}
\arguments{
\item{tps.path}{lines, typically from \link{readLines}, describing a single shape in tps-like format}

\item{curves}{\code{logical} whether to read curves, if any}
}
\value{
a list with components:
\code{coo} a matrix of coordinates; \code{cur} a list of matrices; \code{scale} the scale as a numeric.
}
\description{
And returns a list of coordinates, curves, scale
}
\seealso{
Other babel functions: \code{\link{chc2Out}},
  \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{nef2Coe}},
  \code{\link{ntsrow2Coo}}, \code{\link{pix2chc}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

