% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.closed}
\alias{is.closed}
\title{Tests if shapes are closed}
\usage{
is.closed(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a boolean.
}
\description{
Returns TRUE/FALSE whether the last coordinate of the shapes is the same
as the first one.
}
\examples{
is.closed(matrix(1:10, ncol=2))
is.closed(coo.close(matrix(1:10, ncol=2)))
data(bot)
is.closed(bot)
is.closed(coo.close(bot))
}
\seealso{
\link{coo.close}, \link{coo.unclose}
}
\keyword{ShapeUtilities}

