% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-testers.R
\name{is.Coo}
\alias{is.Coe}
\alias{is.Coo}
\alias{is.LDA}
\alias{is.Ldk}
\alias{is.LdkCoe}
\alias{is.Opn}
\alias{is.OpnCoe}
\alias{is.Out}
\alias{is.OutCoe}
\alias{is.PCA}
\title{Tests if an object is og a given class}
\usage{
is.Coo(x)

is.PCA(x)

is.LDA(x)

is.Out(x)

is.Opn(x)

is.Ldk(x)

is.Coe(x)

is.OutCoe(x)

is.OpnCoe(x)

is.LdkCoe(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE/FALSE
}
\description{
Tests if any of the classes of an object is of a given class. For instance
is.PCA on a PCA object (both 'PCA' and 'prcomp') will return TRUE
}
\examples{
data(bot)
is.Coo(bot)
is.Out(bot)
is.Ldk(bot)
}

