% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coo-utilities.R
\name{is_closed}
\alias{is_closed}
\title{Tests if shapes are closed}
\usage{
is_closed(coo)
}
\arguments{
\item{coo}{either a \code{matrix} of (x; y) coordinates, or a \link{Coo} object.}
}
\value{
a boolean.
}
\description{
Returns TRUE/FALSE whether the last coordinate of the shapes is the same
as the first one.
}
\examples{
is_closed(matrix(1:10, ncol=2))
is_closed(coo_close(matrix(1:10, ncol=2)))
data(bot)
is_closed(bot)
is_closed(coo_close(bot))
}
\seealso{
\link{coo_close}, \link{coo_unclose}
}
\keyword{ShapeUtilities}

