% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{l2a}
\alias{l2a}
\title{Converts a list of coordinates to an array of coordinates}
\usage{
l2a(l)
}
\arguments{
\item{l}{\code{list} of matrices of the same dimension.}
}
\value{
an array of coordinates.
}
\description{
l2a converts a list of \code{k} matrices with \code{m} rows
and \code{n} columns matrices to a \code{m x n x k} array.
}
\details{
May be useful to communicate with other morphometrics packages that use
array of coordinates when handling configurations of landmarks.
}
\examples{
data(wings)
l <- wings$coo
l
a <- l2a(l)
a
}
\seealso{
\link{a2l}.

Other babel functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{chc2Out}}, \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{l2m}}, \code{\link{m2a}}, \code{\link{m2d}},
  \code{\link{m2l}}, \code{\link{nef2Coe}},
  \code{\link{ntsrow2Coo}}, \code{\link{pix2chc}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

