% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{l2a}
\alias{l2a}
\title{Converts a list of coordinates to an array of coordinates}
\usage{
l2a(l)
}
\arguments{
\item{l}{\code{list} of matrices of the same dimension.}
}
\value{
an array of coordinates.
}
\description{
l2a converts a list of \code{k} matrices with \code{m} rows
and \code{n} columns matrices to a \code{m x n x k} array.
}
\details{
May be useful to communicate with other morphometrics packages that use
array of coordinates when handling configurations of landmarks.
}
\examples{
data(wings)
l <- wings$coo
l
a <- l2a(l)
a
}
\seealso{
Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{d2m}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2ll}}, \code{\link{m2l}}

Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{d2m}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2ll}}, \code{\link{m2l}}
}
