% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{l2m}
\alias{l2m}
\title{Converts a list of coordinates to a matrix of coordinates.}
\usage{
l2m(l)
}
\arguments{
\item{l}{a \code{list} with x and y coordinates as components.}
}
\value{
\code{matrix} of (x; y) coordinates.
}
\description{
Converts a \code{list} with x and y components to a two-columns
(colnamed) \code{matrix} of coordinates.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{m2l}.

Other babel functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{chc2Out}}, \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{l2a}}, \code{\link{m2a}}, \code{\link{m2d}},
  \code{\link{m2l}}, \code{\link{nef2Coe}},
  \code{\link{ntsrow2Coo}}, \code{\link{pix2chc}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

