% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{l2m}
\alias{l2m}
\title{Converts a list of coordinates to a matrix of coordinates.}
\usage{
l2m(l)
}
\arguments{
\item{l}{a \code{list} with x and y coordinates as components.}
}
\value{
\code{matrix} of (x; y) coordinates.
}
\description{
Converts a \code{list} with x and y components to a two-columns
(colnamed) \code{matrix} of coordinates.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{m2l}.
}
\keyword{Babel}

