% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{l2m}
\alias{l2m}
\title{Converts a list of coordinates to a matrix of coordinates.}
\usage{
l2m(l)
}
\arguments{
\item{l}{\code{list} with x and y coordinates as components.}
}
\value{
\code{matrix} of (x; y) coordinates.
}
\description{
Converts a \code{list} with x and y components to a two-columns
(colnamed) \code{matrix} of coordinates. Also, if l is a list with a single matrix, then l[[1]] is returned.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{m2l}.

Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{l2a}}, \code{\link{m2a}}, \code{\link{m2d}},
  \code{\link{m2ll}}, \code{\link{m2l}}
}

