% Generated by roxygen2 (4.0.0): do not edit by hand
\name{LDA}
\alias{LDA}
\title{Linear Discriminant Analysis on Coe objects}
\usage{
LDA(Coe, fac)
}
\arguments{
\item{Coe}{the \link{Coe} object}

\item{fac}{the grouping factor (names of one of the $fac column or column id)}
}
\value{
a "LDA" object on which to apply \link{plot.LDA}. #todo
}
\description{
Performs a LDA on Coe objects. Relies on \link{lda} in MASS.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 24)
bot.l <- LDA(bot.f, "type")
bot.l
plot(bot.l)

bot.f$fac$plop <- factor(rep(letters[1:4], each=10))
bot.l <- LDA(bot.f, "plop")
bot.l
plot(bot.l)
}

