% Generated by roxygen2 (4.0.0): do not edit by hand
\name{LDA}
\alias{LDA}
\title{Linear Discriminant Analysis on Coe objects}
\usage{
LDA(Coe, fac)
}
\arguments{
\item{Coe}{the \link{Coe} object}

\item{fac}{the grouping factor (names of one of the $fac column or column id)}
}
\value{
a "LDA" object on which to apply \link{plot.LDA}. #todo
}
\description{
Performs a LDA on OutCoe, (OpnCoe no yet) objects. Relies on \link{lda} in MASS.
}

