% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LDA}
\alias{LDA}
\alias{LDA.PCA}
\alias{LDA.default}
\title{Linear Discriminant Analysis on Coe objects}
\usage{
LDA(x, fac, retain, ...)

\method{LDA}{default}(x, fac, retain, ...)

\method{LDA}{PCA}(x, fac, retain = 5, ...)
}
\arguments{
\item{x}{a  PCA object}

\item{fac}{the grouping factor (names of one of the $fac column or column id)}

\item{retain}{the number of PC axis to retain for LDA.PCA}

\item{...}{additional arguments to feed \link{lda}}
}
\value{
a 'LDA' object on which to apply \link{plot.LDA}, which is a list with components:
\itemize{
 \item \code{x} any \link{Coe} object (or a matrix)
 \item \code{fac} grouping factor used
 \item \code{removed} ids of columns in the original matrix that have been removed since constant (if any)
 \item \code{mod} the raw lda mod from \link{lda}
 \item \code{mod.pred} the predicted model using x and mod
 \item \code{CV.fac} cross-validated classification
 \item \code{CV.tab} cross-validation tabke
 \item \code{CV.correct} proportion of correctly classified individuals
 \item \code{CV.ce} class error
 \item \code{LDs} unstandardized LD scores see Claude (2008)
 \item \code{mshape} mean values of coefficients in the original matrix
 \item \code{method} inherited from the Coe object (if any)
}
}
\description{
Performs a LDA on Coe objects. Relies on \link{lda} in MASS.
}
\examples{
data(bot)
bot.f <- efourier(bot, 24)
bot.l <- LDA(bot.f, 'type')
bot.l
plot(bot.l)
bot.f$fac$plop <- factor(rep(letters[1:4], each=10))
bot.l <- LDA(bot.f, 'plop')
bot.l
plot(bot.l)
}
\seealso{
\link{plot.LDA}, \link{plot_CV}
}
\keyword{Multivariate}

