% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ldk.chull}
\alias{ldk.chull}
\title{Draws convex hulls around landmark positions}
\usage{
ldk.chull(ldk, col = "grey40", lty = 1)
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{col}{a color for drawing the convex hull}

\item{lty}{an lty for drawing the convex hulls}
}
\description{
A wrapper that uses \link{coo.chull}
}
\examples{
data(wings)
coo.plot(mshape(wings))
ldk.chull(wings$coo)
}
\seealso{
\link{coo.chull}, \link{chull}, \link{ldk.confell}, \link{ldk.contour}
}
\keyword{Graphics}

