% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ldk.contour}
\alias{ldk.contour}
\title{Draws kernel density contours around landmark}
\usage{
ldk.contour(ldk, nlevels = 5, grid.nb = 50, col = "grey60")
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{nlevels}{the number of contour lines}

\item{grid.nb}{the grid.nb}

\item{col}{a color for drawing the contour lines}
}
\description{
Using \link{kde2d} in the MASS package.
}
\examples{
data(wings)
coo.plot(mshape(wings))
ldk.contour(wings$coo)
}
\seealso{
\link{kde2d}, \link{ldk.confell}, \link{ldk.chull}
}
\keyword{Graphics}

