% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ldk.labels}
\alias{ldk.labels}
\title{add landmarks to coo.plot}
\usage{
ldk.labels(ldk, d = 0.1, cex = 0.8, ...)
}
\arguments{
\item{ldk}{a matrix of (x; y) coordinates: where to plot the labels}

\item{d}{how far from the coordinates, on a (centroid-landmark) segment}

\item{cex}{the cex for the label}

\item{...}{additional parameters to fed \link{text}}
}
\description{
add landmarks to coo.plot
}

