% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Ldk}
\alias{Ldk}
\alias{Ldk.Ldk}
\alias{Ldk.default}
\alias{Ldk.list}
\alias{LdkCoe}
\title{Builds an Ldk object}
\usage{
Ldk(x, links = NULL, fac = data.frame())
}
\arguments{
\item{x}{a \code{list} of matrices of (x; y) coordinates,
or an array, an Ldk object.}

\item{links}{a matrix of "links" between landmarks, mainly for plotting}

\item{fac}{(optionnal) a \code{data.frame} of factors,
specifying the grouping structure}
}
\value{
an \code{Ldk} object
}
\description{
In Momocs, \code{Ldk} classes objects are wrapping around
lists of configurations of landmarks, along with other informations,
on which generic methods such as plotting methods (e.g. \link{stack})
and specific methods (e.g. todo-Procrustes can be applied.
 \code{Ldk} objects are primarily \code{\link{Coo}} objects.
}
\details{
All the shapes in x must have the same number of landmarks. If you are
trying to make an Ldk object from an Out or an Opn object, try \link{coo.sample}.
}
\examples{
methods(class=Ldk)
}
\seealso{
\link{Coo}, \link{Out}, link{Opn}.
}
\keyword{Ldk}

