% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-domesticplotters.R
\name{ldk_chull}
\alias{ldk_chull}
\title{Draws convex hulls around landmark positions}
\usage{
ldk_chull(ldk, col = "grey40", lty = 1)
}
\arguments{
\item{ldk}{an array (or a list) of landmarks}

\item{col}{a color for drawing the convex hull}

\item{lty}{an lty for drawing the convex hulls}
}
\description{
A wrapper that uses \link{coo_chull}
}
\examples{
coo_plot(MSHAPES(wings))
ldk_chull(wings$coo)
}
\seealso{
\link{coo_chull}, \link{chull}, \link{ldk_confell}, \link{ldk_contour}

Other plotting functions: 
\code{\link{coo_arrows}()},
\code{\link{coo_draw}()},
\code{\link{coo_listpanel}()},
\code{\link{coo_lolli}()},
\code{\link{coo_plot}()},
\code{\link{coo_ruban}()},
\code{\link{ldk_confell}()},
\code{\link{ldk_contour}()},
\code{\link{ldk_labels}()},
\code{\link{ldk_links}()},
\code{\link{plot_devsegments}()},
\code{\link{plot_table}()}

Other ldk plotters: 
\code{\link{ldk_confell}()},
\code{\link{ldk_contour}()},
\code{\link{ldk_labels}()},
\code{\link{ldk_links}()}
}
\concept{ldk plotters}
\concept{plotting functions}
