% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{m2d}
\alias{m2d}
\title{Converts a matrix of coordinates to a data.frame}
\usage{
m2d(m)
}
\arguments{
\item{m}{a matrix (see above).}
}
\value{
a data.frame (see above).
}
\description{
Converts a \code{m x 2} matrix of coordinates named data.frame.
}
\examples{
data(wings)
m2d(wings[3])
}
\seealso{
\link{m2d} the reverse function.

Other babel functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{chc2Out}}, \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{l2a}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2l}}, \code{\link{nef2Coe}},
  \code{\link{ntsrow2Coo}}, \code{\link{pix2chc}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

