% Generated by roxygen2 (4.0.2): do not edit by hand
\name{m2d}
\alias{m2d}
\title{Converts a matrix of coordinates to a data.frame}
\usage{
m2d(m)
}
\arguments{
\item{m}{a matrix (see above).}
}
\value{
a data.frame (see above).
}
\description{
Converts a \code{m x 2} matrix of coordinates named data.frame.
}
\examples{
data(wings)
m2d(wings[3])
}
\seealso{
\link{m2d} the reverse function.
}
\keyword{Babel}

