% Generated by roxygen2 (4.0.0): do not edit by hand
\name{m2l}
\alias{m2l}
\title{Convert a matrix of coordinates to a list of coordinates.}
\usage{
m2l(m)
}
\arguments{
\item{m}{A 2-columns \code{matrix} containing x and y coordinates.}
}
\value{
Returns a \code{list} with \eqn{x; y} components.
}
\description{
\code{m2l} converts a matrix of \eqn{(x; y)}coordinates to a list with
\code{x; y} components.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{l2m}.
}
\keyword{Babel}

