% Generated by roxygen2 (4.0.2): do not edit by hand
\name{m2l}
\alias{m2l}
\title{Converts a matrix of coordinates to a list of coordinates.}
\usage{
m2l(m)
}
\arguments{
\item{m}{a two-columns \code{matrix} of x and y coordinates.}
}
\value{
a \code{list} with x and y components.
}
\description{
Converts a matrix of (x; y) coordinates to a list with
x and y components.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{l2m}.
}
\keyword{Babel}

