% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-bridges.R
\name{m2l}
\alias{m2l}
\title{Converts a matrix of coordinates to a list of coordinates.}
\usage{
m2l(m)
}
\arguments{
\item{m}{a two-columns \code{matrix} of x and y coordinates.}
}
\value{
a \code{list} with x and y components.
}
\description{
Converts a matrix of (x; y) coordinates to a list with
x and y components.
}
\examples{
data(wings)
l <- m2l(wings[1])
l
m <- l2m(l)
m
}
\seealso{
\link{l2m}.

Other bridges functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{coo2cpx}}, \code{\link{cpx2coo}},
  \code{\link{d2m}}, \code{\link{l2a}}, \code{\link{l2m}},
  \code{\link{m2a}}, \code{\link{m2d}}, \code{\link{m2ll}}
}
