% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{Momocs}
\description{
Morphometrics using R
}
\details{
We are very grateful to  Cedric Gaucherel, Sarah Ivorra, Ricardo Kriebel, Neus
Martinez, Marcelo Reginato, Evan Saitta, Norbert Telmon, Bill Venables, Asher Wishkerman for
their helpful contributions, ideas, bug reports, and much more.
}
\references{
Bonhomme V, Picq S, Gaucherel C, Claude J. 2014. Momocs: Outline Analysis Using R.
\emph{Journal of Statistical Software} \bold{56}. \url{http://www.jstatsoft.org/v56/i13}.

Claude J. 2008. \emph{Morphometrics with R}. Springer-Verlag, New-York.
}
\seealso{
Momocs' homepage : \url{http://www.vincentbonhomme.fr/Momocs} with tutorials
and hotline.

Momocs' GitHub repo : \url{https://github.com/vbonhomme/Momocs} to contribute,
among other things.
}
\keyword{Abtract}

