\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{Momocs}
\description{
Morphometrics using R
}
\references{
Bonhomme V, Picq S, Gaucherel C, Claude J. 2014. Momocs:
Outline Analysis Using R. \emph{Journal of Statistical
Software} \bold{56}.
\url{http://www.jstatsoft.org/v56/i13}.

Claude J. 2008. \emph{Morphometrics with R}.
Springer-Verlag, New-York.
}
\seealso{
Momocs' homepage :
\url{http://www.vincentbonhomme.fr/Momocs}

Momocs' GitHub repo :
\url{https://github.com/vbonhomme/Momocs}
}
\keyword{Abtract}

