% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs-package.R
\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{Momocs}
\description{
The goal of Momocs is to provide a complete, convenient,
reproducible and open-source toolkit for 2D morphometrics.
It includes most common 2D morphometrics approaches on outlines,
open outlines, configurations of landmarks, traditional
morphometrics, and facilities for data preparation,
manipulation and visualization with a consistent grammar throughout.
It allows reproducible, pipeable, complex morphometric analyses
and other morphometrics approaches should be easy to
plug in, or develop from, on top of this canvas.
}
\details{
To cite Momocs in publications: \code{citation("Momocs")}.
}
\section{Cheers}{

We are very grateful to (in alphabetical order): Sean Asselin, Laurent Bouby, Matt Bulbert, Simon Crameri, Julia Cooke, April Dinwiddie,
Carl Lipo, Cedric Gaucherel, Catherine Girard, QGouil (GitHub), Christian Steven Hoggard,
Sarah Ivorra, Glynis Jones, Nathalie Keller, Ricardo Kriebel, Remi Laffont, Fabien Lafuma,
Matthias Mace, Stas Malavin, Neus Martinez, Sabrina Renaud, Marcelo Reginato, Evan Saitta, David Siddons, Eleanor Stillman,
Theodore Stammer, Norbert Telmon, Jean-Frederic Terral, Bill Venables, Daniele Ventura, Michael Wallace,
Asher Wishkerman, John Wood for their helpful ideas and bug reports.
}

\references{
\itemize{
\item Bonhomme V, Picq S, Gaucherel C, Claude J. 2014. Momocs: Outline Analysis Using R.
\emph{Journal of Statistical Software} \bold{56}. \url{http://www.jstatsoft.org/v56/i13}.
\item Claude J. 2008. \emph{Morphometrics with R}. Springer-Verlag, New-York.
}
}
\seealso{
\itemize{
\item \bold{Homepage}: \url{https://github.com/MomX/Momocs}
\item \bold{Issues}: \url{https://github.com/MomX/Momocs/issues}
\item \bold{Tutorial}: \code{browseVignettes("Momocs")} or \url{http://momx.github.io/Momocs/}
\item \bold{Email}: \code{bonhomme.vincent@gmail.com} to contribute to dev,
ask for something, propose collaboration, share your data, etc.
}
}
