% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Momocs-package.R
\name{Momocs_help}
\alias{Momocs_help}
\title{Browse Momocs online doc}
\usage{
Momocs_help(topic = NULL)
}
\arguments{
\item{topic}{the function name to access. If not specified the homepage of
the online manual is accessed.}
}
\description{
Launch a browser to an online version of the manual
}

