% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs-package.R
\name{Momocs_update}
\alias{Momocs_update}
\title{Install and load the last version of Momocs}
\usage{
Momocs_update()
}
\description{
Download the last version of Momocs from its GitHub account
from \code{http://www.github.com/vbonhomme/Momocs}), install it and load it (\code{library(Momocs)})
}
\details{
You need devtools, but it is checked anyway.
}

