% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mshapes}
\alias{mshapes}
\alias{mshapes.LdkCoe}
\alias{mshapes.OpnCoe}
\alias{mshapes.OutCoe}
\alias{mshapes.default}
\title{Mean shape calculation from Coe objects}
\usage{
mshapes(Coe, fac, FUN, nb.pts)

\method{mshapes}{default}(Coe, fac, FUN, nb.pts)

\method{mshapes}{OutCoe}(Coe, fac, FUN = mean, nb.pts = 120)

\method{mshapes}{OpnCoe}(Coe, fac, FUN = mean, nb.pts = 120)

\method{mshapes}{LdkCoe}(Coe, fac, FUN = NULL, nb.pts = 120)
}
\arguments{
\item{Coe}{a \link{Coe} object}

\item{fac}{factor from the $fac slot. See examples below.}

\item{FUN}{a function to compute the mean shape (\link{mean} by default, by \link{median} can be considered)}

\item{nb.pts}{numeric the number of points for calculated shapes}
}
\value{
a list with two components: \code{$Coe} object of the same class, and
\code{$shp} a list of matrices of (x, y) coordinates.
}
\description{
Calculates mean shapes on matrices of coefficients by groups (if passed with
a 'fac') or globally (if not), and on \link{Coe} objects.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
mshapes(bot.f) # the mean (global) shape
ms <- mshapes(bot.f, 'type')
ms$Coe
class(ms$Coe)
ms <- ms$shp
coo.plot(ms$beer)
coo.draw(ms$whisky, border='forestgreen')
tps.arr(ms$whisky, ms$beer) #etc.

data(olea)
op <- rawPolynomials(subset(olea, view=='VL'), 5)
ms <- mshapes(op, 'cep') #etc
ms$Coe
panel(Opn(ms$shp), names=TRUE)

data(wings)
wp <- fgProcrustes(wings, tol=1e-4)
ms <- mshapes(wp, 1)
ms$Coe
panel(Ldk(ms$shp), names=TRUE) #etc.
panel(ms$Coe) # equivalent (except the $fac slot)
}
\seealso{
\link{mshape} for operations on raw coordinates.
}
\keyword{Multivariate}

