% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult-MSHAPESs.R
\name{MSHAPES}
\alias{MSHAPES}
\alias{MSHAPES.list}
\alias{MSHAPES.array}
\alias{MSHAPES.Ldk}
\alias{MSHAPES.OutCoe}
\alias{MSHAPES.OpnCoe}
\alias{MSHAPES.LdkCoe}
\alias{MSHAPES.PCA}
\alias{MSHAPES.LDA}
\alias{mshapes}
\title{Mean shape calculation for Coo, Coe, etc.}
\usage{
MSHAPES(x, ...)

\method{MSHAPES}{list}(x, FUN = mean, ...)

\method{MSHAPES}{array}(x, FUN = mean, ...)

\method{MSHAPES}{Ldk}(x, FUN = mean, ...)

\method{MSHAPES}{OutCoe}(x, fac = NULL, FUN = mean, nb.pts = 120, ...)

\method{MSHAPES}{OpnCoe}(x, fac = NULL, FUN = mean, nb.pts = 120, ...)

\method{MSHAPES}{LdkCoe}(x, fac = NULL, FUN = mean, ...)

\method{MSHAPES}{PCA}(x, fac, FUN = mean, ...)

\method{MSHAPES}{LDA}(x, fac, FUN = mean, ...)

mshapes(...)
}
\arguments{
\item{x}{a list, array, Ldk, LdkCoe, OutCoe or OpnCoe or PCA object}

\item{...}{useless here.}

\item{FUN}{a function to compute the mean shape (\link{mean} by default, by \link{median} can be considered)}

\item{fac}{factor specification for \link{fac_dispatcher}}

\item{nb.pts}{numeric the number of points for calculated shapes (only Coe objects)}
}
\value{
the averaged shape; on Coe objects, a list with two components: \code{$Coe} object of the same class, and
\code{$shp} a list of matrices of (x, y) coordinates. On \link{PCA} and \link{LDA} objects, the FUN (typically mean or median)
of scores on \code{PCs} or \code{LDs}. This method used on the latter objects may be moved to another function at some point.
}
\description{
Quite a versatile function that calculates mean (or median, or whatever function)
on list or an array of shapes, an Ldk object. It can also be used on Coe objects.
In that case, the reverse transformation (from coefficients to shapes) is calculated, (within
groups defined with the fac argument if provided) and the Coe object is \emph{also} returned
(in \code{$Coe}) along with a list of shapes (in \code{$shp}) and can then be passed to \code{plot_MSHAPES}.
}
\examples{
#### on shapes
MSHAPES(wings)
MSHAPES(wings$coo)
MSHAPES(coo_sample(bot, 24)$coo)
stack(wings)
coo_draw(MSHAPES(wings))

bot.f <- efourier(bot, 12)
MSHAPES(bot.f) # the mean (global) shape
ms <- MSHAPES(bot.f, 'type')
ms$Coe
class(ms$Coe)
ms <- ms$shp
coo_plot(ms$beer)
coo_draw(ms$whisky, border='forestgreen')
\dontrun{
tps_arr(ms$whisky, ms$beer) #etc.
}

op <- olea \%>\% filter(view=="VL") \%>\% npoly(5)
ms <- MSHAPES(op, ~var) #etc
ms$Coe
panel(Opn(ms$shp), names=TRUE)

wp <- fgProcrustes(wings, tol=1e-4)
ms <- MSHAPES(wp, 1)
plot_MSHAPES(ms)

# on PCA and LDA just return a data_frame
bot.f \%>\% PCA \%>\% MSHAPES(~type)
bot.f \%>\% LDA(~fake) \%>\% MSHAPES(~fake)
# LDA on ~fake but average on type
bot.f \%>\% LDA(~fake) \%>\% MSHAPES(~type)
}
\seealso{
Other multivariate: \code{\link{CLUST}},
  \code{\link{KMEANS}}, \code{\link{KMEDOIDS}},
  \code{\link{LDA}}, \code{\link{MANOVA_PW}},
  \code{\link{MANOVA}}, \code{\link{MDS}},
  \code{\link{NMDS}}, \code{\link{PCA}},
  \code{\link{classification_metrics}}
}
