% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{mutate}
\alias{mutate}
\title{Mutates (ala dplyr) on Momocs objects}
\usage{
mutate(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{comma separated list of unquoted expressions}
}
\value{
a Momocs object of the same class.
}
\description{
Add new variables to the \code{$fac}. See examples and \code{?dplyr::mutate}.
}
\details{
dplyr verbs are maintained.
}
\examples{
olea
mutate(olea, id=factor(1:length(olea)))
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{at_least}}, \code{\link{chop}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{filter}}, \code{\link{rename}},
  \code{\link{rescale}}, \code{\link{rm_harm}},
  \code{\link{rm_uncomplete}}, \code{\link{rw_fac}},
  \code{\link{sample_frac}}, \code{\link{sample_n}},
  \code{\link{select}}, \code{\link{slice}},
  \code{\link{subset.Coo}}, \code{\link{table}}
}
