% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{nef2Coe}
\alias{nef2Coe}
\title{Imports .nef to Coe objects}
\usage{
nef2Coe(nef.path)
}
\arguments{
\item{nef.path}{the path to the .nef file}
}
\description{
Useful to convert .nef files into Coe objects.
It returns a matrix of coefficients that can be passed to \link{Coe}.
}
\note{
I'm not very familiar to other morphometric formats.
So if you have troubles importing your datasets, contact me, I can help. Or if you fix something,
please let met know!
}
\seealso{
Other babel functions: \code{\link{bind_db}},
  \code{\link{chc2Out}}, \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{ntsrow2Coo}}, \code{\link{pix2chc}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

