% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ntsrow2Coo}
\alias{ntscol2Coo}
\alias{ntsrow2Coo}
\title{From .nts to Coo objects}
\usage{
ntsrow2Coo(nts.path, sep = "\\t")
}
\arguments{
\item{nts.path}{the path to the .nts file}

\item{sep}{the separator between data}
}
\description{
Useful to convert .nts files into Coo objects. For .nts provided as rows, use
ntsrow2Coo; for .nts provided as columns of coordinates, try ntscol2Coo. It
returns a list of matrices of coordinates that can be passed to \link{Coo}.
}
\details{
I'm not very familiar to other morphometric formats and these
  functions are (highly) experimental. So if you have troubles importing your
  datasets, please contact me.
}
\examples{
\dontrun{
# That's how wings dataset was created
# made a local copy from http://life.bio.sunysb.edu/morph/data/RohlfSlice1990Mosq.nts
# then :
coo.list  <- ntscol2Coo("~/Desktop/mosquitowings.nts)
fac       <- data.frame(fac=factor(substr(names(coo.list), 1, 2)))
wings <- Ldk(coo.list, fac=fac)
}
}
\keyword{babel}

