% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Opn}
\alias{Opn}
\title{Builds an Opn object}
\usage{
Opn(x, ldk = NULL, fac = data.frame())
}
\arguments{
\item{x}{\code{list} of matrices of \eqn{(x; y)} coordinates}

\item{ldk}{(optionnal) \code{list} of landmarks as row number indices}

\item{fac}{(optionnal) a \code{data.frame} of factors,
specifying the grouping structure}
}
\value{
an \code{Opn} object
}
\description{
In Momocs, \code{Opn} classes objects are wrapping around
lists of \bold{open} outlines, along with other informations,
on which generic methods such as plotting methods (e.g. \link{stack})
and specific methods (e.g. \link{rawPolynomials} can be applied.
 \code{Opn} objects are primarily \code{\link{Coo}} objects.
}
\details{
These methods can be applied on \code{Opn} objects:
\enumerate{
\item Handling: \code{subset, coo.} family;
\item Calibration : \code{nqual, nquant, npow};
\item Graphics: \code{plot, stack, panel};
\item Morpho: \code{rawPolynomials, orthoPolynomials}.
}
}
\seealso{
\link{Coo}, \link{Out}
}
\keyword{Opn}

