% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Out}
\alias{Out}
\title{Builds an Out object}
\usage{
Out(x, ldk = list(), fac = data.frame)
}
\arguments{
\item{x}{a \code{list} of matrices of \eqn{(x; y)} coordinates,
or an array or an Out object or an Ldk object}

\item{ldk}{(optionnal) \code{list} of landmarks as row number indices}

\item{fac}{(optionnal) a \code{data.frame} of factors,
specifying the grouping structure}
}
\value{
an \code{Out} object
}
\description{
In Momocs, \code{Out}-classes objects are lists of closed \bold{out}lines,
with optionnal components, and on which generic methods such as plotting methods (e.g. \link{stack})
and specific methods (e.g. \link{eFourier} can be applied.
 \code{Out} objects are primarily \code{\link{Coo}} objects.
}
\examples{
methods(class=Out)
}
\seealso{
\link{Coo}, \link{Opn}, link{Ldk}.
}
\keyword{Out}

