% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-def-Out.R
\name{Out}
\alias{Out}
\title{Builds an Out object}
\usage{
Out(x, fac = dplyr::tibble(), ldk = list())
}
\arguments{
\item{x}{a \code{list} of matrices of \verb{(x; y)} coordinates,
or an array or an Out object or an Ldk object, or a data.frame (and friends)}

\item{fac}{(optional) a \code{data.frame} of factors and/or numerics
specifying the grouping structure}

\item{ldk}{(optional) \code{list} of landmarks as row number indices}
}
\value{
an \code{Out} object
}
\description{
In Momocs, \code{Out}-classes objects are lists of closed \bold{out}lines,
with optional components, and on which generic methods such as plotting methods (e.g. \link{stack})
and specific methods (e.g. \link{efourier} can be applied.
\code{Out} objects are primarily \code{\link{Coo}} objects.
}
\examples{
methods(class=Out)
}
\seealso{
Other classes: 
\code{\link{Coe}()},
\code{\link{Coo}()},
\code{\link{Ldk}()},
\code{\link{Opn}()},
\code{\link{OpnCoe}()},
\code{\link{OutCoe}()},
\code{\link{TraCoe}()}
}
\concept{classes}
