% Generated by roxygen2 (4.0.0): do not edit by hand
\name{panel}
\alias{panel}
\title{Plot on Coo objects: family picture}
\usage{
panel(Coo, cols, borders, names = NULL, cex.names = 0.6, ...)
}
\arguments{
\item{Coo}{The \code{Coo (Out/Opn)} object  to plot.}

\item{cols}{A \code{vector} of colors for drawing the outlines.
Either a single value or of length exactly equals to the number of coordinates.}

\item{borders}{A \code{vector} of colors for drawing the borders.
Either a single value or of length exactly equals to the number of coordinates.}

\item{names}{whether to plot names or not. If TRUE uses shape names, otherwise
pass a character for the names of the files}

\item{cex.names}{a cex for the names}

\item{...}{further arguments to be passed to \link{coo.list.panel}}
}
\description{
Plots all the outlines from a \code{Coo} side by side
}
\examples{
data(mosquito)
panel(mosquito, names=TRUE, cex.names=0.5)
data(olea)
panel(olea)
}
\seealso{
\link{stack.Coo}, \link{plot.Coo}.
}

