% Generated by roxygen2 (4.0.0): do not edit by hand
\name{panel}
\alias{panel}
\title{Plot on Coo objects: family picture}
\usage{
panel(Coo, cols, borders, fac, reorder, palette = col.summer, names = NULL,
  cex.names = 0.6, points = TRUE, points.pch = 3, points.cex = 0.2,
  points.col, ...)
}
\arguments{
\item{Coo}{The \code{Coo (Out/Opn)} object  to plot.}

\item{cols}{A \code{vector} of colors for drawing the outlines.
Either a single value or of length exactly equals to the number of coordinates.}

\item{borders}{A \code{vector} of colors for drawing the borders.
Either a single value or of length exactly equals to the number of coordinates.}

\item{fac}{a factor within the $fac slot for colors}

\item{reorder}{a factor or a numeric to reorder shapes}

\item{palette}{a color \link{palette}}

\item{names}{whether to plot names or not. If TRUE uses shape names, otherwise
pass a character for the names of the files}

\item{cex.names}{a cex for the names}

\item{points}{logical (for Ldk) whether to draw points}

\item{points.pch}{(for Ldk) and a pch for these points}

\item{points.cex}{(for Ldk) and a cex for these points}

\item{points.col}{(for Ldk) and a col  for these points}

\item{...}{further arguments for the generic}
}
\description{
Plots all the outlines from a \code{Coo} side by side
}
\examples{
data(mosquito)
panel(mosquito, names=TRUE, cex.names=0.5)
data(olea)
panel(olea)
}
\seealso{
\link{stack.Coo}, \link{plot.Coo}.
}

