% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coo.R
\name{panel2}
\alias{panel2}
\title{Family picture of shapes (ggplot2)}
\usage{
panel2(Coo)
}
\arguments{
\item{Coo}{a Coo object}
}
\value{
a ggplot2 object
}
\description{
May replace panel one day.
}
\examples{
data(shapes)
panel2(shapes)
}
\seealso{
Other Coo_graphics: \code{\link{panel}},
  \code{\link{plot.Coo}}, \code{\link{stack.Coo}}
}
