% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PC.contrib}
\alias{PC.contrib}
\alias{PC.contrib.PCA}
\title{Shape variation along PC axes}
\usage{
PC.contrib(PCA, ...)

\method{PC.contrib}{PCA}(PCA, nax = 1:4, sd.r = c(-2, -1, -0.5, 0, 0.5, 1,
  2), main = "PC contribution to shape", xlab = "(Mean + ) SD",
  ylab = "PC axes", ...)
}
\arguments{
\item{PCA}{a \code{\link{PCA}} object}

\item{...}{additional parameter to pass to \code{\link{coo.draw}}}

\item{nax}{a single or a range of PC axes}

\item{sd.r}{a single or a range of mean +/- sd values (eg: c(-1, 0, 1))}

\item{main}{a title for the plot}

\item{xlab}{a title for the x-axis}

\item{ylab}{a title for the y-axis}
}
\description{
Calculates and plots shape variation along Principal Component axes.


}
\examples{
data(bot)
 bot.p <- PCA(eFourier(bot, 12))
 PC.contrib(bot.p)
 PC.contrib(bot.p, nax=1:5, sd.r=c(-5, 2, 1, 0, 1, 2, 5),
   main="A nice title", border="grey40", col="grey80")
}

