% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pca}
\alias{pca}
\title{Principal component analysis on Coe objects}
\usage{
pca(Coe)
}
\arguments{
\item{Coe}{the \link{Coe} object}
}
\value{
a "PCA" object on which to apply \link{plot.PCA}
}
\description{
Performs a PCA on OutCoe, OpnCoe objects.
}

