\name{pca2shp.efourier}
\alias{pca2shp.efourier}
\title{Returns shape for a point on a PC plan}
\usage{
pca2shp.efourier(pos, rot, mshape, amp.shp = 1, pts.shp = 60)
}
\arguments{
  \item{pos}{the position on the plan}

  \item{rot}{the loadings}

  \item{mshape}{the mean shape}

  \item{amp.shp}{the amplification factor}

  \item{pts.shp}{the number of point for drawing the
  shapes}
}
\description{
todo
}
\keyword{graphics}

