% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PCA2shp.polynomials}
\alias{PCA2shp.polynomials}
\title{Calculates shapes from PC plane: polynomials}
\usage{
PCA2shp.polynomials(pos, rot, mshape, amp.shp = 1, pts.shp = 60, ortho,
  baseline1, baseline2)
}
\arguments{
\item{pos}{the position on two PC axis}

\item{rot}{the corresponding loadings}

\item{mshape}{the meanshape}

\item{amp.shp}{amplification factor for the shape deformation}

\item{pts.shp}{number of points to reconstruct the shape}

\item{ortho}{logical whether working with raw or orthogonal polynomials}

\item{baseline1}{the (x; y) coordinates of the first baseline point}

\item{baseline2}{the (x; y) coordinates of the second baseline point}
}
\description{
Calculates shapes from PC plane: polynomials
}
\keyword{Graphics}

