% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pix2chc}
\alias{pix2chc}
\title{Convert (x; y) coordinates to chaincoded coordinates}
\usage{
pix2chc(coo)
}
\arguments{
\item{coo}{(x; y) coordinates passed as a matrix}
}
\description{
Useful to convert (x; y) coordinates to chain-coded coordinates.
}
\examples{
data(shapes)
pix2chc(shapes[1])
}
\references{
Kuhl, F. P., & Giardina, C. R. (1982).
Elliptic Fourier features of a closed contour.
\emph{Computer Graphics and Image Processing}, 18(3), 236-258.
}
\seealso{
\link{chc2pix}
}
\keyword{Babel}

