% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{pix2chc}
\alias{pix2chc}
\title{Convert (x; y) coordinates to chaincoded coordinates}
\usage{
pix2chc(coo)
}
\arguments{
\item{coo}{(x; y) coordinates passed as a matrix}
}
\description{
Useful to convert (x; y) coordinates to chain-coded coordinates.
}
\examples{
data(shapes)
pix2chc(shapes[1])
}
\references{
Kuhl, F. P., & Giardina, C. R. (1982).
Elliptic Fourier features of a closed contour.
\emph{Computer Graphics and Image Processing}, 18(3), 236-258.
}
\seealso{
\link{chc2pix}

Other babel functions: \code{\link{chc2Out}},
  \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{nef2Coe}}, \code{\link{ntsrow2Coo}},
  \code{\link{tie_jpg_txt}}, \code{\link{tps2coo}}
}

