% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.PCA}
\alias{plot.PCA}
\title{Plots Principal Component Analysis}
\usage{
\method{plot}{PCA}(x, fac, xax = 1, yax = 2, points = TRUE,
  col = "#000000", pch = 20, cex = 0.5, palette = col_solarized,
  center.origin = FALSE, zoom = 1, bg = par("bg"), grid = TRUE,
  nb.grids = 3, morphospace = TRUE, pos.shp = "full", amp.shp = 1,
  size.shp = 1, nb.shp = 12, nr.shp = 6, nc.shp = 5, pts.shp = 60,
  border.shp = .transp("#000000", 0.5), lwd.shp = 1,
  col.shp = .transp("#000000", 0.95), stars = FALSE, ellipses = FALSE,
  conf_ellipses = 0.5, ellipsesax = TRUE, conf_ellipsesax = c(0.5, 0.75,
  0.9), lty.ellipsesax = 1, lwd.ellipsesax = sqrt(2), chull = FALSE,
  chull.lty = 3, density = FALSE, lev.density = 20, contour = FALSE,
  lev.contour = 3, n.kde2d = 100, delaunay = FALSE, loadings = FALSE,
  labels = FALSE, col.labels = par("fg"), cex.labels = 0.6,
  abbreviate.labels = TRUE, labelsgroups = TRUE, cex.labelsgroups = 0.8,
  rect.labelsgroups = FALSE, abbreviate.labelsgroups = FALSE,
  axisnames = TRUE, axisvar = TRUE, eigen = TRUE, rug = TRUE,
  title = substitute(x), box = TRUE, old.par = TRUE, ...)
}
\arguments{
\item{x}{an object of class "PCA", typically obtained with \link{PCA}}

\item{fac}{factor, or a name or the column id from the $fac slot, or a formula combining colum names
from the $fac slot (cf. examples)}

\item{xax}{the first PC axis}

\item{yax}{the second PC axis}

\item{points}{logical whether to plot points}

\item{col}{a color for the points (either global, for every level of the fac
or for every individual, see examples)}

\item{pch}{a pch for the points (either global, for every level of the fac
or for every individual, see examples)}

\item{cex}{the size of the points}

\item{palette}{a \link{palette}}

\item{center.origin}{logical whether to center the plot onto the origin}

\item{zoom}{to keep your distances}

\item{bg}{color for the background}

\item{grid}{logical whether to draw a grid}

\item{nb.grids}{and how many of them}

\item{morphospace}{logical whether to add the morphological space}

\item{pos.shp}{either "full", "range", "circle", "xy"
or a data.frame for \link{pos.shapes}}

\item{amp.shp}{amplification factor for shape deformation}

\item{size.shp}{the size of the shapes}

\item{nb.shp}{(pos.shp="circle") the number of shapes on the compass}

\item{nr.shp}{(pos.shp="full" or "range) the number of shapes per row}

\item{nc.shp}{(pos.shp="full" or "range) the number of shapes per column}

\item{pts.shp}{the number of points fro drawing shapes}

\item{border.shp}{the border color of the shapes}

\item{lwd.shp}{the line width for these shapes}

\item{col.shp}{the color of the shapes}

\item{stars}{logical whether to draw "stars"}

\item{ellipses}{logical whether to draw confidence ellipses}

\item{conf_ellipses}{numeric the quantile for the (bivariate gaussian) confidence ellipses}

\item{ellipsesax}{logical whether to draw ellipse axes}

\item{conf_ellipsesax}{one or more numeric, the quantiles for the (bivariate gaussian) ellipses axes}

\item{lty.ellipsesax}{if yes, the lty with which to draw these axes}

\item{lwd.ellipsesax}{if yes, one or more numeric for the line widths}

\item{chull}{logical whether to draw a convex hull}

\item{chull.lty}{if yes, its linetype}

\item{density}{whether to add a 2d density kernel estimation (based on \link{kde2d})}

\item{lev.density}{if yes, the number of levels to plot (through \link{image})}

\item{contour}{whether to add contour lines based on 2d density kernel}

\item{lev.contour}{if yes, the (approximate) number of lines to draw}

\item{n.kde2d}{the number of bins for \link{kde2d}, ie the 'smoothness' of density kernel}

\item{delaunay}{logical whether to add a delaunay 'mesh' between points}

\item{loadings}{logical whether to add loadings for every variables}

\item{labels}{logical whether to add point labels}

\item{col.labels}{a color for these labels}

\item{cex.labels}{a cex for these labels}

\item{abbreviate.labels}{logical whether to abbreviate these labels}

\item{labelsgroups}{logical whether to add labels for groups}

\item{cex.labelsgroups}{ifyes, a numeric for the size of the labels}

\item{rect.labelsgroups}{logical whether to add a rectangle behind groups names}

\item{abbreviate.labelsgroups}{if yes, whether to abbreviate group names}

\item{axisnames}{logical whether to add PC names}

\item{axisvar}{logical whether to draw the variance they explain}

\item{eigen}{logical whether to draw a plot of the eigen values}

\item{rug}{logical whether to add rug to margins}

\item{title}{character a name for the plot}

\item{box}{whether to draw a box around the plotting region}

\item{old.par}{whether to restore the old \link{par}. Set it to \code{FALSE} if you want to reuse the graphical window.}

\item{...}{useless here, just to fit the generic plot}
}
\description{
The Momocs' \code{\link{PCA}} plotter with morphospaces and many graphical options.
}
\details{
Widely inspired by the "layers" philosophy behind graphical functions
of the ade4 R package.
}
\examples{
data(bot)
bot.f <- eFourier(bot, 12)
bot.p <- PCA(bot.f)
bot.p
plot(bot.p, morpho=FALSE)
plot(bot.p, "type")

data(olea)
op <- rawPolynomials(olea, 5)
op.p <- PCA(op)
op.p
plot(op.p, ~ domes + cep, morpho=TRUE) # use of formula

data(wings)
wp <- fgProcrustes(wings, tol=1e-4)
wpp <- PCA(wp)
wpp
plot(wpp, 1)
}
\seealso{
\link{plot.LDA}
}
\keyword{Graphics}
\keyword{Multivariate,}

