% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_CV}
\alias{plot_CV}
\alias{plot_CV.LDA}
\alias{plot_CV.table}
\title{Plots a cross-validation table as an heatmap}
\usage{
plot_CV(x, ...)

\method{plot_CV}{LDA}(x, ...)

\method{plot_CV}{table}(x, freq = TRUE, palette = col_heat, levels = 20,
  cols, pc = TRUE, margin = TRUE, cex = 1, ...)
}
\arguments{
\item{x}{a (typically cross-correlation) table to plot}

\item{...}{not used}

\item{freq}{whether to use row-wise frequencies}

\item{palette}{a color palette such as \link{col_heat} to use,
if \code{cols} is not used.}

\item{levels}{number of levels, otherwise the highest cell in the table}

\item{cols}{a vector of colors}

\item{pc}{whether to use percentages}

\item{margin}{whether to add marginal sums}

\item{cex}{a cex for all values}
}
\description{
Either with frequencies (or percentages) plus marginal sums,
and values as heatmaps. Used in Momocs for plotting cross-validation tables
but may be used for any table (likely with \code{freq=FALSE}).
}
\examples{
data(bot)
bot.p <- PCA(efourier(bot, 12))
bot.l <- LDA(bot.p, 1)
plot_CV(bot.l)
data(olea)
ol <- LDA(PCA(npoly(olea, nb.pts=50)), "cep")
plot_CV(ol)
# raw counts
plot_CV(ol, freq=FALSE, palette=col_india)
# any other count table
m <- matrix(runif(120, 0, 6), 12)
tab <- as.table(round(m))
plot_CV(tab, palette=terrain.colors, levels=5, cex=0.8)
}
\seealso{
\link{LDA}, \link{plot.LDA}, \link{plot_CV2}.
}
\keyword{Graphics}
\keyword{Multivariate}

