% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-cheapbabi-plot_and_layers.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Multivariate plots using cheapbabi layers}
\usage{
plot_PCA(x, f, axes = c(1, 2), palette = col_summer)
}
\arguments{
\item{x}{\code{PCA} object}

\item{f}{\code{factor}. A column name or number from \code{$fac},
or a factor can directly be passed. Accept \code{numeric} as well.}

\item{axes}{\code{numeric} of length two to select PCs to use
(\code{c(1, 2)} by default)}

\item{palette}{\code{color palette} to use \code{col_summer} by default}
}
\description{
Quickly vizualize \link{PCA} objects and build customs plots
using the \link{cheapbabi} layers. See examples.
}
\note{
This method will replace \link{plot.PCA} in further versions.
This \code{cheap base biplot} may be packaged at some point.
All comments are welcome.
}
\examples{
### First prepare two PCA objects.
# Some outlines with bot
bp <- bot \%>\% mutate(fake=sample(letters[1:5], 40, replace=TRUE)) \%>\%
efourier(6) \%>\% PCA
# Some curves with olea
op <- olea \%>\%
mutate(s=coo_area(.)) \%>\%
filter(var != "Cypre") \%>\%
chop(view) \%>\% lapply(opoly, 5, nb.pts=90) \%>\%
combine \%>\% PCA
op$fac$s \%<>\% as.character() \%>\% as.numeric()

op \%>\% plot_PCA \%>\% layer_frame

### Now we can play with layers

op \%>\% plot_PCA \%>\% layer_fullframe \%>\% layer_points
op \%>\% plot_PCA(~s) \%>\% layer_fullframe \%>\% layer_points(pch=3, cex=2)
op \%>\% plot_PCA(~var+domes) \%>\% layer_frame \%>\% layer_ellipses

# you can combine the layers you like with a function like:
myplot <- function(x){
x \%>\%
 layer_fullframe \%>\%
 layer_points \%>\%
 layer_ellipsesaxes \%>\%
 layer_chullfilled
}

# and even continue after this function
op \%>\% plot_PCA(~var) \%>\% myplot() \%>\%
   layer_stars \%>\% layer_title("hi there!") \%>\%
   layer_rug \%>\% layer_eigen \%>\% layer_legend

# You get the idea.
}
