% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCV}
\alias{plotCV}
\title{Plots a cross-validation table as an heatmap}
\usage{
plotCV(x, freq = TRUE, palette = col.heat, levels = 20, cols, pc = TRUE,
  margin = TRUE, cex = 1)
}
\arguments{
\item{x}{a (typically cross-correlation) table to plot}

\item{freq}{whether to use row-wise frequencies}

\item{palette}{a color palette such as \link{col.heat} to use,
if \code{cols} is not used.}

\item{levels}{number of levels, otherwise the highest cell in the table}

\item{cols}{a vector of colors}

\item{pc}{whether to use percentages}

\item{margin}{whether to add marginal sums}

\item{cex}{a cex for all values}
}
\description{
Either with frequencies (or percentages) plus marginal sums,
and values as heatmaps. Used in Momocs for plotting cross-validation tables
but may be used for any table (likely with \code{freq=FALSE}).
}
\examples{
data(bot)
bot.p <- PCA(eFourier(bot, 12))
bot.l <- LDA(bot.p, 1)
tab <- bot.l$CV.tab
tab
plotCV(tab)

data(olea)
ol <- LDA(PCA(rawPolynomials(olea, nb.pts=50)), "cep")
plotCV(ol$CV.tab)
# raw counts
plotCV(tab, freq=FALSE, palette=col.india)
# any other count table
m <- matrix(runif(120, 0, 6), 12)
tab <- as.table(round(m))
plotCV(tab, palette=terrain.colors, levels=5, cex=0.8)
}
\seealso{
\link{LDA}, \link{plot.LDA}, \link{plotCV2}.
}
\keyword{Graphics}
\keyword{Multivariate}

