% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polynomials.i}
\alias{polynomials.i}
\title{Calculate shape from a polynomial model}
\usage{
polynomials.i(pol, nb.pts = 120, reregister = TRUE)
}
\arguments{
\item{pol}{a pol list such as created by \link{polynomials}.}

\item{nb.pts}{the number of points to predict. By default (and can't be higher)
the number of points in the original shape.}

\item{reregister}{logical whether to reregister the shape with the original baseline.}
}
\value{
a matrix of (x; y) coordinates.
}
\description{
Returns a matrix of (x; y) coordinates when passed with a model obtained with
\link{polynomials}.
}
\examples{
data(olea)
o <- olea[5]
coo.plot(o)
for (i in 2:7){
x <- polynomials.i(polynomials(o, i, ortho=TRUE))
coo.draw(x, border=col.summer(7)[i], points=FALSE)  }
}
\keyword{morphoCore}

