\name{pos.shapes}
\alias{pos.shapes}
\title{Calculates nice positions on a plan for drawing shapes}
\usage{
pos.shapes(xy, pos.shp = c("range", "circle")[1], nb.shp = 12, nr.shp = 6,
  nc.shp = 5, circle.r.shp)
}
\arguments{
  \item{xy}{todo}

  \item{pos.shp}{the way shape should be positionned}

  \item{nb.shp}{the total number of shapes}

  \item{nr.shp}{the number of rows to position shapes}

  \item{nc.shp}{the number of cols to position shapes}

  \item{circle.r.shp}{if circle, its radius}
}
\description{
todo
}
\keyword{graphics}

