% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gr-morphospaces.R
\name{pos.shapes}
\alias{pos.shapes}
\title{Calculates nice positions on a plane for drawing shapes}
\usage{
pos.shapes(xy, pos.shp = c("range", "circle", "xy")[1], nb.shp = 12,
  nr.shp = 6, nc.shp = 5, circle.r.shp)
}
\arguments{
\item{xy}{todo}

\item{pos.shp}{how shapes should be positionned: \code{range} of xy,
\code{full} extent of the plane, \code{circle} as a rosewind,
on \code{xy} values provided.}

\item{nb.shp}{the total number of shapes}

\item{nr.shp}{the number of rows to position shapes}

\item{nc.shp}{the number of cols to position shapes}

\item{circle.r.shp}{if circle, its radius}
}
\description{
Calculates nice positions on a plane for drawing shapes
}
\keyword{Graphics}

