% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pProcrustes}
\alias{pProcrustes}
\title{Partial Procrustes alignment between two shapes}
\usage{
pProcrustes(coo1, coo2)
}
\arguments{
\item{coo1}{Configuration matrix to be superimposed onto the centered preshape of coo2.}

\item{coo2}{Reference configuration matrix.}
}
\value{
a list with components
\describe{
\item{coo1}{superimposed centered preshape of coo1 onto the centered preshape of coo2}
\item{coo2}{centered preshape of coo2}
\item{rotation}{rotation matrix}
\item{DP}{partial Procrustes distance between coo1 and coo2}
\item{rho}{trigonometric Procrustes distance}.
}
}
\description{
Directly borrowed from Claude (2008), and called \code{pPsup} there.
}
\references{
Claude, J. (2008). Morphometrics with R. Analysis (p. 316). Springer.
}

