% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{sample_n}
\alias{sample_n}
\alias{sample_n.Coe}
\alias{sample_n.Coo}
\alias{sample_n.default}
\title{Samples n shapes in Momocs objects}
\usage{
sample_n(tbl, size, replace, ...)

\method{sample_n}{default}(tbl, size, replace = FALSE, ...)

\method{sample_n}{Coo}(tbl, size, replace = FALSE, fac = NULL, ...)

\method{sample_n}{Coe}(tbl, size, replace = FALSE, fac = NULL, ...)
}
\arguments{
\item{tbl}{a Momocs object (Coo, Coe)}

\item{size}{numeric how many shapes should we sample}

\item{replace}{logical whether sample should be done with ot without replacement}

\item{...}{additional arguments to dplyr::sample_n and to maintain generic compatibility}

\item{fac}{a column name if a $fac is defined; size is then applied within levels of this factor}
}
\description{
Uses (and maintain) dplyr syntax and verb.
}
\examples{
data(bot)
bot
# samples 5 bottles no matter their type
sample_n(bot, 5)
# 5 bottles of beer and of whisky
table(sample_n(bot, 5, fac="type"))
# many repetitions
table(names(sample_n(bot, 400, replace=TRUE)))
}
\seealso{
\link{sample_frac}
}

