% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coo.R
\name{stack.Coo}
\alias{stack.Coo}
\alias{stack.Ldk}
\alias{stack.OutCoe}
\title{Plots Coo objects: stack (all shapes in the same frame)}
\usage{
\method{stack}{Coo}(x, cols, borders, fac, palette = col_summer,
  coo_sample = 120, points = FALSE, first.point = TRUE, centroid = TRUE,
  ldk = TRUE, ldk.pch = 3, ldk.col = "#FF000055", ldk.cex = 0.5,
  ldk_links = FALSE, ldk_confell = FALSE, ldk_contour = FALSE,
  ldk_chull = FALSE, ldk_labels = FALSE, cur.pch = ".", xy.axis = TRUE,
  title = substitute(x), ...)

\method{stack}{OutCoe}(x, nb.pts = 120, ...)

\method{stack}{Ldk}(x, cols, borders, first.point = TRUE, centroid = TRUE,
  ldk = TRUE, ldk.pch = 3, ldk.col = "#333333", ldk.cex = 0.5,
  ldk_links = FALSE, ldk_confell = FALSE, ldk_contour = FALSE,
  ldk_chull = FALSE, ldk_labels = FALSE, cur.pch = ".", xy.axis = TRUE,
  title = substitute(x), ...)
}
\arguments{
\item{x}{The \code{Coo} object to plot.}

\item{cols}{A \code{vector} of colors for drawing the outlines.
Either a single value or of length exactly equals to the number of coordinates.}

\item{borders}{A \code{vector} of colors for drawing the borders.
Either a single value or of length exactly equals to the number of coordinates.}

\item{fac}{a factor within the $fac slot for colors}

\item{palette}{a color palette to use when fac is provided}

\item{coo_sample}{if not NULL the number of point per shape to display (to plot quickly)}

\item{points}{logical whether to draw or not points}

\item{first.point}{logical whether to draw or not the first point}

\item{centroid}{logical whether to draw or not the centroid}

\item{ldk}{\code{logical}. Whether to display landmarks (if any).}

\item{ldk.pch}{A \code{pch} for these landmarks}

\item{ldk.col}{A color for these landmarks}

\item{ldk.cex}{A \code{cex} fro these landmarks}

\item{ldk_links}{logical whether to draw links (of the mean shape)}

\item{ldk_confell}{logical whether to draw conf ellipses}

\item{ldk_contour}{logical whether to draw contour lines}

\item{ldk_chull}{logical whether to draw convex hull}

\item{ldk_labels}{logical whether to draw landmark labels}

\item{cur.pch}{\code{pch} for semi landmarks}

\item{xy.axis}{whether to draw or not the x and y axes}

\item{title}{a title for the plot. The name of the \code{Coo} by default}

\item{...}{further arguments to be passed to \link{coo_plot}}

\item{nb.pts}{the number of points to use for the shape reconstruction}
}
\description{
Plots all the outlines, on the same graph, from a \link{Coo} (\link{Out}, \link{Opn} or \link{Ldk})
object.
}
\note{
When applied on a \link{OutCoe} object, a wrapper
for \code{stack(as.Out(OutCoe), ...)}. In that case,
\code{...} feeds \code{stack} itself. (same thing for \code{OpnCoe} to come)
}
\examples{
\dontrun{
data(bot)
stack(bot)
bot.f <- efourier(bot, 12)
stack(bot.f)
data(mosquito)
stack(mosquito, borders='#1A1A1A22', first.point=FALSE)
data(hearts)
stack(hearts)
stack(hearts, ldk=FALSE)
stack(hearts, borders='#1A1A1A22', ldk=TRUE, ldk.col=col_summer(4), ldk.pch=20)
stack(hearts, fac="aut", palette=col_sari)
}
}
\seealso{
\link{panel.Coo}, \link{plot.Coo}.
}

