% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-Coo.R
\name{stack2}
\alias{stack2}
\title{Family picture of shapes (ggplot2)}
\usage{
stack2(Coo)
}
\arguments{
\item{Coo}{a Coo object
Family picture of shapes}
}
\value{
a ggplot2 object
}
\description{
Will replace stack soon.
}
\examples{
data(bot)
stack2(bot)
}

