\name{subset.Coo}
\alias{subset.Coo}
\title{Create subsets of Coo objects}
\usage{
\method{subset}{Coo}(x, subset, ...)
}
\arguments{
  \item{x}{and Coo object}

  \item{subset}{logical from the fac or indices}

  \item{...}{(to preserve consistence with subset generic)}
}
\description{
todo
}
\keyword{Coo}

