% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{subset.Coo}
\alias{subset.Coe}
\alias{subset.Coo}
\alias{subset.PCA}
\title{Subsets on Momocs objects}
\usage{
\method{subset}{Coo}(x, subset, ...)

\method{subset}{Coe}(x, subset, ...)

\method{subset}{PCA}(x, subset, ...)
}
\arguments{
\item{x}{a \code{Coo} or a \link{Coe} object.}

\item{subset}{logical taken from the \code{$fac} slot, or indices. See examples.}

\item{...}{useless here but maintains consistence with the generic subset.}
}
\description{
Subset is a wrapper around dplyr's verbs and should NOT be used directly.
}
\examples{
# Do not use subset directly
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{at_least}}, \code{\link{chop}},
  \code{\link{combine}}, \code{\link{dissolve}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{rename}}, \code{\link{rm_uncomplete}},
  \code{\link{rw_fac}}, \code{\link{sample_frac}},
  \code{\link{sample_n}}, \code{\link{select}},
  \code{\link{slice}}
}

