% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl-utilities.R
\name{table}
\alias{table}
\alias{table.Coe}
\alias{table.Coo}
\alias{table.LDA}
\alias{table.PCA}
\alias{table.default}
\title{Cross tabulation of Momocs objects}
\usage{
table(...)

\method{table}{default}(...)

\method{table}{Coo}(...)

\method{table}{Coe}(...)

\method{table}{PCA}(...)

\method{table}{LDA}(...)
}
\arguments{
\item{...}{a list of, first, a Momocs object (Coo, Coe, PCA, etc.), then, column names in the $fac slot. If not specified,
returns a table on the entire $fac data.frame}
}
\description{
Simply extends base \link{table} for a more convenient use on $fac slot.
}
\examples{
data(bot)
table(bot)
data(olea)
table(olea, "var", "domes")
table(olea)
}

