% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel-import.R
\name{tie_jpg_txt}
\alias{tie_jpg_txt}
\title{Binds .jpg outlines and .txt landmarks on them}
\usage{
tie_jpg_txt(lf)
}
\arguments{
\item{lf}{a list of filenames}
}
\description{
Given a list of files (lf) that includes matching filenames with .jpg (black masks)
and .txt (landmark positions on them as .txt), returns an Out with $ldk defined.
Typically be useful if you use ImageJ to define landmarks on your outlines.
}
\note{
Not optimized (images are read twice). Please do not hesitate to contact me
should you have a particular case or need something.
}
\seealso{
Other babel functions: \code{\link{a2l}},
  \code{\link{a2m}}, \code{\link{as_df}},
  \code{\link{chc2Out}}, \code{\link{chc2pix}},
  \code{\link{import_StereoMorph_curve1}},
  \code{\link{import_jpg}}, \code{\link{import_tps}},
  \code{\link{l2a}}, \code{\link{l2m}}, \code{\link{m2a}},
  \code{\link{m2d}}, \code{\link{m2l}},
  \code{\link{nef2Coe}}, \code{\link{ntsrow2Coo}},
  \code{\link{pix2chc}}, \code{\link{tps2coo}}
}

