% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/babel-import-otherformats.R
\name{tps2Coo}
\alias{tps2Coo}
\title{From .tps to Coo objects}
\usage{
tps2Coo(tps.path, sep = " ")
}
\arguments{
\item{tps.path}{the path to the .tps file}

\item{sep}{the separator between data}
}
\description{
Useful to convert .tps files into Coo objects.
It returns a list of matrices of coordinates that can be passed to \link{Coo} (\link{Out}, \link{Opn} or \link{Ldk}).
}
\note{
I'm not very familiar to other morphometric formats.
So if you have troubles importing your datasets, contact me, I can help. Or if you fix something,
please let met know!
}
\keyword{Babel}

