% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tps2Coo}
\alias{tps2Coo}
\title{From .tps to Coo objects}
\usage{
tps2Coo(tps.path, sep = " ")
}
\arguments{
\item{tps.path}{the path to the .tps file}

\item{sep}{the separator between data}
}
\description{
Useful to convert .tps files into Coo objects.
It returns a list of matrices of coordinates that can be passed to \link{Coo}.
}
\details{
I'm not very familiar to other morphometric formats.
So if you have troubles importing your datasets, please contact me.
}
\keyword{babel}

