% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tps2d}
\alias{tps2d}
\title{Thin Plate Splines for 2D data}
\usage{
tps2d(grid0, fr, to)
}
\arguments{
\item{grid0}{a matrix of coordinates on which to calculate deformations}

\item{fr}{the reference \eqn{(x; y)} coordinates}

\item{to}{the target \eqn{(x; y)} coordinates}
}
\value{
a matrix of \code{(x; y)} coordinates with TPS-interpolated
deformations
}
\description{
\code{tps2d} is the core function for Thin Plate Splines. It is used
internally but might be useful elsewhere.
}
\seealso{
\link{tps.grid},\link{tps.iso}, \link{tps.arr} functions use
\code{tps2d}
}
\keyword{ThinPlateSplines}

