% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-def-TraCoe.R
\name{TraCoe}
\alias{TraCoe}
\title{Traditional morphometrics class}
\usage{
TraCoe(coe = matrix(), fac = data.frame())
}
\arguments{
\item{coe}{a matrix of measurements}

\item{fac}{a data.frame for covariates}
}
\description{
Defines the builder for traditional measurement class in Momocs.
Is is intended to ease calculations, data handling and multivariate statistics
just ad the other Momocs' classes
}
\examples{
# let's (more or less) rebuild the flower dataset
fl <- TraCoe(iris[, 1:4], data.frame(sp=iris$Species))
fl \%>\% PCA() \%>\% plot("sp")
}
\seealso{
Other classes: \code{\link{Coe}}, \code{\link{Coo}},
  \code{\link{OpnCoe}}, \code{\link{Opn}},
  \code{\link{OutCoe}}, \code{\link{Out}}
}
