% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-handling.R
\name{transmute}
\alias{transmute}
\title{transmute (ala dplyr) on Momocs classes}
\usage{
transmute(.data, ...)
}
\arguments{
\item{.data}{a \code{Coo}, \code{Coe}, \code{PCA} object}

\item{...}{comma separated list of unquoted expressions}
}
\value{
a Momocs object of the same class.
}
\description{
Add new variables to the \code{$fac} and drop existing ones. See examples and \code{?dplyr::transmute}.
}
\details{
dplyr verbs are maintained.
}
\examples{
olea
transmute(olea, id=factor(1:length(olea)))
}
\seealso{
Other handling functions: \code{\link{arrange}},
  \code{\link{chop}}, \code{\link{combine}},
  \code{\link{dissolve}}, \code{\link{filter}},
  \code{\link{mutate}}, \code{\link{rename}},
  \code{\link{sample_frac}}, \code{\link{sample_n}},
  \code{\link{select}}, \code{\link{slice}}
}

